### 2023-06-30 #5609 @editorialbot
[PRE REVIEW]: PyBCI: Python Framework for User-Friendly and Real-Time Brain Computer Interfaces with LSL
**Submitting author:** @lmbooth (<a href="http://orcid.org/0000-0002-8749-9726">Liam Booth</a>)
**Repository:** https://github.com/LMBooth/pybci
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.1
