### 2023-06-30 #5609 @LMBooth
Sorry, i'm trying to update the version as the most up-to-date is v1.0.1 not v1.0.0 as i had some minor pypi tags and release version edits to make.
