### 2023-07-03 #5609 @LMBooth
Hi @emdupre,

Reviewing the list i think JanCBrammer could be a good choice with his interest in electrophysiology and biosignals, tuliofalmeida or anilbey for a few recommendations?

Is that enough ro would you like me to pick a couple more?
