+++ docs/conf.py
@@ -5,7 +5,7 @@ copyright = '2023, Liam Booth'
 author = 'Liam Booth'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2.4b2'
+release = '1.0.1'
 
 # -- General configuration ---------------------------------------------------
 extensions = [
