### 2023-07-13 #5609 @JanCBrammer
@emdupre, thanks for the consideration. Unfortunately I don't have the bandwidth for a review atm.
I set my status to `unavailable` over at https://reviewers.joss.theoj.org/reviewers for now.
