### 2023-07-28 #5609 @emdupre
Thank you for agreeing to review _PyBCI_, @anilbey and @jsheunis ! As we now have a sufficient number of reviewers, I'll start the review now and will reach out to you both in the dedicated review issue.

Thank you again for considering this invitation, @Ildaron, @rougier, and @tuliofalmeida !
