### 2023-07-28 #5706 @emdupre
👋 Hi @LMBooth, I did have one structural request on the paper to help facilitate the review process. Currently, the compiled paper does not include an associated reference list. Can you please add a references section at the end of the paper? Please also make sure that any relevant Acknowledgements are included as well.

You can see more about structuring your JOSS paper in [our submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
