### 2023-07-28 #5706 @LMBooth
Sorry about that @emdupre , i wasn't sure if a references section was going to be automatically added from the bib in the repository, i'll start add those now.

Thank you again for pointing these points out.
