+++ JOSS/paper.bib
@@ -91,7 +91,7 @@ author={
  pages={2825--2830},
  year={2011}
 }
-@ARTICLE{2020SciPy-NMeth,
+@article{2020SciPy-NMeth,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
             Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
@@ -113,4 +113,14 @@ author={
   pages   = {261--272},
   adsurl  = {https://rdcu.be/b08Wh},
   doi     = {10.1038/s41592-019-0686-2},
+}
+@article{article,
+author = {Bateson, Anthony and Asghar, Aziz},
+year = {2021},
+month = {05},
+pages = {1-1},
+title = {Development and Evaluation of a Smartphone-Based Electroencephalography (EEG) System},
+volume = {PP},
+journal = {IEEE Access},
+doi = {10.1109/ACCESS.2021.3079992}
 }
