### 2023-07-28 #5706 @emdupre
Thank you for your quick response, @LMBooth ! It should be automatically added if you include an additional `# References` heading at the end of the `paper.md` file.

If you encounter any issues, of course, please let me know !
