+++ JOSS/paper.md
@@ -47,6 +47,6 @@ By offering a comprehensive, open-source platform for BCI research, PyBCI has th
 # Acknowledgements
 
 Bateson & Ashgar (2021) created the io:bio mobile EEG device which was used to create an initial port for streaming time-series physiological data in to the Lab Streaming Layer, so we could receive, analyse, record, and classify EMG, ECG and EEG data - enabling prior required experimentation to creating PyBCI.
-The work carried out by Christian Kothe (2014) creating the Lab Streaming Layer and continuous maintenance to the pylsl repository by Chadwick Boulay with some helpful recommendation in the GitHub issue: https://github.com/labstreaminglayer/pylsl/issues/70.
+The work carried out by Christian Kothe (2014) creating the Lab Streaming Layer and continuous maintenance to the pylsl repository by Chadwick Boulay enables unification across many off the shelf devices. Chadwick Boulay also gave helpful recommendation in the GitHub issue: https://github.com/labstreaminglayer/pylsl/issues/70.
 
 # References
