### 2023-07-28 #5706 @LMBooth
Hey again @emdupre, sorry for being a noob, I've added an acknowledgements section and  `# References` at the end of the document, but i can't see any imported references  when i look at the article proof, can you tell if my formatting is off at the end of my [paper.md](https://raw.githubusercontent.com/LMBooth/pybci/main/JOSS/paper.md) or my [paper.bib](https://raw.githubusercontent.com/LMBooth/pybci/main/JOSS/paper.bib)?
