+++ JOSS/paper.bib
@@ -1,4 +1,3 @@
-
 @book{NEURIPS2019_9015,
 	title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
 	booktitle = {Advances in Neural Information Processing Systems 32},
@@ -10,6 +9,7 @@
 	author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
 	editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d{\textquoteright}Alch{\'e}-Buc and E. Fox and R. Garnett}
 }
+
 @book{oliphant2006guide, 
   title={A guide to NumPy}, 
   author={Oliphant, Travis E}, 
@@ -17,6 +17,7 @@
   year={2006}, 
   publisher={Trelgol Publishing USA} 
 }
+
 @misc{tensorflow2015-whitepaper,
 title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
 url={https://www.tensorflow.org/},
@@ -64,6 +65,7 @@ author={
     Xiaoqiang~Zheng},
   year={2015},
 }
+
 @misc{lsl,
   title = {Lab Streaming Layer},
   author = {Kothe, Christian},
@@ -71,6 +73,7 @@ author={
   howpublished = {\url{https://github.com/sccn/labstreaminglayer}},
   note = {Accessed: June 14, 2023}
 }
+
 @misc{vallat_antropy_2023,
   author = {Raphael Vallat},
   title = {AntroPy: entropy and complexity of (EEG) time-series in Python},
@@ -80,6 +83,7 @@ author={
   howpublished = {\url{https://github.com/raphaelvallat/antropy}},
   note = {Python 3 package providing several time-efficient algorithms for computing the complexity of time-series, used for example to extract features from EEG signals}
 }
+
 @article{scikit-learn,
  title={Scikit-learn: Machine Learning in {P}ython},
  author={Pedregosa, F. and Varoquaux, G. and Gramfort, A. and Michel, V.
@@ -91,6 +95,7 @@ author={
  pages={2825--2830},
  year={2011}
 }
+
 @article{2020SciPy-NMeth,
   author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
             Haberland, Matt and Reddy, Tyler and Cournapeau, David and
@@ -114,13 +119,14 @@ author={
   adsurl  = {https://rdcu.be/b08Wh},
   doi     = {10.1038/s41592-019-0686-2},
 }
+
 @article{article,
-author = {Bateson, Anthony and Asghar, Aziz},
-year = {2021},
-month = {05},
-pages = {1-1},
-title = {Development and Evaluation of a Smartphone-Based Electroencephalography (EEG) System},
-volume = {PP},
-journal = {IEEE Access},
-doi = {10.1109/ACCESS.2021.3079992}
+	author = {Bateson, Anthony and Asghar, Aziz},
+	year = {2021},
+	month = {05},
+	pages = {1-1},
+	title = {Development and Evaluation of a Smartphone-Based Electroencephalography (EEG) System},
+	volume = {PP},
+	journal = {IEEE Access},
+	doi = {10.1109/ACCESS.2021.3079992}
 }
