### 2023-07-31 #5706 @emdupre
Hi @LMBooth, always happy to help !

It looks like what's happening is that you're actually writing your citations in plain-text in paper.md, rather than in the correct LaTeX formatting. For example, if you look at the `# Citations` section in [the example paper in the docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), you can see the following styles:

```
For a quick reference, the following citation commands can be used:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```

So, in your paper.md, you'll want to modify every citation accordingly. As an example, right now you have:

```
PyTorch (Paszke et al., 2019), SciKit-learn (Pedregosa et al., 2011) and TensorFlow (Abadi et al., 2016)
```

This should be changed to:

```
PyTorch [@NEURIPS2019_9015], SciKit-learn [@scikit-learn], and TensorFlow [@tensorflow2015-whitepaper]
```
