+++ JOSS/paper.bib
@@ -120,7 +120,7 @@ author={
   doi     = {10.1038/s41592-019-0686-2},
 }
 
-@article{article,
+@article{2021bateson_asghar,
 	author = {Bateson, Anthony and Asghar, Aziz},
 	year = {2021},
 	month = {05},
