### 2023-08-02 LMBooth/pybci#1 @anilbey
specify minimum versions of dependencies
Hello,

Could you specify the minimum required versions of these dependencies?

https://github.com/LMBooth/pybci/blob/118ecbd2227ef056f46351912ec716e6cba2482b/setup.py#L102C1-L108C16

This issue is part of [#5706](https://github.com/openjournals/joss-reviews/issues/5706).
