### 2023-08-02 LMBooth/pybci#2 @anilbey
LSL installation instructions are missing
Hi,

I followed the instructions on the README and got this error.

```python

>>> import pybci

RuntimeError: LSL binary library file was not found. Please make sure that the binary file can be found in the package lib folder
 (/my-env-path/pybci/lib/python3.10/site-packages/pylsl/lib)
 or the system search path. Alternatively, specify the PYLSL_LIB environment variable.
 You can install the LSL library with conda: `conda install -c conda-forge liblsl`
or otherwise download it from the liblsl releases page assets: https://github.com/sccn/liblsl/releases
```

* Could you mention in the installation section that LSL is a dependency and needs to be installed explicitly?
* If there are any recommended installation instructions, could you provide those for the supported operating systems?
* Could you also outline the supported operating systems e.g. linux windows mac?

This issue is part of https://github.com/openjournals/joss-reviews/issues/5706.
