### 2023-08-02 LMBooth/pybci#2 @LMBooth
Hi @anilbey!

Thankyou for pointing this out.
The liblsl dependency is wrapped up with pylsl and should have been installed when downloading pybci by being defined in the 'install_requires'  list provided in setup.py, i'm at work right now but i'll have a go on a clean install of python when i get home to investigate further.

Thankyou again and i'll get in touch when i've had chance to have a play.
