### 2023-08-02 LMBooth/pybci#2 @LMBooth
Do you have a Mac @anilbey? I've just got my friend to check on his Mac and he had a similar error,  if so does `brew install labstreaminglayer/tap/lsl` fix for you?

I'll add some additional instructions to the git and read the docs specifically for Mac in this case, I'll see him tomorrow and have check the process for myself. I'll also double check on an ubuntu machine tomorrow too.
