### 2023-08-03 LMBooth/pybci#2 @anilbey
I am on an Ubuntu 22.04.

> I'll add some additional instructions to the git and read the docs specifically for Mac in this case

That's great.
