### 2023-08-03 LMBooth/pybci#2 @LMBooth
Ahh, i've just seen pylsl does specify in the [prerequisites](https://github.com/labstreaminglayer/pylsl#prerequisites) section that non windows users need to source liblsl, i'll update my installation instructions to notify people accordingly and direct them to the [liblsl repo](https://github.com/sccn/liblsl) for specific instructions.

Thank you again for bringing this up @anilbey.
