+++ setup.py
@@ -99,13 +99,13 @@ setup(
     },
 
     install_requires=[
-        "pylsl",
-        "scipy",
-        "numpy",
-        "antropy",
-        "tensorflow",
-        "scikit-learn",
-        "torch"
+        "pylsl>=1.16.1",
+        "scipy>=1.11.1",
+        "numpy>=1.24.3",
+        "antropy>=0.1.6",
+        "tensorflow>=2.13.0",
+        "scikit-learn>=1.3.0",
+        "torch>=2.0.1"
     ],
     # You can just specify the packages manually here if your project is
     # simple. Or you can use find_packages().
