### 2023-08-03 LMBooth/pybci#2 @LMBooth
Hi again @anilbey ,

I've made some changes to the installation instructions on the[ main github page](https://github.com/LMBooth/pybci#prerequisite-for-non-windows-users) and [read the docs](https://github.com/LMBooth/pybci#prerequisite-for-non-windows-users), if you find they are insufficient please do say and i can expand where needed. If they are sufficient we can close the issue.
