### 2023-08-03 LMBooth/pybci#1 @LMBooth
Hi @anilbey,
I've added minimum required version in [setup.py](https://github.com/LMBooth/pybci/blob/main/setup.py).

I've also added some[ text to the main page ](https://github.com/LMBooth/pybci#prerequisite-for-non-windows-users) of the git explaining which versions are supported and warning previous versions have not been tested.
