### 2023-08-07 LMBooth/pybci#1 @anilbey
Looks great, thanks @LMBooth
