### 2023-08-07 LMBooth/pybci#2 @anilbey
Hi @LMBooth , the "[pylsl repository](https://github.com/LMBooth/pybci/blob/main/home/liam/.local/lib/python3.10/site-packages/pylsl/lib)" link on the readme seems to be broken. It forwards here: https://github.com/LMBooth/pybci/blob/main/home/liam/.local/lib/python3.10/site-packages/pylsl/lib
