### 2023-08-07 LMBooth/pybci#2 @anilbey
The 2nd link was sufficient for me, I can install it now.
