### 2023-08-07 #5706 @anilbey
I completed the installation following the provided instructions. The tool's purpose is well articulated and will certainly provide a significant advantage to the community.

However, I encountered some challenges, notably with the examples within pybci, which require specific sensor hardware that I do not possess. As a result, I was unable to execute the examples and test the software's functionality.

I also noticed the absence of automated tests within the software. Automated testing is vital for maintaining software integrity and enabling continuous integration.

To address these issues, I suggest the implementation of a mock layer that can emulate real hardware streams for each example provided. This solution would facilitate automated testing and empower all contributors, regardless of their hardware capabilities, to understand the code and contribute to the software. The mock streams should ideally behave identically to the hardware streams, thus eliminating the need for additional code to differentiate between the two.
