### 2023-08-09 LMBooth/pybci#2 @LMBooth
Silly me! I'll fix that now.
