### 2023-08-09 #5706 @LMBooth
Hi @anilbey, these are great recommendations that hopefully won't take me too long to implement.

I currently have a [psuedo lsl data generator in the examples](https://github.com/LMBooth/pybci/tree/main/pybci/Examples/PsuedoLSLStreamGenerator). If i take that and make a psuedo class that can be called within pybci, eliminating the need for running the viewer in a separate file. This would also make it simpler to create some automated testing scripts.

I go on holiday tomorrow but am back mid next week so will begin making the changes then. Thankyou again for going through my work and finding these recommendations, i'll post an update here when i've had chance to address.
