+++ pybci/Examples/PsuedoLSLStreamGenerator/README.md
@@ -5,6 +5,7 @@ The PsuedoLSLSreamGenerator found in mainSend.py in this directory can generate 
 A baseline signal is generated on an LSL stream outlet and a PyQt button can be pressed to signify this baseline signal on a separate LSL marker stream. 
 There are a number of markers that can be sent on the LSL marker stream which alters the data stream outlet signal for a given amount of time.
 
+## Python Package Prerequisites
 To run this script requires installation of PyQt5 for the button interface and PyQtGraph for the data plotting.
 
 ### Note:
