+++ pybci/Examples/PsuedoLSLStreamGenerator/README.md
@@ -1,6 +1,6 @@
 
 ## Description
-The PsuedoLSLSreamGenerator found in mainSend.py in this directory can generate multiple channels on a given stream type at a given sample rate.
+The PseudoLSLSreamGenerator found in mainSend.py in this directory can generate multiple channels on a given stream type at a given sample rate.
 
 A baseline signal is generated on an LSL stream outlet and a PyQt button can be pressed to signify this baseline signal on a separate LSL marker stream. 
 There are a number of markers that can be sent on the LSL marker stream which alters the data stream outlet signal for a given amount of time.
