### 2023-08-24 #5706 @LMBooth
Hi all!

For a brief update, i've managed to create a pseudo device which can be called as a class, i found the GIL to be troublesome when running both the pseudo device (data + marker generation + lsl streams) and pybci at the same time, when viewing the generated data in an LSL viewer (brainvision) there was a lot of starting and stopping in the generation, to overcome this i've created a controller class which is used to call the pseudodevice in either another core or another thread depending on the users choice, the threaded method seems to be okay and a bit more work is needed on the multiprocessed version, hopefully by the end of this weekend i can push an updated version (likely 1.1.0).
