### 2023-09-02 050a23f
pseudodevice - alpha
pybci pseudodevice creation
