+++ build/lib/pybci/__init__.py
@@ -1,0 +1,2 @@
+from .pybci import PyBCI
+#from .Utils import Classifier, FeatureExtractor, LSLScanner
