+++ pybci/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.0.2'
+__version__ = '1.1.0-alpha'

[…]
