+++ install_pybci.egg-info/SOURCES.txt
@@ -1,29 +1,0 @@
-LICENSE
-MANIFEST.in
-README.md
-setup.py
-install_pybci.egg-info/PKG-INFO
-install_pybci.egg-info/SOURCES.txt
-install_pybci.egg-info/dependency_links.txt
-install_pybci.egg-info/requires.txt
-install_pybci.egg-info/top_level.txt
-pybci/__init__.py
-pybci/pybci.py
-pybci/version.py
-pybci/Configuration/EpochSettings.py
-pybci/Configuration/FeatureSettings.py
-pybci/Configuration/PsuedoDeviceSettings.py
-pybci/Configuration/__init__.py
-pybci/ThreadClasses/AsyncDataReceiverThread.py
-pybci/ThreadClasses/ClassifierThread.py
-pybci/ThreadClasses/DataReceiverThread.py
-pybci/ThreadClasses/FeatureProcessorThread.py
-pybci/ThreadClasses/MarkerThread.py
-pybci/ThreadClasses/OptimisedDataReceiverThread.py
-pybci/ThreadClasses/__init__.py
-pybci/Utils/Classifier.py
-pybci/Utils/FeatureExtractor.py
-pybci/Utils/LSLScanner.py
-pybci/Utils/Logger.py
-pybci/Utils/PseudoDevice.py
-pybci/Utils/__init__.py
