+++ docs/index.rst
@@ -3,7 +3,7 @@ Welcome to the PyBCI documentation!
 
 **PyBCI**  is a Python package to create a Brain Computer Interface (BCI) with data synchronisation and pipelining handled by the `Lab Streaming Layer <https://github.com/sccn/labstreaminglayer>`_, machine learning with  `Pytorch <https://pytorch.org/>`_, `scikit-learn <https://scikit-learn.org/stable/#>`_ or `TensorFlow <https://www.tensorflow.org/install>`_, leveraging packages like `Antropy <https://github.com/raphaelvallat/antropy>`_, `SciPy <https://scipy.org/>`_ and `NumPy <https://numpy.org/>`_ for generic time and/or frequency based feature extraction or optionally have the users own custom feature extraction class used.
 
-The goal of PyBCI is to enable quick iteration when creating pipelines for testing human machine and brain computer interfaces, namely testing applied data processing and feature extraction techniques on custom machine learning models. Training the BCI requires LSL enabled devices and an LSL marker stream for training stimuli. (The `examples folder <https://github.com/LMBooth/pybci/tree/main/pybci/Examples>`_ found on the github has a `pseudo LSL data generator and marker creator <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/PsuedoLSLStreamGenerator>`_ in the `mainSend.py <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/PsuedoLSLStreamGenerator/mainSend.py>`_ file so the examples can run without the need of LSL capable hardware.)
+The goal of PyBCI is to enable quick iteration when creating pipelines for testing human machine and brain computer interfaces, namely testing applied data processing and feature extraction techniques on custom machine learning models. Training the BCI requires LSL enabled devices and an LSL marker stream for training stimuli. (The `examples folder <https://github.com/LMBooth/pybci/tree/main/pybci/Examples>`_ found on the github has a `pseudo LSL data generator and marker creator <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/PseudoLSLStreamGenerator>`_ in the `mainSend.py <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/PseudoLSLStreamGenerator/mainSend.py>`_ file so the examples can run without the need of LSL capable hardware.)
 
 `Github repo here! <https://github.com/LMBooth/pybci/>`_
 
