+++ build/lib/pybci/pybci.py
@@ -1,348 +1,0 @@
-from .Utils.LSLScanner import LSLScanner
-from .Utils.Logger import Logger
-from .ThreadClasses.FeatureProcessorThread import FeatureProcessorThread
-from .ThreadClasses.DataReceiverThread import DataReceiverThread
-from .ThreadClasses.AsyncDataReceiverThread import AsyncDataReceiverThread
-from .ThreadClasses.OptimisedDataReceiverThread import OptimisedDataReceiverThread
-from .ThreadClasses.MarkerThread import MarkerThread
-from .ThreadClasses.ClassifierThread import ClassifierThread
-from .Configuration.EpochSettings import GlobalEpochSettings, IndividualEpochSetting
-from .Configuration.FeatureSettings import GeneralFeatureChoices
-import queue, threading, copy
-import tensorflow as tf
-#import torch
-import torch.nn as nn
-
-#tf.get_logger().setLevel('ERROR')
-
-class PyBCI:
-    globalEpochSettings = GlobalEpochSettings()
-    customEpochSettings = {}
-    minimumEpochsRequired = 10
-    markerThread = []
-    dataThreads = []
-    streamChsDropDict= {}
-    dataStreams = []
-    markerStream = None
-    connected = False
-    epochCounts = {} # holds markers received, their target ids and number received of each
-    classifierInformation = []
-    clf= None
-    model = None 
-    torchModel = None
-    def __init__(self, dataStreams = None, markerStream= None, streamTypes = None, markerTypes = None, loggingLevel = Logger.INFO,
-                 globalEpochSettings = GlobalEpochSettings(), customEpochSettings = {}, streamChsDropDict = {},
-                 streamCustomFeatureExtract = {},
-                 minimumEpochsRequired = 10, clf= None, model = None, torchModel = None):
-        """
-        The PyBCI object stores data from available lsl time series data streams (EEG, pupilometry, EMG, etc.)
-        and holds a configurable number of samples based on lsl marker strings.
-        If no marker strings are available on the LSL the class will close and return an error.
-        Parameters
-        ----------
-        dataStreams: List[str] 
-            Allows user to set custom acceptable EEG stream definitions, if None defaults to streamTypes scan
-        markerStream: List[str] 
-            Allows user to set custom acceptable Marker stream definitions, if None defaults to markerTypes scan
-        streamTypes: List[str] 
-            Allows user to set custom acceptable EEG type definitions, ignored if dataStreams not None
-        markerTypes: List[str] 
-            Allows user to set custom acceptable Marker type definitions, ignored if markerStream not None
-        loggingLevel: string 
-            Sets PyBCI print level, ('INFO' prints all statements, 'WARNING' is only warning messages, 'TIMING' gives estimated time for feature extraction, and  classifier training or testing,  'NONE' means no prints from PyBCI)
-        globalEpochSettings (GlobalEpochSettings): 
-            Sets global timing settings for epochs.
-        customEpochSettings: dict 
-            Sets individual timing settings for epochs. {markerstring1:IndividualEpochSettings(),markerstring2:IndividualEpochSettings()}
-        streamChsDropDict: dict 
-            Keys for dict should be respective datastreams with corresponding list of which channels to drop.  {datastreamstring1: list(ints), datastreamstring2: list(ints)}
-        streamCustomFeatureExtract: dict
-            Allows dict to be passed of datastream with custom feature extractor class for analysing data.  {datastreamstring1: customClass1(), datastreamstring2: customClass1(),}
-        minimumEpochsRequired: int 
-            Minimm number of required epochs before model fitting begins, must be of each type of received markers and mroe then 1 type of marker to classify.
-        clf: sklearn.base.ClassifierMixin 
[…]
