+++ build/lib/pybci/__init__.py
@@ -1,2 +1,0 @@
-from .pybci import PyBCI
-#from .Utils import Classifier, FeatureExtractor, LSLScanner
