+++ build/lib/pybci/ThreadClasses/AsyncDataReceiverThread.py
@@ -1,126 +1,0 @@
-import threading, time
-from collections import deque
-import itertools
-from bisect import bisect_left
-def slice_fifo_by_time(fifo, start_time, end_time):
-        """Find the slice of fifo between start_time and end_time using binary search."""
-        # separate times and data for easier indexing
-        times, data = zip(*fifo)
-        # find the index of the first time that is not less than start_time
-        start_index = bisect_left(times, start_time)
-        # find the index of the first time that is not less than end_time
-        end_index = bisect_left(times, end_time)
-        # return the slice of data between start_index and end_index
-        return data[start_index:end_index]
-
-class AsyncDataReceiverThread(threading.Thread):
-    """Responsible for receiving data from accepted LSL outlet, slices samples based on tmin+tmax basis, 
-    starts counter for received samples after marker is received in ReceiveMarker. Relies on timestamps to slice array, 
-    suspected more computationally intensive then synchronous method.
-    """
-    startCounting = False
-    currentMarker = ""
-    def __init__(self, closeEvent, trainTestEvent, dataQueueTrain,dataQueueTest, dataStreamInlet,  
-                 customEpochSettings, globalEpochSettings,devCount,  streamChsDropDict = [], maxExpectedSampleRate=100):
-        # large maxExpectedSampleRate can incur marker drop and slow procesing times for slicing arrays
-        super().__init__()
-        self.trainTestEvent = trainTestEvent
-        self.closeEvent = closeEvent
-        self.dataQueueTrain = dataQueueTrain
-        self.dataQueueTest = dataQueueTest
-        self.dataStreamInlet = dataStreamInlet
-        self.customEpochSettings = customEpochSettings
-        self.globalEpochSettings = globalEpochSettings
-        self.streamChsDropDict = streamChsDropDict
-        self.sr = maxExpectedSampleRate
-        #self.dataType = dataStreamInlet.info().type()
-        self.devCount = devCount # used for tracking which device is sending data to feature extractor
-        
-    def run(self):
-        posCount = 0
-        chCount = self.dataStreamInlet.info().channel_count()
-        maxTime = (self.globalEpochSettings.tmin + self.globalEpochSettings.tmax)
-        if len(self.customEpochSettings.keys())>0:
-            if  max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings]) > maxTime:
-                maxTime = max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings])
-        fifoLength = int(self.sr*maxTime)
-        window_end_time = 0
-        dataFIFOs = [deque(maxlen=fifoLength) for ch in range(chCount - len(self.streamChsDropDict))]
-        while not self.closeEvent.is_set():
-            sample, timestamp = self.dataStreamInlet.pull_sample(timeout = 1)
-            if sample != None:
-                for index in sorted(self.streamChsDropDict, reverse=True):
-                    del sample[index] # remove the desired channels from the sample
-                for i,fifo in enumerate(dataFIFOs):
-                    fifo.append((timestamp, sample[i]))
-                if self.trainTestEvent.is_set(): # We're training!
-                    if self.startCounting: # we received a marker
-                        posCount += 1
-                        if posCount >= self.desiredCount:  # enough samples are in FIFO, chop up and put in dataqueue
-                            if len(self.customEpochSettings.keys())>0: #  custom marker received
-                                if self.customEpochSettings[self.currentMarker].splitCheck: # slice epochs into overlapping time windows
-                                    window_length = self.customEpochSettings[self.currentMarker].windowLength
-                                    window_overlap = self.customEpochSettings[self.currentMarker].windowOverlap
[…]
