### 2023-09-03 3e31dfa
Update PseudoDevice.rst
