### 2023-09-03 093a56b
Update PseudoDeviceController.rst
