### 2023-09-03 96675de
Update PseudoDeviceController.rst
