+++ docs/api/PseudoDeviceController.rst
@@ -1,4 +1,4 @@
-pybci.Configuration.PseudoDevice import PseudoDeviceController
+pybci.Utils.PseudoDevice import PseudoDeviceController
 ======================
 
 .. class:: PseudoDeviceController(execution_mode='process', *args, **kwargs)
