### 2023-09-03 7d1ef0a
pseudodevice v1.1.0!
Creation of pseudo device and integration into examples.

Documentation to be updated over coming days
