+++ pybci/Examples/PseudoLSLStreamGenerator/mainSend.py
@@ -1,207 +1,0 @@
-################ main.py ########################
-# Main file for practice lsl markers            #
-# Please note! sample rate is not exact,        #
-# expect some drop over time!                   #
-# Written by Liam Booth 14/05/2023              #
-#################################################
-from PyQt5 import QtWidgets, QtCore
-from PyQt5.QtGui import QFont
-from PyQt5.QtCore import QTimer
-import pylsl
-import random, time
-from collections import deque
-from pyqtgraph import PlotWidget, plot
-import pyqtgraph as pg
-import numpy as np
-import threading
-
-def precise_sleep(duration):
-    end_time = time.time() + duration
-    while time.time() < end_time:
-        pass
-
-class PsuedoEMGDataConfig:
-    duration = 1.0 
-    noise_level = 0.1
-    amplitude = 0.2
-    frequency = 1.0
-
-class Window(QtWidgets.QWidget):
-    commandStrings = ["Marker1", "Marker2", "Marker3", "Marker4", "Marker5"]
-    commandDataConfigs = [PsuedoEMGDataConfig(), PsuedoEMGDataConfig(), PsuedoEMGDataConfig(), PsuedoEMGDataConfig(),PsuedoEMGDataConfig()]
-    baselineConfig  = PsuedoEMGDataConfig()
-    currentMarker = "Marker1"
-    totchs = 8
-    x = deque([x/250 for x in range(250*5)]) #np.zeros(250*5) # this can be edited to change width of x axis 
-    y = [deque([0 for x in range(250*5)])  for y in range(totchs)]
-    markerOccurred = False
-    signal = [deque([0 for x in range(250*5)])  for y in range(totchs)] 
-    signalcount = 0
-    chunkCount = 0
-    sampleRate = 250
-    #de = deque(maxlen=250/10) maybe make a timestap deque to keep consistent timing?
-
-    def __init__(self, *args, **kwargs):
-        super(Window, self).__init__(*args, **kwargs)
-        #self.showMaximized()
-        self.lock = threading.Lock()  # Lock for thread safety
-        # setting up different signal varables
-        self.commandDataConfigs[0].amplitude = 1
-        self.commandDataConfigs[0].noise_level = 1
-        self.commandDataConfigs[0].frequency = 2
-        self.commandDataConfigs[1].amplitude = 1
-        self.commandDataConfigs[1].noise_level = 0.5
-        self.commandDataConfigs[1].frequency = 5
-        self.commandDataConfigs[2].amplitude = 2
-        self.commandDataConfigs[2].noise_level = 0.3
-        self.commandDataConfigs[2].frequency = 15
-        self.commandDataConfigs[3].amplitude = 2
-        self.commandDataConfigs[3].noise_level = 0.6
-        self.commandDataConfigs[3].frequency = 0.2
-        self.commandDataConfigs[4].amplitude = 3
-        self.commandDataConfigs[4].noise_level = 0.1
-        self.commandDataConfigs[4].frequency = 2
-
-        self.button = QtWidgets.QPushButton('Start Signal Test')
-        self.button.clicked.connect(self.BeginTest)
-        self.button2 = QtWidgets.QPushButton('Change Marker')
-        self.button2.clicked.connect(self.ChangeString)
-        self.button3 = QtWidgets.QPushButton('Send Marker')
-        self.button3.clicked.connect(self.SendMarker)
-        self.buttonBaseline = QtWidgets.QPushButton('Send Baseline')
-        self.buttonBaseline.clicked.connect(self.SendBaseline)
-        self.text = QtWidgets.QLabel("Command!")
-        self.text.setAlignment(QtCore.Qt.AlignCenter)
[…]
