+++ pybci/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.1.0-alpha'
+__version__ = '1.1.0'

[…]
