### 2023-09-03 8898de4
Update Pseudo_Device.rst
