### 2023-09-03 41618b7
Update Pseudo_Device.rst
