### 2023-09-03 fdb54c0
Update PyBCI.rst
