+++ pybci/Examples/README.md
@@ -2,15 +2,14 @@
 
 This folder holds multiple scripts illustrating the functions and configurations available within the PyBCI package.
 
-NOTE: All the examples shown that are not in a dedicated folder work with the PsuedoLSLStreamGenerator found in [PsuedoLSLSreamGenerator/mainSend.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/PsuedoLSLStreamGenerator/mainSend.py), if using with own LSL capable hardware you may need to adjust the scripts accordingly.
+NOTE: all the examples shown that are not in a dedicated folder work with the createPseudoDevice variable set to True when instantiating PBCI(). If using with own LSL capable hardware you may need to adjust the scripts accordingly, namely set `createPseudoDevice=False`.
 
 | Example File | Description |
 |--------------|-------------|
 | [ArduinoHandGrasp/](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/) | Folder contains an LSL marker creator in [MarkerMaker.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/MarkerMaker.py) using PyQt5 as an on screen text stimulus, illustrates how LSL markers can be used to train. [ServoControl.ino](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/ServoControl/ServoControl.ino) is designed for an arduino uno which controls 5 servo motors, each of which control the position of an indidividual finger for a 3D printed hand which can be controlled via serial commands. There is also a [Myoware Muscle Sensor](https://myoware.com/products/muscle-sensor/) attached to analog pin A0 being read continuously over the serial connection. [ArduinoToLSL.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/ArduinoToLSL.py) is used to send and receive serial data to and from the arduino, whilst pushing the A0 data to an LSL outlet which is classified in [testArduinoHand.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/testArduinoHand.py), whilst simultaneously receiving a marker stream from testArduinoHand.py to inform which hand position to do.|
-| [PsuedoLSLSreamGenerator/mainSend.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/PsuedoLSLStreamGenerator/mainSend.py) | Generates multiple channels on a given stream type at a given sample rate. A baseline signal is generated on an LSL stream outlet and a PyQt button can be pressed to signify this signal on a separate LSL marker stream. The signal can be altered by 5 distinct markers for a configurable amount of time, allowing the user to play with various signal patterns for clasification. NOTE: Requires `PyQt5` and `pyqtgraph` installs for data viewer. |
+| [MultimodalPupilLabsEEG/](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/MultimodalPupilLabsEEG) | Advanced example illustrating two devices, pupil labs gaze device stream wth custom feature extractor class and Hull University ioBio EEG device, specifically channels Fp1 and Fp2. The PupilLabsRightLeftEyeClose folder has a youtube video illustrating the multimodal example in action. |
[…]
