### 2023-09-03 eefb6f6
Update Pseudo_Device.rst
