### 2023-09-03 7f846d0
Update Pseudo_Device.rst
