### 2023-09-03 6df7521
Update Pseudo_Device.rst
