### 2023-09-03 #5706 @LMBooth
Took longer then expected, but i've managed to update the examples and documenttion regarding the newly implemented pseudo device class which used in all the primary examples to allow the scripts to run optionally run automiatically if the user has no LSL hardware.

If anyone has nay issues with using this new feature please do let me know or open an issue in the git!

Thanks again @anilbey!
