### 2023-09-11 9c6c5e7
Update Pseudo_Device.rst
