+++ docs/BackgroundInformation/Pseudo_Device.rst
@@ -11,6 +11,8 @@ The PseudoDevice class holds marker information and generates signal data based 
 
 The PseudoDeviceController can have the string "process" or "thread" set to decide whether the pseudo device should be a multiprocessed or threaded operation respectively, by default it is set to "process", then passes all the same configuration arguments to PseudoDevice.
 
+Any generic LSLViewer can be used to view the generated data, example [viewers found here](https://labstreaminglayer.readthedocs.io/info/viewers.html).
+
 .. _configuring-pseudo-device:
 Configuring the Pseudo Device
 =========================================================
