### 2023-09-11 0415d64
Update Pseudo_Device.rst
