### 2023-09-11 58d2200
Update Pseudo_Device.rst
