### 2023-09-11 f04148d
Update Pseudo_Device.rst
