### 2023-09-13 LMBooth/pybci#3 @jsheunis
JOSS: paper title seems redundant
> PyBCI: A Python Package for Brain-Computer Interface (BCI) Design/An Open Source Brain-Computer Interface Framework in Python

These two titles seem like different phrasing of the same idea/tool. Is there a reason why both are provided in the title field? I would propose using a single title.

Ping https://github.com/openjournals/joss-reviews/issues/5706
