### 2023-09-13 LMBooth/pybci#5 @jsheunis
JOSS: several paper comments
- The summary section states "integration into other systems..." - which other systems, could this statement be a bit more specific?
- The software functionality section states "end-to-end solution for BCI research" but also ""no additional visualization...". These statements seem contradictory, since I would think an end to end solution includes everything one needs to conduct the research. Relatedly, the paper states researchers can focus on their experiments and don't need extensive software development skills. If they have to integrate it with visualization tools or other systems in order to have the whole system function correctly, this statement might not be completely true.

Ping https://github.com/openjournals/joss-reviews/issues/5706
