### 2023-09-13 LMBooth/pybci#6 @jsheunis
PyPI package name
I could only find your package on PyPI: https://pypi.org/search/?q=pybci

Is the conflict with a different package still a problem? If not, I would suggest renaming the package to `pybci`, because the current installation process is slightly confusing with the double "install": https://github.com/LMBooth/pybci/blob/main/docs/BackgroundInformation/Introduction.rst?plain=1#L12-L14

If you can't rename the package, perhaps consider naming it `pybci-installer` instead.

https://github.com/openjournals/joss-reviews/issues/5706
