### 2023-09-13 LMBooth/pybci#6 @LMBooth
Hi @jsheunis, thanks for bringing this up, the conflict is still a problem with no response as of yet.

I could rename to pybci-installer if you think this would be preferable and more clear for users, i'll have a look at formalising this tonight when i get home from work.
