### 2023-09-13 LMBooth/pybci#3 @LMBooth
Thanks for catching this @jsheunis , i think they are both titles i was playing with and hadn't got round to deleting one!

I'll amend now and if you agree its acceptable we can close the issue.
