### 2023-09-13 LMBooth/pybci#6 @jsheunis
actually, on second thought, I don't think my suggestion will be clearer for users. But I also think the current name isn't clear either. The "install" part suggest that you are downloading an installer of some kind, which then needs to be executed in order to install the package. In reality, it is the actual package that's being downloaded and installed. In the end it's your decision, but you could perhaps try a name with some extra characters (perhaps `py-bci` i.e. with a hyphen, or something else that is just not as confusing as `install`).

And then the associated docs would need to be updated.
