### 2023-09-13 LMBooth/pybci#8 @jsheunis
Installation process suggestions
I suggest adding the following to the readme and to the docs:

#### The use of a virtual environment for installing the package, e.g.:

```
python -m venv my_env
source my_env/bin/activate
```
followed by `pip install`

#### Installing from the source on github, for those wanting latest un-released features or for developers:

```
pip install git+https://github.com/LMBooth/pybci.git
```

or

```
git clone https://github.com/LMBooth/pybci.git
cd pybci
pip install -e .
```

Ping https://github.com/openjournals/joss-reviews/issues/5706
