### 2023-09-13 LMBooth/pybci#8 @jsheunis
I also think, even though the minimum Python requirement is implicit in the setup.py code, it would be useful to state this explicitly in the docs, ideally in the Installation section.
