+++ JOSS/paper.md
@@ -29,9 +29,6 @@ bibliography: paper.bib
 # Summary:
 PyBCI is a comprehensive, open-source Python framework developed to facilitate brain-computer interface (BCI) research. It encompasses data acquisition, data labelling, feature extraction, and machine learning. PyBCI provides a streamlined, user-friendly platform for creating real-time BCI applications. The software uses the Lab Streaming Layer (LSL) [@lsl] protocol for the unified collection of time-series measurement that handles both the networking, time-synchronization and (near-) real-time access (supported LSL devices found here: https://labstreaminglayer.readthedocs.io/info/supported_devices.html). At least one LSL data stream is required and a single marker stream is used for labelling training data. PyTorch [@NEURIPS2019_9015], SciKit-learn [@scikit-learn] and TensorFlow [@tensorflow2015-whitepaper] are leveraged for applying machine learning classifiers. NumPy [@oliphant2006guide], SciPy [@2020SciPy-NMeth], and Antropy [@vallat_antropy_2023] are utilised for generic time and/or frequency feature extraction examples.
 
-
-PyBCI is a comprehensive, open-source Python framework developed to facilitate brain-computer interface (BCI) research. It encompasses data acquisition, data labelling, feature extraction, and machine learning. PyBCI provides a streamlined, user-friendly platform for conducting real-time BCI applications. The software uses the Lab Streaming Layer (LSL) [@lsl] protocol for data acquisition on various LSL enabled data streams. The LSL is a system for the unified collection of measurement time series in research experiments that handles both the networking (supported LSL devices found here: https://labstreaminglayer.readthedocs.io/info/supported_devices.html). At least one LSL data stream is required and a single marker stream is used for labelling training data. PyTorch [@NEURIPS2019_9015], SciKit-learn [@scikit-learn] and TensorFlow [@tensorflow2015-whitepaper] are leveraged for applying machine learning classifiers. NumPy [@oliphant2006guide], SciPy [@2020SciPy-NMeth], and Antropy [@vallat_antropy_2023] are utilised for generic time and/or frequency feature extraction examples.
-
 # Statement of Need:
 
 PyBCI puts emphasis on quick and easy customisation of applied time-series feature extraction techniques and machine learning models, live data classification, and integration into other systems. Remaining lightweight with solely Python packages and no additional visualisation and recording tools, it's assumed a user will have these tools available with their LSL hardware. 
