### 2023-09-13 LMBooth/pybci#8 @LMBooth
Thankyou for these recommendation @jsheunis , i've made them and shall now close this issue. Please drop a comment or re-open if you find any related issues.
