### 2023-09-13 LMBooth/pybci#5 @jsheunis
> I think the point i was trying to get at is that it can be integrated in to larger python projects, but i feel its a non sequitur for where i've placed it - Hopefully this is more clear and concise now.

Yup 👍

> I did think about adding data plotters and visualisation but thought there are plenty of packages already available for these purposes and it would help keep the package dependencies slimmer. How about instead of "PyBCI provides an end-to-end solution for BCI research." "PyBCI accelerates the pace of BCI research by streamlining data collection, processing, and model analysis."?

Sounds good to me.

> For the latter half of the second point, they don't necessarily need any software skills to integrate, if they have an automated lsl marker delivery from something like Psychopy, a pre-built data viewer like the one from brainvision or eeglab, an off the shelf device with an lsl port it's possible to get the classification right from pycbi in a command prompt. My thought process being someone from the psychology field will have more chance at getting something up and running then trying to write it from scratch. Do you think i should redact this statement? " It provides researchers with the tools necessary to conduct advanced BCI experiments without requiring extensive software development skills. "

Thanks for the clarification. I don't think you have to redact, but perhaps you could work the example you just gave into the text to illustrate the point you are trying to make?
