+++ docs/BackgroundInformation/Introduction.rst
@@ -5,7 +5,7 @@ Introduction
 Installation
 ===================
 
-For stable releases use: :code:`pip install install-pybci`
+For stable releases use: :code:`pip install pybci-package`
 
 For development versions use: :code:`pip install git+https://github.com/LMBooth/pybci.git` or 
 
@@ -26,7 +26,7 @@ Windows:
 
    python -m venv my_env
    .\my_env\Scripts\Activate
-   pip install install-pybci  # For stable releases
+   pip install pybci-package  # For stable releases
    # OR
    pip install git+https://github.com/LMBooth/pybci.git  # For development version
 
@@ -36,7 +36,7 @@ Linux/MaxOS:
 
    python3 -m venv my_env
    source my_env/bin/activate
-   pip install install-pybci  # For stable releases
+   pip install pybci-package  # For stable releases
    # OR
    pip install git+https://github.com/LMBooth/pybci.git  # For development version
 
