### 2023-09-15 LMBooth/pybci#6 @LMBooth
`py-bci` , `py.bci` were too similar to another package and i tried `py-bci-`, `-py-bci` which broke pypi naming conventions needing to start and end on a alpha-numeric, so instead i've gone for `pybci-package`, and updated the documentation accordingly, hopefully this is more intuitive.
