+++ JOSS/paper.bib
@@ -130,3 +130,41 @@ author={
 	journal = {IEEE Access},
 	doi = {10.1109/ACCESS.2021.3079992}
 }
+
+@article{OpenViBE,
+  author={Renard, Yann and Lotte, Fabien and Gibert, Guillaume and Congedo, Marco and Maby, Emmanuel and Delannoy, Vincent and Bertrand, Olivier and Lécuyer, Anatole},
+  journal={Presence}, 
+  title={OpenViBE: An Open-Source Software Platform to Design, Test, and Use Brain–Computer Interfaces in Real and Virtual Environments}, 
+  year={2010},
+  volume={19},
+  number={1},
+  pages={35-53},
+  doi={10.1162/pres.19.1.35}}
+
+@article{BCI2000,
+  author={Schalk, G. and McFarland, D.J. and Hinterberger, T. and Birbaumer, N. and Wolpaw, J.R.},
+  journal={IEEE Transactions on Biomedical Engineering}, 
+  title={BCI2000: a general-purpose brain-computer interface (BCI) system}, 
+  year={2004},
+  volume={51},
+  number={6},
+  pages={1034-1043},
+  doi={10.1109/TBME.2004.827072}}
+
+@article{BciPy,
+	author = {Tab Memmott and Aziz Koçanaoğulları and Matthew Lawhead and Daniel Klee and Shiran Dudy and Melanie Fried-Oken and Barry Oken},
+	title = {BciPy: brain–computer interface software in Python},
+	journal = {Brain-Computer Interfaces},
+	volume = {8},
+	number = {4},
+	pages = {137-153},
+	year  = {2021},
+	publisher = {Taylor & Francis},
+	doi = {10.1080/2326263X.2021.1878727},
+	URL = { 
+	        https://doi.org/10.1080/2326263X.2021.1878727
+	},
+	eprint = { 
+	        https://doi.org/10.1080/2326263X.2021.1878727
+	}
+}
