+++ JOSS/paper.md
@@ -27,13 +27,17 @@ bibliography: paper.bib
 ---
 
 # Summary:
-PyBCI is a comprehensive, open-source Python framework developed to facilitate brain-computer interface (BCI) research. It encompasses data acquisition, data labelling, feature extraction, and machine learning. PyBCI provides a streamlined, user-friendly platform for creating real-time BCI applications. The software uses the Lab Streaming Layer (LSL) [@lsl] protocol for the unified collection of time-series measurement that handles both the networking, time-synchronization and (near-) real-time access (supported LSL devices found here: https://labstreaminglayer.readthedocs.io/info/supported_devices.html). At least one LSL data stream is required and a single marker stream is used for labelling training data. PyTorch [@NEURIPS2019_9015], SciKit-learn [@scikit-learn] and TensorFlow [@tensorflow2015-whitepaper] are leveraged for applying machine learning classifiers. NumPy [@oliphant2006guide], SciPy [@2020SciPy-NMeth], and Antropy [@vallat_antropy_2023] are utilised for generic time and/or frequency feature extraction examples.
+PyBCI is an open-source Python framework designed to streamline brain-computer interface (BCI) research. It offers a comprehensive platform for real-time data acquisition, labeling, classification and analysis. PyBCI is compatible with a wide range of time-series hardware and software data sources, thanks to its integration with the Lab Streaming Layer (LSL) protocol [@lsl].
 
 # Statement of Need:
 
-PyBCI puts emphasis on quick and easy customisation of applied time-series feature extraction techniques, applied machine learning models, and live data classification. Remaining lightweight with solely Python packages and no additional visualisation and recording tools, it's assumed a user will have these tools available with their LSL hardware. 
+BCI research brings together diverse fields like neuroscience, engineering, and data science, requiring specialized tools for data acquisition, feature extraction, and real-time analysis. Existing solutions may offer partial functionalities or be cumbersome to use, slowing down the pace of innovation. PyBCI addresses these challenges by providing a flexible, Python-based platform aimed at researchers and developers in the BCI domain. Assuming a foundational understanding of Python, the software serves as a comprehensive solution for both academic and industry professionals.
 
-BCIs have the potential to revolutionise the way we interact with computers. However, BCI research is often limited by the lack of availability of open-source, comprehensive, and user-friendly software. PyBCI addresses this gap, offering a simple, flexible, and robust Python-based platform for conducting BCI research. It integrates seamlessly with the LSL for data acquisition and labelling, utilizing popular machine learning packages such as PyTorch, Scikit-learn and TensorFlow, employing Antropy, NumPy, and SciPy for an example feature extraction or optionally allowing the user to pass a custom feature extractor to process specific device data. This combination of software features allows researchers to focus on their experiments rather than software development, expediting the advancement of BCI technology.
+Designed to be lightweight and user-friendly, PyBCI emphasizes quick customization and integrates seamlessly with the Lab Streaming Layer (LSL) for data acquisition and labeling [@lsl]. The platform incorporates reputable machine learning libraries like PyTorch [@NEURIPS2019_9015], TensorFlow [@tensorflow2015-whitepaper], and Scikit-learn [@scikit-learn], as well as feature extraction tools such as Antropy [@vallat_antropy_2023], NumPy [@oliphant2006guide], and SciPy [@2020SciPy-NMeth]. This integration allows users to focus more on their research and less on software development. While a detailed comparison with other software solutions will follow in the 'State of the Field' section, PyBCI sets itself apart through its emphasis on ease of use and technological integration.
+
+# State of the Field:
+
[…]
