### 2023-09-18 LMBooth/pybci#9 @jsheunis
DOC: Introduction section seems unintuitive
The Introduction section of the docs seems a bit unintuitive. To me, "Introduction" is more about an overview of the tool, and not installation and usage. Currently it contains "Installation", "Example Usage", and "What is PyBCI?". I suggest reordering it at the top level to a "Get started" section, which contains all the information a user would need to get started. This would be installation and example usage. Then I would move the "What is PyBCI?" into its own section, maybe "Introduction" or "Overview".

Ping https://github.com/openjournals/joss-reviews/issues/5706
