### 2023-09-18 LMBooth/pybci#10 @jsheunis
DOC: Simple Implementation section suggestions
This section is currently too minimal to bring users up to speed on what they should be doing to get started. The only text is "For example". What should a user do? Should they always first write a python script to get started? should they copy the sample script and update it to their own parameters? What exactly should be in place on the hardware level before they implement a script?

This section is also a good place to mention and link to the pseudo-device description, so that users can know that they don't necessarily need a hardware setup in order to start using pybci.

https://github.com/openjournals/joss-reviews/issues/5706
