### 2023-09-18 LMBooth/pybci#9 @jsheunis
I would also add a place for prerequisites before the installation. Then the paragraph about LSL for non-windows systems can move there, and you can also mention the minimum python version.
