### 2023-09-18 LMBooth/pybci#11 @jsheunis
DOC: statement of need
According to the JOSS requirements, the documentation should contain the following:

> Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Please make sure this goes into the overview / "What is PyBCI" section.

Ping https://github.com/openjournals/joss-reviews/issues/5706
