### 2023-09-18 LMBooth/pybci#12 @jsheunis
ENH: consider implementing a command line interface
The first thing I tried after installing pybci was running `pybci` via the command line, but this command was not recognized. Perhaps I have a bias towards CLIs, but I think it would be useful to use argparse to implement a simple command or two. At the least, I think it would be useful to be able to type `pybci --help` in order to get a description of what the tool is and what it can do.

In addition, this is up to your discretion, but I think it would be useful to have some sort of toy example that can be run via the command line. E.g. `pybci hello-world`, which then starts the pseudo-device and runs pybci with some sort of minimal visualization.

If something like this is implemented, a relevant description could also form part of the "Getting started" section in the docs.

Ping https://github.com/openjournals/joss-reviews/issues/5706
