### 2023-09-18 LMBooth/pybci#13 @jsheunis
DOC: inconsistent syntax highlighting
I've noticed on several pages of the docs that class names or python expressions or similar are not always rendered in the expected syntax highlighting. E.g. className vs ``className`` or X = 1 vs ``X = 1``. Please run through all the docs again to make sure you use this consistently.

Ping https://github.com/openjournals/joss-reviews/issues/5706
