### 2023-09-18 LMBooth/pybci#14 @jsheunis
DOC: examples
This section of the documentation is too minimal IMO, and its' a section that users will be likely to visit.

Generally, the same comments I made in https://github.com/LMBooth/pybci/issues/10 also apply here, i.e. what does a user have to do to get any individual example running? Just linking them to an online script does not provide an intuitive user experiencing.

In addition:
- The table does not seem like a good way to represent the different examples, I had to scroll horizontally in order to be able to read each paragraph. I suggest putting all examples into their own subsection
- You could consider adding a tag (e.g. using https://shields.io/) to each example that shows whether it can be used with the pseudodevice or not. If this is useful, then these tags can be used for other categorizations of examples as well.

Regarding the statement:
> if using with own LSL capable hardware you may need to adjust the scripts accordingly, namely set createPseudoDevice=False.

I think it's suboptimal if users have to change scripts in order to make examples work. I think passing an argument to the script itself would make it easier for them. This is another indication that a command line API would be useful (see https://github.com/LMBooth/pybci/issues/12).

Ping https://github.com/openjournals/joss-reviews/issues/5706
