### 2023-09-18 LMBooth/pybci#15 @jsheunis
TEST: there are no tests or CI implementation
There aren't any unit or integration tests, e.g. with [pytest](https://docs.pytest.org/en/7.4.x/) and no continuous integration implementation, e.g. with [appveyor](https://www.appveyor.com/). Adding these components are essential to ensure that the package runs successfully on multiple platforms and that new code contributions don't cause unintended breakage. It is also useful and encouraged for developers to run tests on their own systems before creating a PR with their contributions to your code base.

Ping https://github.com/openjournals/joss-reviews/issues/5706
