### 2023-09-18 c250d7a
Update Pseudo_Device.rst
Remove typo
