### 2023-09-18 LMBooth/pybci#17 @jsheunis
DOC: Community guidelines
Ideally, the software should contain guidelines for contributors to allow them to contribute productively to the code, docs or any other aspect of the package. This would include info such as:
- how/where to ask questions
- what process to follow to contribute code
- how to run tests
- anything else worth knowing when contributing

This usually takes the form of a CONTRIBUTING.md document at the root of the repo as well as a section with the same content in the docs.

Ping https://github.com/openjournals/joss-reviews/issues/5706
