+++ build/lib/pybci/ThreadClasses/OptimisedDataReceiverThread.py
@@ -1,0 +1,139 @@
+import threading, time
+from collections import deque
+import itertools
+import numpy as np
+from bisect import bisect_left
+
+class OptimisedDataReceiverThread(threading.Thread):
+    """Responsible for receiving data from accepted LSL outlet, slices samples based on tmin+tmax basis, 
+    starts counter for received samples after marker is received in ReceiveMarker. Relies on timestamps to slice array, 
+    suspected more computationally intensive then synchronous method.
+    """
+    markerReceived = False
+    currentMarker = ""
+    def __init__(self, closeEvent, trainTestEvent, dataQueueTrain,dataQueueTest, dataStreamInlet,  
+                 customEpochSettings, globalEpochSettings,devCount,  streamChsDropDict = [], maxExpectedSampleRate=100):
+        # large maxExpectedSampleRate can incur marker drop and slow procesing times for slicing arrays
+        super().__init__()
+        self.trainTestEvent = trainTestEvent
+        self.closeEvent = closeEvent
+        self.dataQueueTrain = dataQueueTrain
+        self.dataQueueTest = dataQueueTest
+        self.dataStreamInlet = dataStreamInlet
+        self.customEpochSettings = customEpochSettings
+        self.globalEpochSettings = globalEpochSettings
+        self.streamChsDropDict = streamChsDropDict
+        self.sr = maxExpectedSampleRate
+        self.devCount = devCount # used for tracking which device is sending data to feature extractor
+        
+    def run(self):
+        chCount = self.dataStreamInlet.info().channel_count()
+        maxTime = (self.globalEpochSettings.tmin + self.globalEpochSettings.tmax)
+        if len(self.customEpochSettings.keys()) > 0: # min time used for max_samples and temp array,  maxTime used for longest epochs and permanentDataBuffers
+            if max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings]) > maxTime:
+                maxTime = max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings])
+        if self.globalEpochSettings.splitCheck:
+            window_length = self.globalEpochSettings.windowLength
+            window_overlap = self.globalEpochSettings.windowOverlap
+        else:
+            window_length = self.globalEpochSettings.tmin+self.globalEpochSettings.tmax
+        minfifoLength = int(self.sr * window_length * (1-self.globalEpochSettings.windowOverlap)) #  sets global window length with overlap as factor for minimum delay in test mode
+        dataBuffers = np.zeros((minfifoLength,chCount))
+        chs_to_drop = np.ones(chCount, dtype=bool)
+        chs_to_drop[self.streamChsDropDict] = False
+        fifoLength = int(self.sr * (maxTime+20)) # adds twenty seconds to give more timestamps when buffering  (assuming devices dont timeout for longer then 20.0 seconds, migth be worth making configurable)
+        permanentDataBuffers = np.zeros((fifoLength, chCount - len(self.streamChsDropDict)))
+        permanentTimestampBuffer = np.zeros(fifoLength)
+        next_window_time = 0 # sets testing mode window time, duration based on windowlength and overlap
+        while not self.closeEvent.is_set():
+            _, timestamps = self.dataStreamInlet.pull_chunk(timeout=0.0, max_samples=dataBuffers.shape[0], dest_obj=dataBuffers) # optimised method of getting data to pull_sample, dest_obj saves memory re-allocation
+            if timestamps:
+                if len(self.streamChsDropDict) == 0:
+                    dataBufferView = dataBuffers[:len(timestamps), :] # [:, :len(timestamps)]
+                else:
[…]
