+++ build/lib/pybci/Configuration/FeatureSettings.py
@@ -1,0 +1,15 @@
+class GeneralFeatureChoices:
+    psdBand = False
+    appr_entropy = False
+    perm_entropy = False
+    spec_entropy = False
+    svd_entropy = False
+    samp_entropy = False
+    rms = True
+    meanPSD = True
+    medianPSD = False
+    variance = False
+    meanAbs = False
+    waveformLength = False
+    zeroCross = False
+    slopeSignChange = False
