+++ build/lib/pybci/version.py
@@ -1,2 +1,1 @@
 __version__ = '1.2.0-a0'
-

[…]
