+++ build/lib/pybci/Utils/__init__.py
@@ -1,4 +1,0 @@
-#from .Classifier import Classifier
-#from .FeatureExtractor import FeatureExtractor
-#from .LSLScanner import LSLScanner
-
