+++ pybci/cli.py
@@ -1,46 +1,50 @@
 import argparse
-from Tests.testSimple import main as testSimple
-from Tests.testSklearn import main as testSklearn
-from Tests.testTensorflow import main as testTensorflow
-from Tests.testPyTorch import main as testPyTorch
+from .Tests.testSimple import main as testSimple
+from .Tests.testSklearn import main as testSklearn
+from .Tests.testTensorflow import main as testTensorflow
+from .Tests.testPyTorch import main as testPyTorch
 
 def main():
-    parser = argparse.ArgumentParser(description='PyBCI: A Brain-Computer Interface package.')
+    parser = argparse.ArgumentParser(description='PyBCI: A Brain-Computer Interface package. Visit https://pybci.readthedocs.io/en/latest/ for more information!')
     
     subparsers = parser.add_subparsers(title='Commands', description='Available example commands')
     
     testSimple_parser = subparsers.add_parser('testSimple', help='Runs simple setup where sklearn support-vector-machine is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testSimple.py in the examples folder.')
-    testSimple_parser.add_argument('--createPseudoDevice',default=True,  type=bool, help='Set to True or False to enable or disable pseudo device creation.')
+    testSimple_parser.add_argument('--create_pseudo_device',default=True,  type=bool, help='Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.')
     testSimple_parser.add_argument('--min_epochs_train', default=4, type=int, help='Minimum epochs to collect before model training commences, must be less than, min_epochs_test. If less than min_epochs_test defaults to min_epochs_test+1.')
     testSimple_parser.add_argument('--min_epochs_test', default=10, type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSimple_parser.set_defaults(func=testSimple)
     
     testSklearn_parser = subparsers.add_parser('testSklearn', help='Sklearn multi-layer perceptron is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testSklearn.py in the examples folder.')
-    testSklearn_parser.add_argument('--createPseudoDevice',default=True, type=bool, help='Set to True or False to enable or disable pseudo device creation.')
+    testSklearn_parser.add_argument('--create_pseudo_device',default=True, type=bool, help='Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.')
     testSklearn_parser.add_argument('--min_epochs_train', default=4,type=int, help='Minimum epochs to collect before model training commences, must be less than, min_epochs_test. If less than min_epochs_test defaults to min_epochs_test+1.')
     testSklearn_parser.add_argument('--min_epochs_test', default=10,type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSklearn_parser.set_defaults(func=testSklearn)
     
     testTensorflow_parser = subparsers.add_parser('testTensorflow', help='Tensorflow GRU is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testTensorflow.py in the examples folder.')
-    testTensorflow_parser.add_argument("--createPseudoDevice", default=True, type=bool, help="Set to True or False to enable or disable pseudo device creation.")
+    testTensorflow_parser.add_argument("--create_pseudo_device", default=True, type=bool, help="Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.")
[…]
