+++ pybci/version.py
@@ -1,2 +1,2 @@
-__version__ = '1.2.0-a0'
+__version__ = '1.2.0'
 
