+++ pybci/cli.py
@@ -1,8 +1,17 @@
-import argparse
+import argparse, time
 from .Tests.testSimple import main as testSimple
 from .Tests.testSklearn import main as testSklearn
 from .Tests.testTensorflow import main as testTensorflow
 from .Tests.testPyTorch import main as testPyTorch
+
+from .Utils.PseudoDevice import PseudoDeviceController
+
+
+def RunPseudo():
+    pseudoDevice = PseudoDeviceController()
+    pseudoDevice.BeginStreaming()
+    while True:
+        time.sleep(1)
 
 def main():
     parser = argparse.ArgumentParser(description='PyBCI: A Brain-Computer Interface package. Visit https://pybci.readthedocs.io/en/latest/ for more information!')
@@ -12,19 +21,19 @@ def main():
     testSimple_parser = subparsers.add_parser('testSimple', help='Runs simple setup where sklearn support-vector-machine is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testSimple.py in the examples folder.')
     testSimple_parser.add_argument('--create_pseudo_device',default=True,  type=bool, help='Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.')
     testSimple_parser.add_argument('--min_epochs_train', default=4, type=int, help='Minimum epochs to collect before model training commences, must be less than, min_epochs_test. If less than min_epochs_test defaults to min_epochs_test+1.')
-    testSimple_parser.add_argument('--min_epochs_test', default=10, type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
+    testSimple_parser.add_argument('--min_epochs_test', default=14, type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSimple_parser.set_defaults(func=testSimple)
     
     testSklearn_parser = subparsers.add_parser('testSklearn', help='Sklearn multi-layer perceptron is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testSklearn.py in the examples folder.')
     testSklearn_parser.add_argument('--create_pseudo_device',default=True, type=bool, help='Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.')
     testSklearn_parser.add_argument('--min_epochs_train', default=4,type=int, help='Minimum epochs to collect before model training commences, must be less than, min_epochs_test. If less than min_epochs_test defaults to min_epochs_test+1.')
-    testSklearn_parser.add_argument('--min_epochs_test', default=10,type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
+    testSklearn_parser.add_argument('--min_epochs_test', default=14,type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSklearn_parser.set_defaults(func=testSklearn)
     
     testTensorflow_parser = subparsers.add_parser('testTensorflow', help='Tensorflow GRU is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testTensorflow.py in the examples folder.')
     testTensorflow_parser.add_argument("--create_pseudo_device", default=True, type=bool, help="Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.")
[…]
