### 2023-09-23 LMBooth/pybci#12 @LMBooth
Excellent recommendation @jsheunis , i've now implemented 4 simple tests that can be run right from the command line. The pseudodevice can be initiated as a command too. I'll close this issue for now and keep #14 open until i finish the complementary documentation.
