### 2023-09-23 LMBooth/pybci#14 @LMBooth
Note to self: add documentation for cli commands
