### 2023-09-23 40b5a33
Update PyBCI.rst
