+++ docs/api/PyBCI.rst
@@ -1,144 +1,61 @@
 PyBCI
 =====
 
-.. class:: PyBCI(dataStreams=None, markerStream=None, streamTypes=None, markerTypes=None, loggingLevel=Logger.INFO, globalEpochSettings=GlobalEpochSettings(), customEpochSettings={}, streamChsDropDict={}, streamCustomFeatureExtract={}, minimumEpochsRequired=10, createPseudoDevice=False, pseudoDeviceArgs=None, clf=None, model=None, torchModel=None)
+.. class:: PyBCI(dataStreams=None, markerStream=None, streamTypes=None, markerTypes=None, loggingLevel=Logger.INFO, globalEpochSettings=Global EpochSettings(), custom EpochSettings={}, streamChsDropDict={}, streamCustomFeatureExtract={}, minimum EpochsRequired=10, createPseudoDevice=False, pseudoDeviceArgs=None, clf=None, model=None, torchModel=None)
 
-   The `PyBCI` object is the main controller for interfacing with all relevant threads. When initialised sets up the main operation of the BCI and can be queried for relevant information.
+   The PyBCI object is the main controller for interfacing with all relevant threads. When initialised, it sets up the main operation of the BCI and can be queried for relevant information.
 
-   **Parameters:**
+   :param dataStreams: list(str) or None: Allows the user to set custom acceptable EEG stream definitions. Defaults to `streamTypes` scan if `None`.
+   :param markerStream: str or None: Allows the user to set custom acceptable Marker stream definition. Defaults to `markerTypes` scan if `None`.
+   :param streamTypes: list(str) or None: Allows the user to set custom acceptable EEG type definitions, ignored if `dataStreams` is not `None`.
+   :param markerTypes: list(str) or None: Allows the user to set custom acceptable Marker type definitions, ignored if `markerStream` is not `None`.
+   :param loggingLevel: string: Sets PyBCI print level. Options are 'INFO', 'WARNING', 'TIMING', and 'NONE'.
+   :param global EpochSettings: Global EpochSettings: Sets global timing settings for epochs.
+   :param custom EpochSettings: dict: Sets individual timing settings for epochs.
+   :param streamChsDropDict: dict: Specifies which channels to drop for each data stream.
+   :param streamCustomFeatureExtract: dict: Allows a custom feature extractor class for each data stream.
+   :param minimum EpochsRequired: int: Minimum number of required epochs before model fitting begins.
+   :param createPseudoDevice: bool: If True, auto-generates LSL marker and LSL data.
+   :param pseudoDeviceArgs: dict: Dictionary of arguments to initialize pseudo device.
+   :param clf: sklearn.base.ClassifierMixin or None: Allows custom Sklearn model to be passed.
+   :param model: tf.keras.model or None: Allows custom TensorFlow model to be passed.
+   :param torchModel: custom def or None: Custom torch function should be passed with 4 inputs.
 
-.. _pybci-datastreams:
+   .. note::
+      For more information on epoch settings, see `Global EpochSettings()` and `Individual EpochSettings()`.
 
-.. py:attribute:: dataStreams
-    :type: list(str) or None
-    :value: Allows the user to set custom acceptable EEG stream definitions. If `None`, it defaults to `streamTypes` scan.
+   .. py:method:: __enter__()
 
-.. _pybci-markerstream:
+      Connects to the BCI. Same as __init__.
 
-.. py:attribute:: markerStream
-    :type: str or None
-    :value: Allows the user to set custom acceptable Marker stream definition. If `None`, it defaults to `markerTypes` scan.
+   .. py:method:: __exit__(exc_type, exc_val, exc_tb)
 
-.. _pybci-streamtypes:
+      Stops all threads of the BCI.
 
-.. py:attribute:: streamTypes
-    :type: list(str) or None
[…]
