### 2023-09-24 57dd0d5
folder reshuffle + pytest
intiial pytest and appveyor integration as well as folder reshuffle for CliTests
