+++ setup.py
@@ -105,7 +105,7 @@ setup(
         "antropy>=0.1.6",
         "tensorflow>=2.13.0",
         "scikit-learn>=1.3.0",
-        "torch>=2.0.1"
+        "torch>=2.0.0"
     ],
     # You can just specify the packages manually here if your project is
     # simple. Or you can use find_packages().
