+++ appveyor.yml
@@ -12,7 +12,7 @@ install:
   - "python -m pip install --upgrade pip"  # Upgrade pip
   - "python --version"  # Print Python version for debugging
   - "pip --version"  # Print pip version for debugging
-  - "pip install torch"  # Try installing torch directly
+  - "pip install https://files.pythonhosted.org/packages/d0/c8/f0dc8642e3ce0a3ae5f05e5149ab9df5375d569294f7be9a1ab1d95a1d76/torch-2.0.1-cp311-cp311-win_amd64.whl#sha256=dbd68cbd1cd9da32fe5d294dd3411509b3d841baecb780b38b3b7b06c7754434"  # Try installing torch directly
   - "python setup.py install"  # Install your package
   - "pip install pytest"
 
