+++ appveyor.yml
@@ -2,10 +2,10 @@ version: 1.0.{build}
 image: Visual Studio 2019  # Choose an image that meets your project needs
 environment:
   matrix:
-    - PYTHON: "C:\\Python311"
-    - PYTHON: "C:\\Python310"
-    - PYTHON: "C:\\Python39"
-    - PYTHON: "C:\\Python38"
+    - PYTHON: "C:\\Python311-x64"
+    - PYTHON: "C:\\Python310-x64"
+    - PYTHON: "C:\\Python39-x64"
+    - PYTHON: "C:\\Python38-x64"
 
 install:
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
