+++ appveyor.yml
@@ -1,5 +1,5 @@
 version: 1.0.{build}
-image: Visual Studio 2019  # Choose an image that meets your project needs
+image: Visual Studio 2019
 environment:
   matrix:
     - PYTHON: "C:\\Python311-x64"
@@ -9,19 +9,19 @@ environment:
 
 install:
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
-  - "python -m pip install --upgrade pip"  # Upgrade pip
-  - "python --version"  # Print Python version for debugging
-  - "pip --version"  # Print pip version for debugging
-    # Download liblsl library (replace the URL with the correct one for your needs)
+  - "python -m pip install --upgrade pip"
+  - "python --version"
+  - "pip --version"
+  # Download liblsl library (replace the URL with the correct one for your needs)
   - "curl -o liblsl.zip https://github.com/sccn/liblsl/releases/download/your_version/liblsl.zip"
   # Extract and move (replace paths as necessary)
   - "unzip liblsl.zip -d %PYTHON%\\Lib\\site-packages\\"
   # Install other dependencies
-  - "pip install https://files.pythonhosted.org/packages/d0/c8/f0dc8642e3ce0a3ae5f05e5149ab9df5375d569294f7be9a1ab1d95a1d76/torch-2.0.1-cp311-cp311-win_amd64.whl#sha256=dbd68cbd1cd9da32fe5d294dd3411509b3d841baecb780b38b3b7b06c7754434"  # Try installing torch directly
-  - "python setup.py install"  # Install your package
+  - "pip install https://files.pythonhosted.org/packages/d0/c8/f0dc8642e3ce0a3ae5f05e5149ab9df5375d569294f7be9a1ab1d95a1d76/torch-2.0.1-cp311-cp311-win_amd64.whl#sha256=dbd68cbd1cd9da32fe5d294dd3411509b3d841baecb780b38b3b7b06c7754434"
+  - "python setup.py install"
   - "pip install pytest"
 
-build: off  # Disabling the build step as it's a Python project
+build: off
 
 test_script:
   - "pytest Tests/"
