+++ appveyor.yml
@@ -13,7 +13,7 @@ install:
   - "python --version"
   - "pip --version"
   # Download liblsl library (replace the URL with the correct one for your needs)
-  - "curl -o liblsl.zip https://github.com/sccn/liblsl/releases/download/your_version/liblsl.zip"
+  - "curl -o liblsl.zip https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip"
   # Extract and move (replace paths as necessary)
   - "unzip liblsl.zip -d %PYTHON%\\Lib\\site-packages\\"
   # Install other dependencies
