+++ appveyor.yml
@@ -21,7 +21,7 @@ install:
   # Set the PYLSL_LIB environment variable
   - "set PYLSL_LIB=%PYTHON%\\Lib\\site-packages\\lsl.dll"
   # Install other dependencies
-  - "pip install https://files.pythonhosted.org/packages/d0/c8/f0dc8642e3ce0a3ae5f05e5149ab9df5375d569294f7be9a1ab1d95a1d76/torch-2.0.1-cp311-cp311-win_amd64.whl#sha256=dbd68cbd1cd9da32fe5d294dd3411509b3d841baecb780b38b3b7b06c7754434"
+  #- "pip install https://files.pythonhosted.org/packages/d0/c8/f0dc8642e3ce0a3ae5f05e5149ab9df5375d569294f7be9a1ab1d95a1d76/torch-2.0.1-cp311-cp311-win_amd64.whl#sha256=dbd68cbd1cd9da32fe5d294dd3411509b3d841baecb780b38b3b7b06c7754434"
   - "python setup.py install"
   - "pip install pytest"
 
