+++ appveyor.yml
@@ -10,7 +10,6 @@ environment:
 install:
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
   - "python -m pip install --upgrade pip"
-  - "pip install --upgrade urllib3>=2.0.5"  # Add this line
   - "python --version"
   - "pip --version"
   # Download and unzip liblsl
@@ -22,6 +21,7 @@ install:
   - "set PYLSL_LIB=%PYTHON%\\Lib\\site-packages\\lsl.dll"
   # Install other dependencies
   - "python setup.py install"
+  - "pip install --upgrade urllib3>=2.0.5"  # Moved to this position
   - "pip install pytest"
 
 build: off
