+++ appveyor.yml
@@ -25,7 +25,7 @@ install:
   - "set PYLSL_LIB=%PYTHON%\\Lib\\site-packages\\lsl.dll"
   # Uninstall and reinstall urllib3
   - "pip uninstall -y urllib3"
-  - "pip install --upgrade 'urllib3>=2.0.5'"
+  - "pip install --upgrade urllib3>=2.0.5"
   # Install other dependencies
   - "python setup.py install"
   # Install pytest
