+++ appveyor.yml
@@ -38,6 +38,6 @@ build: off
 test_script:
   - "echo 'About to run tests...'"
   # Set timeout for pytest (e.g., 300 seconds)
-  - "pytest Tests/ --timeout=300"
+  - "pytest Tests/ --timeout=750"
   - "echo 'Finished running tests.'"
   - "if %errorlevel% neq 0 exit /b %errorlevel%"
