+++ appveyor.yml
@@ -37,7 +37,7 @@ build: off
 
 test_script:
   - "echo 'About to run tests...'"
-  # Set timeout for pytest (e.g., 300 seconds)
-  - "pytest Tests/ --timeout=1000"
+  # Set timeout for pytest (e.g., 300 seconds)--timeout=1000
+  - "pytest Tests/"
   - "echo 'Finished running tests.'"
   - "if %errorlevel% neq 0 exit /b %errorlevel%"
