+++ appveyor.yml
@@ -13,7 +13,7 @@ install:
   # Update PATH
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
   # Upgrade pip, setuptools, wheel
-  - "python -m pip install --upgrade pip setuptools wheel"
+  - "python -m pip install --upgrade pip"
   - "python --version"
   - "pip --version"
   # Download and unzip liblsl
@@ -27,7 +27,7 @@ install:
   - "pip uninstall -y urllib3"
   - "pip install --upgrade urllib3>=2.0.5"
   # Install other dependencies
-  - "python setup.py install"
+  - "pip install ."
   # Install pytest
   - "pip install pytest"
   # Exit if error
