### 2023-10-01 e7f5c51
bug fixes
thread and processing cleanly closes for pseudodevice now, should help with appveyor testing.
