+++ appveyor.yml
@@ -38,6 +38,7 @@ build: off
 
 test_script:
   - "echo 'About to run tests...'"
+  - "set PYTEST_ERRORLEVEL=0"
   - "pytest Tests/ || SET PYTEST_ERRORLEVEL=%errorlevel%"
   - "if %PYTEST_ERRORLEVEL% neq 0 (echo 'Pytest failed. Terminating python process.' & taskkill /IM python.exe /F) else (echo 'Pytest succeeded.')"
   - "echo 'Finished running tests.'"
