+++ appveyor.yml
@@ -5,7 +5,6 @@ environment:
     - PYTHON: "C:\\Python311-x64"
     - PYTHON: "C:\\Python310-x64"
     - PYTHON: "C:\\Python39-x64"
-    - PYTHON: "C:\\Python38-x64"
 cache:
   - liblsl -> appveyor.yml
 
