+++ setup.py
@@ -83,9 +83,6 @@ setup(
         'Operating System :: POSIX :: Linux',
         'Operating System :: MacOS',
 
-        # Specify the Python versions you support here. In particular, ensure
-        # that you indicate whether you support Python 2, Python 3 or both.
-        'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
