+++ README.md
@@ -49,7 +49,7 @@ Once the liblsl library has been downloaded ```pip install pybci-package``` shou
 [Examples of supported LSL hardware here!](https://labstreaminglayer.readthedocs.io/info/supported_devices.html)
 
 ## Python Package Dependencies Version Minimums
-Tested on Python 3.8+ ([appveyor.yml](https://github.com/LMBooth/pybci/blob/main/appveyor.yml))
+Tested on Python 3.9+ ([appveyor.yml](https://github.com/LMBooth/pybci/blob/main/appveyor.yml))
 
 The following package versions define the minimum supported by PyBCI, also defined in setup.py:
 
