### 2023-10-03 #5706 @LMBooth
HI @emdupre, thanks for checking in .  I've recently managed to get my appveyor scripts running for a couple simple automated tests, think i want to extend to a couple more cases. Also now we have some scripts to run from a CLI interface, getting both those done had my past couple weekends spare time.

I still need to make a few changes to the documentation but in comparison that should be a lot quicker and easier to get through. I am going away to enjoy the sun next week. I'd like to aim for all edits and changes to be made by the 20th of October to set a target.
