### 2023-10-13 b7ab584
split introduction in to getting started and what is pybci pages
