### 2023-10-13 LMBooth/pybci#9 @LMBooth
Thankyou for these recommendations @jsheunis, i've had a reshuffle and will close the issue for now. Hopefully the information is now presented in a more intuitive manner.
