+++ docs/BackgroundInformation/Examples.rst
@@ -6,49 +6,67 @@ Examples
 The following examples can all be found on the `PyBCI GitHub repository <https://github.com/LMBooth/pybci/tree/main/pybci/Examples>`_.
 
 .. note:: 
-   All the examples shown that are not in a dedicated folder work with the `createPseudoDevice` variable set to `True` when instantiating `PBCI()`. If using your own LSL-capable hardware, you may need to adjust the scripts accordingly, namely set `createPseudoDevice=False`.
+   All the examples shown that are not in a dedicated folder work with the `createPseudoDevice` variable set to `True` when instantiating `PyBCI()`. If using your own LSL-capable hardware, you may need to adjust the scripts accordingly, namely set `createPseudoDevice=False`.
 
 ArduinoHandGrasp
 ----------------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Not_Available-blue
+   :alt: pseudo device not available shield
 - **GitHub Link**: `ArduinoHandGrasp/ <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/ArduinoHandGrasp>`_
 - **Description**: This folder contains an LSL marker creator in `MarkerMaker.py`, which uses PyQt5 as an on-screen text stimulus. It also includes `ServoControl.ino`, designed for an Arduino Uno to control 5 servo motors. A `Myoware Muscle Sensor` is attached to analog pin A0. The `ArduinoToLSL.py` script sends and receives serial data, while `testArduinoHand.py` classifies the data.
 
 PupilLabsRightLeftEyeClose
 --------------------------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Not_Available-blue
+   :alt: pseudo device not available shield
 - **GitHub Link**: `PupilLabsRightLeftEyeClose/ <https://github.com/LMBooth/pybci/blob/main/pybci/Examples/PupilLabsRightLeftEyeClose/>`_
 - **Description**: This folder contains a basic Pupil Labs example with a custom extractor class. `RightLeftMarkers.py` uses Tkinter to generate visual stimuli. `bciGazeExample.py` shows how a custom feature extractor class can be used.
 
 MultimodalPupilLabsEEG
 -----------------------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Not_Available-blue
+   :alt: pseudo device not available shield
 - **GitHub Link**: `MultimodalPupilLabsEEG/ <https://github.com/LMBooth/pybci/tree/main/pybci/Examples/MultimodalPupilLabsEEG>`_
 - **Description**: An advanced example illustrating the use of two devices: Pupil Labs and Hull University ioBio EEG device. Includes a YouTube video demonstrating the multimodal example.
 
 testEpochTimingsConfig
 -----------------------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Available-blue
+   :alt: pseudo device not available shield
 - **GitHub Link**: `testEpochTimingsConfig <https://github.com/LMBooth/pybci/blob/main/pybci/Examples/testEpochTimingsConfig.py>`_
 - **Description**: A simple example showing custom global epoch settings.
 
 testPytorch
 -----------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Available-blue
+   :alt: pseudo device not available shield
 - **GitHub Link**: `testPytorch <https://github.com/LMBooth/pybci/blob/main/pybci/Examples/testPytorch.py>`_
 - **Description**: Provides an example of using a PyTorch Neural Net Model as the classifier.
 
 testRaw
 -------
+.. image:: https://img.shields.io/badge/Pseudo_Device-Available-blue
+   :alt: pseudo device not available shield
[…]
