### 2023-10-13 LMBooth/pybci#10 @LMBooth
Excellent suggestion @jsheunis , I've amended the documentation to be more informative to the user. Thankyou for bringing this to my attention.
