### 2023-10-13 LMBooth/pybci#14 @LMBooth
Hi @jsheunis, thanks for pointing out these recommendations. I've amended the hard to navigate table and added shields and more informative text describing what is needed to get the user to understand how to use the examples.

In regard to arg passing for the examples - I've now implemented a few [CLI based examples](https://github.com/LMBooth/pybci/tree/main/pybci/CliTests)  where arguments can be passed. For the [example scripts here](https://github.com/LMBooth/pybci/tree/main/pybci/Examples) my intention is to make it simple for people to take the scripts, edit and implement their own use cases as needed. Since CLI based examples are available i feel adding argument parsing to the other examples will increase the cognitive workload to deduce what is necessary or not. Would you agree that with the addition of the CLI examples the addition of arg passing to the other examples is no longer needed and makes them clearer to read and understand?
