### 2023-10-13 LMBooth/pybci#15 @LMBooth
We now have appveyor tests, i would like to extend the appveyor tests to two or three more use cases before closing this issue.

Thankyou for recommending these tests and enhancements @jsheunis. I have to admit it was a learning curve as i haven't developed used an automated testing service before like appveyor which is awesome an addition and  I'm glad you recommended it.
