### 2023-10-13 be462c0
Update Pseudo_Device.rst
