+++ pybci/Examples/README.md
@@ -2,16 +2,26 @@
 
 This folder holds multiple scripts illustrating the functions and configurations available within the PyBCI package.
 
-NOTE: all the examples shown that are not in a dedicated folder work with the createPseudoDevice variable set to True when instantiating PyBCI(). If using with own LSL capable hardware you may need to adjust the scripts accordingly, namely set `createPseudoDevice=False`.
+NOTE:    The examples have shields describing whether they work with PyBCI's pseudoDevice class and what additional external hardware is required. If using your own LSL-capable hardware  and marker stream set `createPseudoDevice=False` or optionally pass `True` or `False` as an arguement to each script.
+
+PyBCI requires an LSL marker stream for defining when time series data should be attributed to an action/marker/epoch and an LSL data stream to create time-series data. 
+
+If the user has no available LSL hardware to hand they can set `createPseudoDevice=True` when instantiating the PyBCI object to enable a pseudo LSL data stream to generate time-series data and LSL marker stream for epoching the data. More information on PyBCI's Pseudo Device class can be found here: :ref:`what-pseudo-device`. 
+
+The (example scripts)[https://pybci.readthedocs.io/en/latest/BackgroundInformation/Examples.html] illustrate various applied ML libraries (SKLearn, Tensorflow, PyTorch) or provide examples of how to integrate LSL hardware.
+
+The code snippet can be used below to run a simple classification task using the Pseudo Device, alternatively call pybci in the command line to get a list of CLI commands and tests:
+
 
 | Example File | Description |
 |--------------|-------------|
-| [ArduinoHandGrasp/](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/) | Folder contains an LSL marker creator in [MarkerMaker.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/MarkerMaker.py) using PyQt5 as an on screen text stimulus, illustrates how LSL markers can be used to train. [ServoControl.ino](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/ServoControl/ServoControl.ino) is designed for an arduino uno which controls 5 servo motors, each of which control the position of an indidividual finger for a 3D printed hand which can be controlled via serial commands. There is also a [Myoware Muscle Sensor](https://myoware.com/products/muscle-sensor/) attached to analog pin A0 being read continuously over the serial connection. [ArduinoToLSL.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/ArduinoToLSL.py) is used to send and receive serial data to and from the arduino, whilst pushing the A0 data to an LSL outlet which is classified in [testArduinoHand.py](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/ArduinoHandGrasp/testArduinoHand.py), whilst simultaneously receiving a marker stream from testArduinoHand.py to inform which hand position to do.|
-| [MultimodalPupilLabsEEG/](https://github.com/LMBooth/pybci/blob/main/pybci/Examples/MultimodalPupilLabsEEG) | Advanced example illustrating two devices, pupil labs gaze device stream wth custom feature extractor class and Hull University ioBio EEG device, specifically channels Fp1 and Fp2. The PupilLabsRightLeftEyeClose folder has a youtube video illustrating the multimodal example in action. |
[…]
