### 2023-10-14 c7e86ec
Update What_is_PyBCI.rst
