### 2023-10-14 6bc749d
Update What_is_PyBCI.rst
