### 2023-10-14 d75ca28
Update Theory_Operation.rst
