### 2023-10-14 45eed35
Update Theory_Operation.rst
