### 2023-10-14 dd97cf9
Update Theory_Operation.rst
