### 2023-10-14 b2dd2ca
Update Pseudo_Device.rst
