### 2023-10-14 2a680a5
Update Pseudo_Device.rst
