### 2023-10-14 LMBooth/pybci#13 @LMBooth
Hi @jsheunis, i've had a comb over and standardised so all py attributes, methods and classes are highlighted appropriately and all variables/equations are now in plain text too. Thanks for spotting this.
