### 2023-10-15 8ddadd2
Create CODE_OF_CONDUCT.md
