### 2023-10-15 LMBooth/pybci#17 @LMBooth
Added [contributing.md](https://github.com/LMBooth/pybci/blob/main/CONTRIBUTING.md) to root of repository and [equivalent in the read the docs.](https://pybci.readthedocs.io/en/latest/BackgroundInformation/Contributing.html)
