### 2023-10-15 dfb8585
Update What_is_PyBCI.rst
