### 2023-10-15 8cdcae6
Update What_is_PyBCI.rst
