### 2023-10-15 LMBooth/pybci#11 @LMBooth
added a dedicated statement of need in the[ "What is PyBCI" page](https://pybci.readthedocs.io/en/latest/BackgroundInformation/What_is_PyBCI.html#statement-of-need)
