+++ Tests/TestSimple.py
@@ -1,12 +1,11 @@
 from pybci import PyBCI
-import time
+import time, os
 import pytest
-from sklearn.neural_network import MLPClassifier
+
 # Test case using the fixture
 @pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_bci():
-    clf = MLPClassifier(max_iter = 1000, solver ="lbfgs")#solver=clf, alpha=alpha,hidden_layer_sizes=hid)
-    bci = PyBCI(minimumEpochsRequired=5, createPseudoDevice=True,  clf = clf)
+    bci = PyBCI(minimumEpochsRequired=5, createPseudoDevice=True)
     while not bci.connected:
         bci.Connect()
         time.sleep(1)
@@ -23,7 +22,7 @@ def test_run_bci():
             if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
                 classInfo = bci.CurrentClassifierInfo() # hangs if called too early
                 accuracy = classInfo["accuracy"]###
-                if accuracy > 0.75:
+                if accuracy > 0.5:
                     accuracy_achieved = True
                     bci.StopThreads()
                     break
