### 2023-10-15 LMBooth/pybci#15 @LMBooth
I've extended to 5tests to cover: simple implementation, the pseudo device creation, sklearn, tensorflow and pytorch.
