### 2023-10-15 7555835
update pseudo tests
added pytorch, tensorflow and pseudodevice pytests
