+++ Tests/test_PseudoDevice.py
@@ -3,7 +3,7 @@ import pytest
 from pybci import PyBCI
 from pybci.Utils.PseudoDevice import PseudoDeviceController
 # Test case using the fixture
-@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
+#@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_pseudo():
     bci = PyBCI(minimumEpochsRequired=5, createPseudoDevice=True)
     while not bci.connected:
