+++ Tests/test_Tensorflow.py
@@ -20,7 +20,7 @@ model.compile(loss='sparse_categorical_crossentropy',# using sparse_categorical 
               optimizer='adam',
               metrics=['accuracy'])
 
-@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
+#@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_bci():
     bci = PyBCI(minimumEpochsRequired = 4, createPseudoDevice=True, model = model)
     while not bci.connected:
