### 2023-10-15 #5706 @LMBooth
Hi @emdupre, @anilbey  and @jsheunis, I've finished updating my documentation and automated tests now relevant to @jsheunis comments and git issues raised. All the feedback was incredibly useful and has made PyBCI in to a much more comprehensive package, now highlighting how others can help contribute toward the project
