### 2023-10-16 f8fcdba
Update Pseudo_Device.rst
