### 2023-10-20 #5706 @emdupre
Thank you, @LMBooth !

@anilbey and @jsheunis, would you be able to revisit your reviewer checklists in light of these updates ? It looks like https://github.com/LMBooth/pybci/issues/14 is still awaiting reviewer confirmation to close as well.

If you have any questions or concerns, please don't hesitate to ping me in this main thread or in the relevant issues ! And thanks again for your work in reviewing _PyBCI_ to date.
