### 2023-10-21 LMBooth/pybci#18 @anilbey
some import errors such as NameError
I've noticed a potential oversight in `PseudoDevice.py` where certain lines are likely to raise `NameError`s due to the absence of the `multiprocessing` import. Specifically, these lines are:

1) https://github.com/LMBooth/pybci/blob/f8fcdbaacded50138a0e27a88f36683e42b22c4e/pybci/Utils/PseudoDevice.py#L119
2) https://github.com/LMBooth/pybci/blob/f8fcdbaacded50138a0e27a88f36683e42b22c4e/pybci/Utils/PseudoDevice.py#L125

For clarity, here's a snippet illustrating the error:

```python
>>> from multiprocessing import Queue
>>> multiprocessing.Queue
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'multiprocessing' is not defined
>>>
```

This brings up three key concerns:

1) NameError due to the missing import statements in two locations.
2) The tests currently do not capture this NameError, suggesting potential gaps in our test coverage.
3) Additionally, there are several imports present in the codebase that are not in use.

To mitigate the first and third issues, utilizing code linters, such as `ruff`, could be beneficial:

```shell
pip install ruff
ruff .
pybci/ThreadClasses/OptimisedDataReceiverThread.py:1:19: F401 [*] `time` imported but unused
pybci/ThreadClasses/OptimisedDataReceiverThread.py:2:25: F401 [*] `collections.deque` imported but unused
pybci/ThreadClasses/OptimisedDataReceiverThread.py:3:8: F401 [*] `itertools` imported but unused
pybci/ThreadClasses/OptimisedDataReceiverThread.py:5:20: F401 [*] `bisect.bisect_left` imported but unused
pybci/ThreadClasses/OptimisedDataReceiverThread.py:128:35: E712 Comparison to `False` should be `cond is False` or `if not cond:`
pybci/Utils/Classifier.py:4:8: F401 [*] `tensorflow` imported but unused
```

Incorporating a linter into the CI pipeline will help prevent such oversights in future commits.

For the second concern, generating a code coverage report would provide insights into the areas currently untested.

This issue is a part of https://github.com/openjournals/joss-reviews/issues/5706.
