### 2023-10-22 56865fc
cleaned up unused imports
used ruff linter to clean code
