+++ Tests/pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
+addopts = --cov=pybci --cov-report=xml
 filterwarnings =
     ignore::DeprecationWarning:sre_constants:19
