### 2023-10-22 8f19676
small cli updates
