+++ pybci/CliTests/testPyTorch.py
@@ -6,8 +6,8 @@ from torch.utils.data import DataLoader, TensorDataset
 from torch import nn
 
 
-def main(create_pseudo_device=True, min_epochs_train=4, min_epochs_test=10, num_chs = 8, num_feats = 2, num_classes = 4):
-    if create_pseudo_device:
+def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, num_chs = 8, num_feats = 2, num_classes = 4):
+    if createPseudoDevice:
         num_chs = 8 # 8 channels are created in the PseudoLSLGenerator
         num_feats = 2 # default is mean freq and rms to keep it simple
         num_classes = 4 # number of different triggers (can include baseline) sent, defines if we use softmax of binary
@@ -60,29 +60,33 @@ def main(create_pseudo_device=True, min_epochs_train=4, min_epochs_test=10, num_
             accuracy = correct / len(y_test)
         return accuracy, model
 
-    bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=True,  torchModel = PyTorchModel)
+    bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=createPseudoDevice,  torchModel = PyTorchModel)
     while not bci.connected: # check to see if lsl marker and datastream are available
         bci.Connect()
         time.sleep(1)
     bci.TrainMode() # now both marker and datastreams available start training on received epochs
     accuracy = 0
+    test = False
     try:
         while(True):
-            currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
-            time.sleep(0.5) # wait for marker updates
-            print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
-            if len(currentMarkers) > 1:  # check there is more then one marker type received
-                if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
-                    classInfo = bci.CurrentClassifierInfo() # hangs if called too early
-                    accuracy = classInfo["accuracy"]
-                if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
-                    bci.TestMode()
-                    break
-        while True:
-            markerGuess = bci.CurrentClassifierMarkerGuess() # when in test mode only y_pred returned
-            guess = [key for key, value in currentMarkers.items() if value[0] == markerGuess]
-            print("Current marker estimation: " + str(guess), end="           \r")
-            time.sleep(0.2)
+            if test is False:
+                currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
+                time.sleep(0.5) # wait for marker updates
+                print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
+                if len(currentMarkers) > 1:  # check there is more then one marker type received
+                    if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
+                        classInfo = bci.CurrentClassifierInfo() # hangs if called too early
+                        accuracy = classInfo["accuracy"]
+                    if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
+                        bci.TestMode()
+                        break
+            else:
+                markerGuess = bci.CurrentClassifierMarkerGuess() # when in test mode only y_pred returned
+                guess = [key for key, value in currentMarkers.items() if value[0] == markerGuess]
+                print("Current marker estimation: " + str(guess), end="           \r")
+                time.sleep(0.2)
+
[…]
