+++ pybci/ThreadClasses/DataReceiverThread.py
@@ -1,99 +1,0 @@
-import threading
-from collections import deque
-import itertools
-
-class DataReceiverThread(threading.Thread):
-    """Responsible for receiving data from accepted LSL outlet, slices samples based on tmin+tmax basis, 
-    starts counter for received samples after marker is received in ReceiveMarker.
-    """
-    startCounting = False
-    currentMarker = ""
-    def __init__(self, closeEvent, trainTestEvent, dataQueueTrain,dataQueueTest, dataStreamInlet,  
-                 customEpochSettings, globalEpochSettings,devCount,  streamChsDropDict = []):
-        super().__init__()
-        self.trainTestEvent = trainTestEvent
-        self.closeEvent = closeEvent
-        self.dataQueueTrain = dataQueueTrain
-        self.dataQueueTest = dataQueueTest
-        self.dataStreamInlet = dataStreamInlet
-        self.customEpochSettings = customEpochSettings
-        self.globalEpochSettings = globalEpochSettings
-        self.streamChsDropDict = streamChsDropDict
-        self.sr = dataStreamInlet.info().nominal_srate()
-        self.devCount = devCount # used for tracking which device is sending data to feature extractor
-        
-    def run(self):
-        posCount = 0
-        chCount = self.dataStreamInlet.info().channel_count()
-        maxTime = (self.globalEpochSettings.tmin + self.globalEpochSettings.tmax)
-        if len(self.customEpochSettings.keys())>0:
-            if  max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings]) > maxTime:
-                maxTime = max([self.customEpochSettings[x].tmin + self.customEpochSettings[x].tmax for x in self.customEpochSettings])
-        fifoLength = int(self.dataStreamInlet.info().nominal_srate()*maxTime)
-        dataFIFOs = [deque(maxlen=fifoLength) for ch in range(chCount - len(self.streamChsDropDict))]
-        while not self.closeEvent.is_set():
-            sample, timestamp = self.dataStreamInlet.pull_sample(timeout = 1)
-            if sample is not None:
-                for index in sorted(self.streamChsDropDict, reverse=True):
-                    del sample[index] # remove the desired channels from the sample
-                for i,fifo in enumerate(dataFIFOs):
-                    fifo.append(sample[i])
-                if self.trainTestEvent.is_set(): # We're training!
-                    if self.startCounting: # we received a marker
-                        posCount+=1
-                        if posCount >= self.desiredCount:  # enough samples are in FIFO, chop up and put in dataqueue
-                            if len(self.customEpochSettings.keys())>0: #  custom marker received
-                                if self.customEpochSettings[self.currentMarker].splitCheck: # slice epochs in to overlapping time windows
-                                    window_samples =int(self.customEpochSettings[self.currentMarker].windowLength * self.sr) #number of samples in each window
-                                    increment = int((1-self.customEpochSettings[self.currentMarker].windowOverlap)*window_samples) # if windows overlap each other by how many samples
-                                    while posCount - window_samples > 0:
-                                        sliceDataFIFOs = [list(itertools.islice(d, posCount - window_samples, posCount)) for d in dataFIFOs]
-                                        self.dataQueueTrain.put([sliceDataFIFOs, self.currentMarker, self.sr,  self.devCount])
-                                        posCount-=increment
-                                else: # don't slice just take tmin to tmax time
-                                    sliceDataFIFOs = [list(itertools.islice(d, fifoLength - int((self.customEpochSettings[self.currentMarker].tmax+self.customEpochSettings[self.currentMarker].tmin) * self.sr), fifoLength))for d in dataFIFOs]
[…]

[…]
