+++ .github/workflows/ci.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: windows-latest
     strategy:
       matrix:
-        python-version: [3.11, 3.10, 3.9]
+        python-version: [3.11]
         
     steps:
     - name: Checkout code
