+++ .github/workflows/ci.yml
@@ -48,7 +48,12 @@ jobs:
     - name: Run Tests
       run: |
         pytest Tests/ --cov=pybci --cov-report=xml
-
+  - name: Upload coverage to GitHub Artifacts
+      uses: actions/upload-artifact@v2
+      with:
+        name: coverage-report
+        path: coverage.xml
+    
     - name: Upload Coverage to Codecov
       uses: codecov/codecov-action@v2
       with:
