### 2023-10-22 LMBooth/pybci#18 @LMBooth
Hi @anilbey , thanks for recommending the linter and coverage report! I've ran ruff and cleaned up a lot of the unnecessary imports and added it in to the appveyor process to help identify future issues like you've recommended.

For your other recomendation, I ran some coverage checks locally which is returning results but seem to be having a bit of trouble getting anything back using Codecov in a workflow, hopefuly will crack it when i finish work tomorrow evening. I've tried extending some of the tests to also work with the cli commands too but been having trouble getting them to return properly so may exclude them from the coverage for now.
