+++ .github/workflows/ci.yml
@@ -44,6 +44,13 @@ jobs:
         pip install --upgrade urllib3>=2.0.5
         pip install ruff
         pip install .
+
+    - name: Update Environment Variables
+      run: |
+        echo "Updating Environment Variables"
+        $env:PYTHONPATH="${{ github.workspace }}"
+        $env:PYLSL_LIB="${{ env.pythonLocation }}\Lib\site-packages\lsl.dll"
+
         
     - name: Debug Information
       run: |
