+++ appveyor.yml
@@ -42,6 +42,6 @@ test_script:
   - "set PYTEST_ERRORLEVEL=0"
   - "pytest Tests/ || SET PYTEST_ERRORLEVEL=%errorlevel%"
   - "if %PYTEST_ERRORLEVEL% neq 0 (echo 'Pytest failed. Terminating python process.' & taskkill /IM python.exe /F) else (echo 'Pytest succeeded.')"
-  - "ruff ."
+  #- "ruff ."
   - "echo 'Finished running tests.'"
   - "if %errorlevel% neq 0 exit /b %errorlevel%"
