+++ Tests/pytest.ini
@@ -1,4 +1,3 @@
 [pytest]
-addopts = --cov=pybci --cov-report=xml
 filterwarnings =
     ignore::DeprecationWarning:sre_constants:19
