+++ .github/workflows/ci.yml
@@ -61,7 +61,7 @@ jobs:
 
     - name: Run Tests
       run: |
-        pytest Tests/ --cov=pybci --cov-report=xml
+        pytest Tests/ --cov=pybci --cov-report=xml -vv
     - name: Upload coverage to GitHub Artifacts
       uses: actions/upload-artifact@v2
       with:
