+++ .github/workflows/ci.yml
@@ -62,6 +62,12 @@ jobs:
     - name: Run Tests
       run: |
         pytest Tests/ --cov=pybci --cov-report=xml -vv
+
+    - name: Generate Report
+      run: |
+        pip install coverage
+        coverage run -m pytest
+        
     - name: Upload coverage to GitHub Artifacts
       uses: actions/upload-artifact@v2
       with:
@@ -69,6 +75,7 @@ jobs:
         path: coverage.xml
     - name: List working directory
       run: ls -al
+      
     - name: Upload Coverage to Codecov
       uses: codecov/codecov-action@v2
       with:
