+++ Tests/test_Pytorch.py
@@ -61,7 +61,7 @@ def test_run_bci():
     bci.TrainMode()
     accuracy_achieved = False
     marker_received = False
-    accuracy=0
+    accuracy=None
     while True:
         currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
         time.sleep(0.5) # wait for marker updates
@@ -71,7 +71,7 @@ def test_run_bci():
             if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
                 classInfo = bci.CurrentClassifierInfo() # hangs if called too early
                 accuracy = classInfo["accuracy"]###
-                if accuracy > 0:
+                if accuracy >= 0:
                     # set to above 0 to show some accuracy was retruend from model
                     accuracy_achieved = True
                     bci.StopThreads()
