+++ Tests/test_Simple.py
@@ -14,7 +14,7 @@ def test_run_bci():
     while True:
         currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
         time.sleep(0.5) # wait for marker updates
-        print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
+        #print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
         if len(currentMarkers) > 1:  # check there is more then one marker type received
             marker_received = True
             if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
