+++ .github/workflows/ci.yml
@@ -65,8 +65,8 @@ jobs:
 
     - name: Generate Report
       run: |
-        coverage run -m pytest
-        coverage xml
+        coverage run --source=pybci/ -m pytest
+        coverage xml -i
     #- name: Upload coverage to GitHub Artifacts
     #  uses: actions/upload-artifact@v2
     #  with:
