### 2023-10-24 e5aa92b
test and feature extractor update
removed sample entropy in feature selection
added cli test
added extra feature tests
