+++ pybci/version.py
@@ -1,2 +1,2 @@
-__version__ = '1.2.6'
+__version__ = '1.3.0'
 
