### 2023-10-24 69a8626
change test name due to improper closure
