+++ pybci/CliTests/testTensorflow.py
@@ -2,8 +2,45 @@ import time
 import argparse
 from pybci import PyBCI
 import tensorflow as tf# bring in tf for custom model creation
+import threading
 
-def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, num_chs = 8, num_feats = 2, num_classes = 4):
+def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, num_chs = 8, num_feats = 2, num_classes = 4, timeout=None):
+    def loop(bci):
+        while not bci.connected: # check to see if lsl marker and datastream are available
+            bci.Connect()
+            time.sleep(1)
+        bci.TrainMode() # now both marker and datastreams available start training on received epochs
+        accuracy = 0
+        test = False
+        try:
+            while(True):
+                if test is False:
+                    currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
+                    time.sleep(0.5) # wait for marker updates
+                    print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
+                    if len(currentMarkers) > 1:  # check there is more then one marker type received
+                        if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
+                            classInfo = bci.CurrentClassifierInfo() # hangs if called too early
+                            accuracy = classInfo["accuracy"]
+                        if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
+                            bci.TestMode()
+                            break
+                else:
+                    markerGuess = bci.CurrentClassifierMarkerGuess() # when in test mode only y_pred returned
+                    guess = [key for key, value in currentMarkers.items() if value[0] == markerGuess]
+                    print("Current marker estimation: " + str(guess), end="           \r")
+                    time.sleep(0.2)
+
+            return None
+        except KeyboardInterrupt: # allow user to break while loop
+            print("\nLoop interrupted by user.")
+
+    
+    def stop_after_timeout(bci):
+        time.sleep(timeout)
+        print("\nTimeout reached. Stopping threads.")
+        bci.StopThreads()
+
     if createPseudoDevice:
         num_chs = 8 # 8 channels are created in the PseudoLSLGenerator
         num_feats = 2 # default is mean freq and rms to keep it simple
@@ -21,34 +58,13 @@ def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, num_ch
               optimizer='adam',
               metrics=['accuracy'])
     bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=createPseudoDevice, model = model)
-    while not bci.connected: # check to see if lsl marker and datastream are available
-        bci.Connect()
-        time.sleep(1)
-    bci.TrainMode() # now both marker and datastreams available start training on received epochs
-    accuracy = 0
-    test = False
-    try:
-        while(True):
-            if test is False:
-                currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
-                time.sleep(0.5) # wait for marker updates
-                print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
-                if len(currentMarkers) > 1:  # check there is more then one marker type received
-                    if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
-                        classInfo = bci.CurrentClassifierInfo() # hangs if called too early
-                        accuracy = classInfo["accuracy"]
[…]
