+++ pybci/CliTests/testSimple.py
@@ -1,43 +1,62 @@
 import argparse
 from pybci import PyBCI
 import time 
+import threading
 
-def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10):
+def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, timeout=None):
+    def loop(bci):
+        while not bci.connected: # check to see if lsl marker and datastream are available
+            bci.Connect()
+            time.sleep(1)
+        bci.TrainMode() # now both marker and datastreams available start training on received epochs
+        accuracy = 0
+        test = False
+        try:
+            while(True):
+                if test is False:
+                    currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
+                    time.sleep(0.5) # wait for marker updates
+                    print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
+                    if len(currentMarkers) > 1:  # check there is more then one marker type received
+                        if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
+                            classInfo = bci.CurrentClassifierInfo() # hangs if called too early
+                            accuracy = classInfo["accuracy"]
+                        if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
+                            bci.TestMode()
+                            test = True
+                else:
+                    markerGuess = bci.CurrentClassifierMarkerGuess() # when in test mode only y_pred returned
+                    guess = [key for key, value in currentMarkers.items() if value[0] == markerGuess]
+                    print("Current marker estimation: " + str(guess), end="           \r")
+                    time.sleep(0.2)
+        except KeyboardInterrupt: # allow user to break while loop
+            print("\nLoop interrupted by user.")
+
+    def stop_after_timeout(bci):
+        time.sleep(timeout)
+        print("\nTimeout reached. Stopping threads.")
+        bci.StopThreads()
+    
     if min_epochs_test <= min_epochs_train:
         min_epochs_test = min_epochs_train+1
     bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=createPseudoDevice)
-    while not bci.connected: # check to see if lsl marker and datastream are available
-        bci.Connect()
-        time.sleep(1)
-    bci.TrainMode() # now both marker and datastreams available start training on received epochs
-    accuracy = 0
-    test = False
-    try:
-        while(True):
-            if test is False:
-                currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
-                time.sleep(0.5) # wait for marker updates
-                print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
-                if len(currentMarkers) > 1:  # check there is more then one marker type received
-                    if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
-                        classInfo = bci.CurrentClassifierInfo() # hangs if called too early
-                        accuracy = classInfo["accuracy"]
-                    if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
-                        bci.TestMode()
-                        test = True
-            else:
-                markerGuess = bci.CurrentClassifierMarkerGuess() # when in test mode only y_pred returned
-                guess = [key for key, value in currentMarkers.items() if value[0] == markerGuess]
-                print("Current marker estimation: " + str(guess), end="           \r")
-                time.sleep(0.2)
[…]
