### 2023-10-24 LMBooth/pybci#18 @LMBooth
For an update, managed to get the codecov report working and added the shield to the readme.md in the root of the repository. Currently trying to expand the tests to cover more of the codebase. Thanks again @anilbey , i'll leave the issue open for now until i manage to finish expanding my test coverage.
