### 2023-10-24 1f414c1
Update test_cli.py
