+++ Tests/test_cli.py
@@ -1,5 +1,5 @@
 import subprocess
-
+import time
 def run_cli_command(command):
     process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
     stdout, stderr = process.communicate()
@@ -13,7 +13,11 @@ def run_cli_command(command):
 # Example usage
 def test_cli():
     run_cli_command("pybci testSimple --timeout=10")
+    time.sleep(15)
     run_cli_command("pybci testSklearn --timeout=10")
+    time.sleep(15)
     run_cli_command("pybci testPyTorch --timeout=10")
+    time.sleep(15)
     run_cli_command("pybci testTensorflow --timeout=10")
+    time.sleep(15)
     assert True
