### 2023-10-24 b8e38fc
Update test_cli.py
