### 2023-10-24 2169cdd
Update test_cli.py
