### 2023-10-25 60baf11
Update test_cli.py
