### 2023-10-25 745218c
Update test_cli.py
