### 2023-10-25 f6e4241
cli update
updates cli calls to exit more gracefully, and added coverage tests for cli
