+++ pybci/CliTests/testSimple.py
@@ -3,52 +3,85 @@ from pybci import PyBCI
 import time 
 import threading
 
-def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, timeout=None):
-    def loop(bci):
-        while not bci.connected: # check to see if lsl marker and datastream are available
-            bci.Connect()
+stop_signal = threading.Event()  # Global event to control the main loop
+
+def command_listener():
+    while  not stop_signal.is_set():
+        command = input("PyBCI: [CLI] - Enter 'stop' to terminate\n")
+        if command == 'stop':
+            stop_signal.set()
+            break
+
+class CLI_testSimpleWrapper:
+    def __init__(self, createPseudoDevice, min_epochs_train, min_epochs_test, timeout):
+        self.createPseudoDevice = createPseudoDevice
+        self.timeout = timeout
+        self.min_epochs_train = min_epochs_train
+        self.min_epochs_test = min_epochs_test
+        self.accuracy = 0
+        self.currentMarkers = {}
+        if self.min_epochs_test <= self.min_epochs_train:
+            self.min_epochs_test = self.min_epochs_train+1
+        self.bci = PyBCI(minimumEpochsRequired = self.min_epochs_train, createPseudoDevice=self.createPseudoDevice)
+        main_thread = threading.Thread(target=self.loop)
+        main_thread.start()
+        if self.timeout:
+            print("PyBCI: [CLI] - starting timeout thread")
+            self.timeout_thread = threading.Thread(target=self.stop_after_timeout)
+            self.timeout_thread.start()
+        main_thread.join()
+        if timeout is not None:
+            self.timeout_thread.join()
+        
+    def loop(self):
+        while not self.bci.connected: # check to see if lsl marker and datastream are available
+            self.bci.Connect()
             time.sleep(1)
-        bci.TrainMode() # now both marker and datastreams available start training on received epochs
-        accuracy = 0
+        self.bci.TrainMode() # now both marker and datastreams available start training on received epochs
+        self.accuracy = 0
         test = False
         try:
-            while(True):
+            while not stop_signal.is_set():  # Add the check here
                 if test is False:
-                    currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
+                    self.currentMarkers = self.bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
                     time.sleep(0.5) # wait for marker updates
-                    print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
-                    if len(currentMarkers) > 1:  # check there is more then one marker type received
-                        if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
-                            classInfo = bci.CurrentClassifierInfo() # hangs if called too early
-                            accuracy = classInfo["accuracy"]
-                        if min([currentMarkers[key][1] for key in currentMarkers]) > min_epochs_test:  
-                            bci.TestMode()
+                    print("Markers received: " + str(self.currentMarkers) +" Accuracy: " + str(round(self.accuracy,2)), end="         \r")
+                    if len(self.currentMarkers) > 1:  # check there is more then one marker type received
+                        if min([self.currentMarkers[key][1] for key in self.currentMarkers]) > self.bci.minimumEpochsRequired:
+                            classInfo = self.bci.CurrentClassifierInfo() # hangs if called too early
+                            self.accuracy = classInfo["accuracy"]
[…]
