+++ pybci/CliTests/testPyTorch.py
@@ -6,104 +6,151 @@ from torch.utils.data import DataLoader, TensorDataset
 from torch import nn
 import threading
 
+stop_signal = threading.Event()  # Global event to control the main loop
 
-def main(createPseudoDevice=True, min_epochs_train=4, min_epochs_test=10, num_chs = 8, num_feats = 2, num_classes = 4, timeout=None):
-    if createPseudoDevice:
-        num_chs = 8 # 8 channels are created in the PseudoLSLGenerator
-        num_feats = 2 # default is mean freq and rms to keep it simple
-        num_classes = 4 # number of different triggers (can include baseline) sent, defines if we use softmax of binary
+global num_chs_g, num_feats_g, num_classes_g
 
-    class SimpleNN(nn.Module):
-        def __init__(self, input_size, hidden_size, num_classes):
-            super(SimpleNN, self).__init__()
-            self.fc1 = nn.Linear(input_size, hidden_size)
-            self.bn1 = nn.BatchNorm1d(hidden_size)
-            self.relu = nn.ReLU(inplace=True)  # In-place operation
-            self.fc2 = nn.Linear(hidden_size, hidden_size)
-            self.bn2 = nn.BatchNorm1d(hidden_size)
-            self.fc3 = nn.Linear(hidden_size, num_classes)
+class SimpleNN(nn.Module):
+    def __init__(self, input_size, hidden_size, num_classes):
+        super(SimpleNN, self).__init__()
+        self.fc1 = nn.Linear(input_size, hidden_size)
+        self.bn1 = nn.BatchNorm1d(hidden_size)
+        self.relu = nn.ReLU(inplace=True)  # In-place operation
+        self.fc2 = nn.Linear(hidden_size, hidden_size)
+        self.bn2 = nn.BatchNorm1d(hidden_size)
+        self.fc3 = nn.Linear(hidden_size, num_classes)
+    
+    def forward(self, x):
+        out = self.fc1(x)
+        if out.shape[0] > 1:  # Skip BatchNorm if batch size is 1
+            out = self.bn1(out)
+        out = self.relu(out)
+        out = self.fc2(out)
+        if out.shape[0] > 1:  # Skip BatchNorm if batch size is 1
+            out = self.bn2(out)
+        out = self.relu(out)
+        out = self.fc3(out)
+        return out
+
+def PyTorchModel(x_train, x_test, y_train, y_test):
+    input_size = num_feats_g*num_chs_g # num of channels multipled by number of default features (rms and mean freq)
+    hidden_size = 100
+    #num_classes = num_classes # default in pseudodevice
+    model = SimpleNN(input_size, hidden_size, num_classes_g)
+    model.train()
+    criterion = nn.CrossEntropyLoss()
+    optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
+    epochs = 10
+    train_data = TensorDataset(torch.Tensor(x_train), torch.Tensor(y_train).long())
+    train_loader = DataLoader(dataset=train_data, batch_size=32, shuffle=True, drop_last=True)  # Drop last incomplete batch
+    for epoch in range(epochs):
+        for inputs, labels in train_loader:
+            optimizer.zero_grad()
+            outputs = model(inputs)
+            loss = criterion(outputs, labels)
+            loss.backward()
+            optimizer.step()
+    model.eval()
+    accuracy = 0
+    with torch.no_grad():
+        test_outputs = model(torch.Tensor(x_test))
+        _, predicted = torch.max(test_outputs.data, 1)
+        correct = (predicted == torch.Tensor(y_test).long()).sum().item()
+        accuracy = correct / len(y_test)
+    return accuracy, model
         
-        def forward(self, x):
[…]

[…]
