### 2023-10-25 #5706 @LMBooth
Hi all, thanks to @anilbey suggestions,  i've managed to integrate a linter in to the appveyor CI process as well as including a coverage report via codecov.
