### 2023-10-26 #5706 @emdupre
Thank you, @LMBooth ! Apologies for the extra ping ; I hadn't removed the reminder after your latest follow-up.

Thanks, too, to @anilbey for your follow-up comments ! @jsheunis and @anilbey, please let me know your thoughts after re-visiting the reviewer checklist with _PyBCI_'s latest updates.
