### 2023-10-27 8021ede
cli and import update
- Specify necessary import to speed up module loading time instead of bringing whole libraries in.

- Changed printing format of cli tests so out is more readable, confusion made with input in cli as delayed marker print estimation.
