+++ pybci/cli.py
@@ -1,14 +1,25 @@
 import argparse
 import time
-from .CliTests.testSimple import main as testSimple
-from .CliTests.testSklearn import main as testSklearn
-from .CliTests.testTensorflow import main as testTensorflow
-from .CliTests.testPyTorch import main as testPyTorch
 
-from .Utils.PseudoDevice import PseudoDeviceController
 
+def import_and_run_testSimple(**kwargs):
+    from .CliTests.testSimple import main as testSimple
+    testSimple(**kwargs)
+
+def import_and_run_testSklearn(**kwargs):
+    from .CliTests.testSklearn import main as testSklearn
+    testSklearn(**kwargs)
+
+def import_and_run_testTensorflow(**kwargs):
+    from .CliTests.testTensorflow import main as testTensorflow
+    testTensorflow(**kwargs)
+
+def import_and_run_testPyTorch(**kwargs):
+    from .CliTests.testPyTorch import main as testPyTorch
+    testPyTorch(**kwargs)
 
 def RunPseudo():
+    from .Utils.PseudoDevice import PseudoDeviceController
     pseudoDevice = PseudoDeviceController()
     pseudoDevice.BeginStreaming()
     while True:
@@ -25,7 +36,7 @@ def main():
     testSimple_parser.add_argument('--min_epochs_test', default=14, type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSimple_parser.add_argument("--timeout", default=None, type=int, help="Timeout in seconds for the script to automatically stop.")
 
-    testSimple_parser.set_defaults(func=testSimple)
+    #testSimple_parser.set_defaults(func=testSimple)
     
     testSklearn_parser = subparsers.add_parser('testSklearn', help='Sklearn multi-layer perceptron is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testSklearn.py in the examples folder.')
     testSklearn_parser.add_argument('--createPseudoDevice',default=True, type=bool, help='Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.')
@@ -33,7 +44,7 @@ def main():
     testSklearn_parser.add_argument('--min_epochs_test', default=14,type=int, help='Minimum epochs to collect before model testing commences, if less than min_epochs_test defaults to min_epochs_test+1.')
     testSklearn_parser.add_argument("--timeout", default=None, type=int, help="Timeout in seconds for the script to automatically stop.")
 
-    testSklearn_parser.set_defaults(func=testSklearn)
+    #testSklearn_parser.set_defaults(func=testSklearn)
     
     testTensorflow_parser = subparsers.add_parser('testTensorflow', help='Tensorflow GRU is used for model and pseudodevice generates 8 channels of 3 marker types and baseline. Similar to the testTensorflow.py in the examples folder.')
     testTensorflow_parser.add_argument("--createPseudoDevice", default=True, type=bool, help="Set to True or False to enable or disable pseudo device creation. pseudodevice generates 8 channels of 3 marker types and baseline.")
@@ -43,7 +54,7 @@ def main():
     testTensorflow_parser.add_argument("--num_classes", default=4, type=int, help='Num of classes in marker stream to configure tensorflow model, if PseudoDevice==True defaults to 4.')
     testTensorflow_parser.add_argument("--timeout", default=None, type=int, help="Timeout in seconds for the script to automatically stop.")
 
-    testTensorflow_parser.set_defaults(func=testTensorflow)
+    #testTensorflow_parser.set_defaults(func=testTensorflow)
 
[…]
