+++ appveyor.yml
@@ -1,47 +1,76 @@
 version: 1.2.{build}
-image: Visual Studio 2019
-environment:
-  matrix:
-    - PYTHON: "C:\\Python311-x64"
-    - PYTHON: "C:\\Python310-x64"
-    - PYTHON: "C:\\Python39-x64"
-cache:
-  - liblsl -> appveyor.yml
+build: off
 
+environment:
+  global:
+    PYLSL_LIB: lsl.dll
+  matrix:
+    - image: Visual Studio 2019
+      PYTHON: "C:\\Python311-x64"
+    - image: Visual Studio 2019
+      PYTHON: "C:\\Python310-x64"
+    - image: Visual Studio 2019
+      PYTHON: "C:\\Python39-x64"
+    - image: ubuntu
+      PYTHON: "/usr/bin/python3"
+    - image: macOS
+      PYTHON: "/usr/local/bin/python3"
+
+matrix:
+  fast_finish: true
+
+
+Sure, here's the install section of your appveyor.yml file with adaptations for Windows, Linux, and macOS:
+
+yml
+Copy code
 install:
   # Update PATH
-  - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
+  - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
+  
   # Upgrade pip, setuptools, wheel
   - "python -m pip install --upgrade pip"
   - "python --version"
   - "pip --version"
+  
   # Download and unzip liblsl
-  - "powershell Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'"
-  - "unzip -o liblsl.zip -d liblsl"
-  # Move DLL to correct directory
-  - "move liblsl\\bin\\lsl.dll %PYTHON%\\Lib\\site-packages\\"
-  # Set PYLSL_LIB environment variable
-  - "set PYLSL_LIB=%PYTHON%\\Lib\\site-packages\\lsl.dll"
+  - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'; Expand-Archive -Path liblsl.zip -DestinationPath liblsl }
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.tar.bz2 -O liblsl.tar.bz2; tar xjf liblsl.tar.bz2 -C liblsl; fi
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-macOS.tar.bz2 -O liblsl.tar.bz2; tar xjf liblsl.tar.bz2 -C liblsl; fi
+  
[…]
