+++ appveyor.yml
@@ -5,16 +5,17 @@ environment:
   global:
     PYLSL_LIB: lsl.dll
   matrix:
+    - image: ubuntu
+      PYTHON: "/usr/bin/python3"
+    - image: macOS
+      PYTHON: "/usr/local/bin/python3"
     - image: Visual Studio 2019
       PYTHON: "C:\\Python311-x64"
     - image: Visual Studio 2019
       PYTHON: "C:\\Python310-x64"
     - image: Visual Studio 2019
       PYTHON: "C:\\Python39-x64"
-    - image: ubuntu
-      PYTHON: "/usr/bin/python3"
-    - image: macOS
-      PYTHON: "/usr/local/bin/python3"
+
 
 matrix:
   fast_finish: true
@@ -34,14 +35,31 @@ install:
   - "pip --version"
   
   # Download and unzip liblsl
-  - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'; Expand-Archive -Path liblsl.zip -DestinationPath liblsl }
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.tar.bz2 -O liblsl.tar.bz2; tar xjf liblsl.tar.bz2 -C liblsl; fi
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-macOS.tar.bz2 -O liblsl.tar.bz2; tar xjf liblsl.tar.bz2 -C liblsl; fi
+  - ps: |
+      if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") {
+          Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'
+          Expand-Archive -Path liblsl.zip -DestinationPath liblsl
+      }
+  - sh: |
+      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
+          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.tar.bz2 -O liblsl.tar.bz2
+          tar xjf liblsl.tar.bz2 -C liblsl
+      elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
+          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-macOS.tar.bz2 -O liblsl.tar.bz2
+          tar xjf liblsl.tar.bz2 -C liblsl
+      fi
   
   # Move DLL to correct directory (Windows) or Set PYLSL_LIB environment variable (Linux, macOS)
-  - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { Move-Item liblsl/bin/lsl.dll $env:PYTHON/Lib/site-packages/; $env:PYLSL_LIB="$env:PYTHON\Lib\site-packages\lsl.dll" }
[…]
