+++ appveyor.yml
@@ -28,6 +28,30 @@ install:
   - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
   - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
   - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
+
+install:
+  - ps: |
+      $ErrorActionPreference = 'Stop'
+      if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") {
+        $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH"
+        Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'
+        Expand-Archive -Path liblsl.zip -DestinationPath liblsl
+        Move-Item liblsl/bin/lsl.dll $env:PYTHON/Lib/site-packages/
+        $env:PYLSL_LIB="$env:PYTHON\Lib\site-packages\lsl.dll"
+      }
+  - sh: |
+      set -e
+      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]] || [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
+        export PATH=$PYTHON:$PYTHON/Scripts:$PATH
+        if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
+          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.tar.bz2 -O liblsl.tar.bz2
+        else
+          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-macOS.tar.bz2 -O liblsl.tar.bz2
+        fi
+        tar xjf liblsl.tar.bz2 -C liblsl
+        export PYLSL_LIB=$PWD/liblsl/lib/liblsl64.so
+      fi
+
   
   # Upgrade pip, setuptools, wheel
   - "python -m pip install --upgrade pip"
