+++ appveyor.yml
@@ -5,58 +5,35 @@ environment:
   global:
     PYLSL_LIB: lsl.dll
   matrix:
-    - ID: Ubu20core
-      APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
-    - image: ubuntu
+    - ID: Ubuntu
+      image: ubuntu
       PYTHON: "/usr/bin/python3"
-    - image: macOS
+    - ID: macOS
+      image: macOS
       PYTHON: "/usr/local/bin/python3"
-    - image: Visual Studio 2019
+    - ID: VS2019_Py311
+      image: Visual Studio 2019
       PYTHON: "C:\\Python311-x64"
-    - image: Visual Studio 2019
+    - ID: VS2019_Py310
+      image: Visual Studio 2019
       PYTHON: "C:\\Python310-x64"
-    - image: Visual Studio 2019
+    - ID: VS2019_Py39
+      image: Visual Studio 2019
       PYTHON: "C:\\Python39-x64"
 
 
 matrix:
   fast_finish: true
 
-
 install:
-  - cmd: |
-      @echo off
-      if "%APPVEYOR_BUILD_WORKER_IMAGE%" == "Visual Studio 2019" (
-        powershell.exe -Command "& {
-          $ErrorActionPreference = 'Stop';
-          $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH";
-          Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip';
-          Expand-Archive -Path liblsl.zip -DestinationPath liblsl;
-          Move-Item liblsl/bin/lsl.dll $env:PYTHON/Lib/site-packages/;
-          $env:PYLSL_LIB="$env:PYTHON\Lib\site-packages\lsl.dll";
-        }"
-      ) else (
-        if "%APPVEYOR_BUILD_WORKER_IMAGE%" == "ubuntu" (
-          bash -c "
-            export PATH=$PYTHON:$PYTHON/Scripts:$PATH;
-            wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.tar.bz2 -O liblsl.tar.bz2;
-            tar xjf liblsl.tar.bz2 -C liblsl;
-            export PYLSL_LIB=$PWD/liblsl/lib/liblsl64.so;
-          "
-        ) else (
-          bash -c "
-            export PATH=$PYTHON:$PYTHON/Scripts:$PATH;
-            wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-macOS.tar.bz2 -O liblsl.tar.bz2;
-            tar xjf liblsl.tar.bz2 -C liblsl;
-            export PYLSL_LIB=$PWD/liblsl/lib/liblsl64.so;
-          "
-        )
-      )
-  - sh: |
-    if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
-        sudo apt-get update
-        sudo apt-get install -y python3-pip
-    fi
+  - ps: $ErrorActionPreference = 'Stop'; if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
[…]
