+++ appveyor.yml
@@ -34,9 +34,9 @@ install:
   - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
   
   # Upgrade pip, setuptools, wheel
-  - "python -m pip install --upgrade pip"
+  - "python -m pip3 install --upgrade pip3"
   - "python --version"
-  - "pip --version"
+  - "pip3 --version"
   
   # Download and unzip liblsl
   - ps: |
