+++ appveyor.yml
@@ -82,19 +82,20 @@ install:
           export PYLSL_LIB=$PWD/liblsl/lib/liblsl64.so
       fi
     
-  # Uninstall and reinstall urllib3
-  - "pip3 uninstall -y urllib3"
-  - "pip3 install --upgrade urllib3>=2.0.5"
-  
-  # Install other dependencies
-  - "pip3 install ."
-  
-  # Install pytest
-  - "pip3 install pytest"
-  - "pip3 install pytest-timeout"
-  
-  # Install ruff
-  - "pip3 install ruff"
+  # install dependencies
+  - ps: |
+    - "pip uninstall -y urllib3"
+    - "pip install --upgrade urllib3>=2.0.5"
+    - "pip install ."
+    - "pip install pytest"
+    - "pip install pytest-timeout"
+    - "pip install ruff"
+  - sh: |
+    - "pip3 install --upgrade urllib3>=2.0.5"
+    - "pip3 install ."
+    - "pip3 install pytest"
+    - "pip3 install pytest-timeout"
+    - "pip3 install ruff"
   
   # Exit if error
   - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio" -and $LastExitCode -ne 0) { Write-Host "Command exited with code $LastExitCode" }
