+++ appveyor.yml
@@ -7,10 +7,10 @@ environment:
   matrix:
     - ID: Ubuntu
       APPVEYOR_BUILD_WORKER_IMAGE: ubuntu
-      PYTHON: "/usr/bin/python3"
+      PYTHON:  "/usr/bin/python3.9"
     - ID: macOS
       APPVEYOR_BUILD_WORKER_IMAGE: macOS
-      PYTHON: "/usr/local/bin/python3"
+      PYTHON: "/usr/local/opt/python@3.9/bin/python3.9"
     - ID: VS2019_Py311
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       PYTHON: "C:\\Python311-x64"
@@ -25,10 +25,16 @@ matrix:
   fast_finish: true
 
 install:
-  - ps: $ErrorActionPreference = 'Stop'; if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
-  # Update PATH
+  # Install Python 3.9 for max and linux
+  - sh: |
+      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
+          sudo apt-get update
+          sudo apt-get install -y python3.9 python3-pip
+          alias python3=python3.9
+          alias pip3=pip3.9
+      elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
+          brew install python@3.9
+      fi
   - sh: |
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
           sudo apt-get update
@@ -41,7 +47,10 @@ install:
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") {
           $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH"
       }
-  
+      
+  - ps: $ErrorActionPreference = 'Stop'; if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
+  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
   # Upgrade pip, setuptools, wheel
   - sh: |
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" || $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
