+++ appveyor.yml
@@ -30,10 +30,12 @@ install:
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
           sudo apt-get update
           sudo apt-get install -y python3.9 python3-pip
-          alias python3=python3.9
-          alias pip3=pip3.9
+          sudo update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.9 1
+          sudo update-alternatives --install /usr/bin/pip3 pip3 /usr/bin/pip3.9 1
       elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
           brew install python@3.9
+          ln -s /usr/local/opt/python@3.9/bin/python3.9 /usr/local/bin/python3
+          ln -s /usr/local/opt/python@3.9/bin/pip3.9 /usr/local/bin/pip3
       fi
   - sh: |
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
@@ -100,11 +102,11 @@ install:
       pip install pytest-timeout
       pip install ruff
   - sh: |
-      $PYTHON -m pip install --upgrade urllib3>=2.0.5
-      $PYTHON -m pip install .
-      $PYTHON -m pip install pytest
-      $PYTHON -m pip install pytest-timeout
-      $PYTHON -m pip install ruff
+      python3 -m pip install --upgrade urllib3>=2.0.5
+      python3 -m pip install .
+      python3 -m pip install pytest
+      python3 -m pip install pytest-timeout
+      python3 -m pip install ruff
 
 # Exit if error
   - ps: if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio" -and $LastExitCode -ne 0) { Write-Host "Command exited with code $LastExitCode" }
