+++ appveyor.yml
@@ -24,23 +24,22 @@ environment:
 matrix:
   fast_finish: true
   
-
 install:
   # Install Python 3.9 for max and linux
   - sh: |
-    if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
-        sudo add-apt-repository ppa:deadsnakes/ppa -y
-        sudo apt-get update
-        sudo apt-get install -y python3.9 python3.9-venv python3.9-dev python3-pip
-        python3.9 -m pip install --upgrade pip
-        which python3.9
-        sudo update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.9 1
-        sudo update-alternatives --install /usr/bin/pip3 pip3 /usr/bin/pip3.9 1
-    elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
-        brew install python@3.9
-        ln -s /usr/local/opt/python@3.9/bin/python3.9 /usr/local/bin/python3
-        ln -s /usr/local/opt/python@3.9/bin/pip3.9 /usr/local/bin/pip3
-    fi
+      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
+          sudo add-apt-repository ppa:deadsnakes/ppa -y
+          sudo apt-get update
+          sudo apt-get install -y python3.9 python3.9-venv python3.9-dev python3-pip
+          python3.9 -m pip install --upgrade pip
+          which python3.9
+          sudo update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.9 1
+          sudo update-alternatives --install /usr/bin/pip3 pip3 /usr/bin/pip3.9 1
+      elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
+          brew install python@3.9
+          ln -s /usr/local/opt/python@3.9/bin/python3.9 /usr/local/bin/python3
+          ln -s /usr/local/opt/python@3.9/bin/pip3.9 /usr/local/bin/pip3
+      fi
       
   - sh: |
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
