+++ appveyor.yml
@@ -6,7 +6,7 @@ environment:
     PYLSL_LIB: lsl.dll
   matrix:
     - ID: Ubuntu
-      APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2104
+      APPVEYOR_BUILD_WORKER_IMAGE: ubuntu2104
       PY:  3.9
     - ID: macOS
       APPVEYOR_BUILD_WORKER_IMAGE: macOS
