+++ appveyor.yml
@@ -82,7 +82,6 @@ install:
       fi
 # install dependencies
   - ps: |
-      pip uninstall -y urllib3
       pip install --upgrade urllib3>=2.0.5
       pip install .
       pip install pytest
