+++ appveyor.yml
@@ -22,41 +22,27 @@ environment:
 
 matrix:
   fast_finish: true
-  
-install:
 
+install:
   - sh: |
-      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
-          PYTHON="/usr/bin"
-          export PATH=$PYTHON:$PYTHON/Scripts:$PATH
-          sudo apt-get update
-          sudo apt-get install -y python3-pip
-          python3 -m pip install --upgrade pip
-      elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
-          export PATH=$PYTHON:$PYTHON/Scripts:$PATH
-      fi
+      case $APPVEYOR_BUILD_WORKER_IMAGE in
+      "ubuntu")
+        sudo apt-get update
+        sudo apt-get install -y python3-pip
+        python3 -m pip install --upgrade pip
+        ;;
+      "macOS")
+        export PATH=$PY:$PY/Scripts:$PATH
+        ;;
+      esac
+
   - ps: |
-      if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") {
+      switch ($env:APPVEYOR_BUILD_WORKER_IMAGE) {
+        "Visual Studio 2019" {
           $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH"
+        }
       }
-      
-  - ps: $ErrorActionPreference = 'Stop'; if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") { $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH" }
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
-  - sh: if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then set -e; export PATH=$PYTHON:$PYTHON/Scripts:$PATH; fi
-  # Upgrade pip, setuptools, wheel
-  - sh: |
-      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" || $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
-          python3 -m pip install --upgrade pip
-          python3 --version
-          pip3 --version
-      fi
-  - ps: |
-      if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") {
-          python -m pip install --upgrade pip
-          python --version
-          pip --version
-      }
-  
+
   # Download and unzip liblsl
   - ps: |
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") {
@@ -64,43 +50,44 @@ install:
           Expand-Archive -Path liblsl.zip -DestinationPath liblsl
       }
   - sh: |
-      if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu" ]]; then
-          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.deb -O liblsl.deb
-          dpkg -x liblsl.deb liblsl
-      elif [[ $APPVEYOR_BUILD_WORKER_IMAGE == "macOS" ]]; then
-          wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-OSX_amd64.tar.bz2 -O liblsl.tar.bz2
-          tar xjf liblsl.tar.bz2 -C liblsl
-      fi
+      case $APPVEYOR_BUILD_WORKER_IMAGE in
+      "ubuntu")
[…]
