### 2023-10-27 #5706 @emdupre
👋  Hi everyone, just a small note that I will be out-of-office 30 October - 10 November. I will answer any comments or questions as soon as I'm back, though, so please don't hesitate to ping this thread.
