+++ appveyor.yml
@@ -83,8 +83,8 @@ install:
       python3 -m pip install pytest pytest-timeout ruff
 
 # Exit if error
-  - ps: if ($LastExitCode -ne 0) { Write-Host "Command exited with code $LastExitCode"; exit $LastExitCode }
-  - sh: if [[ $? != 0 ]]; then echo "Command exited with code $?"; exit $?; fi
+  #- ps: if ($LastExitCode -ne 0) { Write-Host "Command exited with code $LastExitCode"; exit $LastExitCode }
+  #- sh: if [[ $? != 0 ]]; then echo "Command exited with code $?"; exit $?; fi
 
 test_script:
   # Run tests
