+++ appveyor.yml
@@ -94,6 +94,8 @@ test_script:
   - sh: |
       if [[ $APPVEYOR_BUILD_WORKER_IMAGE == "ubuntu2204" ]]; then
         echo 'About to run tests on Ubuntu...'
+        ls -alh $PWD/liblsl/usr/lib/
+        echo $PYLSL_LIB
         which python3
         python3 --version
         ls -alh Tests/
