+++ appveyor.yml
@@ -61,7 +61,7 @@ install:
         conda update -q conda
         # Create a new conda environment with the desired Python version
         conda create -q -n test-environment python=$PYTHON
-        source activate test-environment
+        conda activate test-environment  # Updated this line
         # Install LSL
         conda install -c conda-forge liblsl
         ;;
