+++ appveyor.yml
@@ -31,28 +31,6 @@ install:
         sudo apt-get update
         sudo apt-get install -y python3-pip
         python3 -m pip install --upgrade pip
-        ;;
-      "macOS")
-        export PATH=$PY:$PY/Scripts:$PATH
-        ;;
-      esac
-
-  - ps: |
-      switch ($env:APPVEYOR_BUILD_WORKER_IMAGE) {
-        "Visual Studio 2019" {
-          $env:PATH = "$env:PYTHON;$env:PYTHON\Scripts;$env:PATH"
-        }
-      }
-
-  # Download and unzip liblsl
-  - ps: |
-      if ($env:APPVEYOR_BUILD_WORKER_IMAGE -match "Visual Studio 2019") {
-          Invoke-WebRequest -Uri 'https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-Win_amd64.zip' -OutFile 'liblsl.zip'
-          Expand-Archive -Path liblsl.zip -DestinationPath liblsl
-      }
-  - sh: |
-      case $APPVEYOR_BUILD_WORKER_IMAGE in
-      "ubuntu")
         # Install Miniconda
         wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
         bash miniconda.sh -b -p $HOME/miniconda
@@ -61,15 +39,9 @@ install:
         conda update -q conda
         # Create a new conda environment with the desired Python version
         conda create -q -n test-environment python=$PYTHON
-        conda activate test-environment  # Updated this line
+        source activate test-environment || conda activate test-environment  # Updated this line
         # Install LSL
         conda install -c conda-forge liblsl
-        # Diagnostic: List the contents of the Conda environment
-        conda list
-        # Diagnostic: Locate the LSL library
-        find $HOME/miniconda -name 'liblsl*'
-        # Diagnostic: Check the library path
-        echo $LD_LIBRARY_PATH
         ;;
       "macOS")
         # Install Brew (if not already installed)
@@ -78,6 +50,12 @@ install:
         brew install sccn/labstreaminglayer/lsl
         ;;
       esac
+  - sh: |
+      # Diagnostic Commands (separated from the above block to ensure they run)
+      echo 'Diagnostic Information:'
+      conda list || echo 'Failed to list conda packages.'
+      find $HOME/miniconda -name 'liblsl*' || echo 'Failed to find liblsl.'
+      echo $LD_LIBRARY_PATH
   
  # Move DLL to correct directory (Windows) or Set PYLSL_LIB environment variable (Linux, macOS)
   - ps: |
