+++ appveyor.yml
@@ -33,10 +33,15 @@ install:
         echo "Updating system..."
         sudo apt-get update
 
-        echo "Installing LSL dependency..."
-        wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.deb -O liblsl.deb
-        dpkg -x liblsl.deb liblsl
-        export PYLSL_LIB=$PWD/liblsl/usr/lib/liblsl64.so.1.16.0  # Update this line
+        wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
+        sha256sum Miniconda3-latest-Linux-x86_64.sh
+        bash Miniconda3-latest-Linux-x86_64.sh
+        conda install -c conda-forge liblsl
+        
+        #echo "Installing LSL dependency..."
+        #wget https://github.com/sccn/liblsl/releases/download/v1.16.2/liblsl-1.16.2-bionic_amd64.deb -O liblsl.deb
+        #dpkg -x liblsl.deb liblsl
+        #export PYLSL_LIB=$PWD/liblsl/usr/lib/liblsl64.so.1.16.0  # Update this line
         
         #sudo apt-get install -y python3-pip cmake git
         #echo "Cloning liblsl repository..."
