+++ appveyor.yml
@@ -91,9 +91,7 @@ test_script:
         echo 'About to run tests on Ubuntu...'
         # Diagnostic commands
         echo "PYLSL_LIB: $PYLSL_LIB"
-        ls -alh /usr/local/lib/
         ls -alh /home/appveyor/.local/lib/python3.10/site-packages/pylsl/lib
-        ls -alh $PWD/liblsl/usr/lib/
         echo $PYLSL_LIB
         which python3
         python3 --version
