### 2023-10-28 5fb5896
Update test_PseudoDevice.py
