+++ Tests/test_Pytorch.py
@@ -64,21 +64,20 @@ def test_run_bci():
     accuracy_achieved = False
     marker_received = False
     accuracy=None
-    assert True
-    #while True:
-    #    currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
-    #    time.sleep(0.5) # wait for marker updates
+    while True:
+        currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
+        time.sleep(0.5) # wait for marker updates
         #print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
-    #    if len(currentMarkers) > 1:  # check there is more then one marker type received
-    #        marker_received = True
-    #        if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
-    #            classInfo = bci.CurrentClassifierInfo() # hangs if called too early
-    #            accuracy = classInfo["accuracy"]###
-    #            if accuracy >= 0:
+        if len(currentMarkers) > 1:  # check there is more then one marker type received
+            marker_received = True
+            if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
+                classInfo = bci.CurrentClassifierInfo() # hangs if called too early
+                accuracy = classInfo["accuracy"]###
+                if accuracy >= 0:
                     # set to above 0 to show some accuracy was retruend from model
-    #                accuracy_achieved = True
-    #                bci.StopThreads()
-    #                break
+                    accuracy_achieved = True
+                    bci.StopThreads()
+                    break
             #if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired+4:
             #    break
-    #assert accuracy_achieved and marker_received
+    assert accuracy_achieved and marker_received
