+++ pybci/Examples/testSimple.py
@@ -2,16 +2,21 @@ from pybci import PyBCI
 import time 
 
 if __name__ == '__main__': # Note: this line is needed when calling pseudoDevice as by default runs in a multiprocessed operation
+    
+    print("Starting bci")
     bci = PyBCI(minimumEpochsRequired = 5, createPseudoDevice=True)
+    print("attemting connect to bci")
     while not bci.connected: # check to see if lsl marker and datastream are available
         bci.Connect()
         time.sleep(1)
+    print("Connected to device")
     bci.TrainMode() # now both marker and datastreams available start training on received epochs
+    print("Training mode started")
     accuracy = 0
     try:
         while(True):
             currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
-            time.sleep(0.5) # wait for marker updates
+            time.sleep(0.1) # wait for marker updates
             print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
             if len(currentMarkers) > 1:  # check there is more then one marker type received
                 if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
