+++ pybci/Utils/FeatureExtractor.py
@@ -2,7 +2,6 @@ import numpy as np
 from scipy.signal import welch
 
 import warnings
-import antropy
 from ..Configuration.FeatureSettings import GeneralFeatureChoices
 # Filter out UserWarning messages from the scipy package, could be worth moving to init and applying printdebug print levels? (typically nans, 0 and infs causing errors)
 warnings.filterwarnings("ignore", category=UserWarning, module="scipy") # used to reduce print statements from constant signals being applied
@@ -10,6 +9,7 @@ warnings.filterwarnings("ignore", category=UserWarning, module="antropy") # used
 warnings.filterwarnings("ignore", category=RuntimeWarning, module="antropy") # used to reduce print statements from constant signals being applied
 warnings.filterwarnings("ignore", category=RuntimeWarning, module="numpy") # used to reduce print statements from constant signals being applied
 #warnings.filterwarnings("ignore", category=RuntimeWarning, module="pybci") # used to reduce print statements from constant signals being applied
+import antropy
 
 
 class GenericFeatureExtractor():
