+++ Tests/test_Pytorch.py
@@ -56,24 +56,24 @@ def PyTorchModel(x_train, x_test, y_train, y_test):
 #@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_bci():
     pdc = PseudoDeviceController(execution_mode="thread")
-    bci = PyBCI(minimumEpochsRequired = 4, createPseudoDevice=True,pseudoDeviceController=pdc, torchModel = PyTorchModel)
+    bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True,torchModel = PyTorchModel, pseudoDeviceController=pdc)
     while not bci.connected:
         bci.Connect()
         time.sleep(1)
     bci.TrainMode()
     accuracy_achieved = False
     marker_received = False
-    accuracy=None
+    accuracy=0
     while True:
         currentMarkers = bci.ReceivedMarkerCount() # check to see how many received epochs, if markers sent to close together will be ignored till done processing
         time.sleep(0.5) # wait for marker updates
-        #print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
+        print("Markers received: " + str(currentMarkers) +" Accuracy: " + str(round(accuracy,2)), end="         \r")
         if len(currentMarkers) > 1:  # check there is more then one marker type received
             marker_received = True
             if min([currentMarkers[key][1] for key in currentMarkers]) > bci.minimumEpochsRequired:
                 classInfo = bci.CurrentClassifierInfo() # hangs if called too early
                 accuracy = classInfo["accuracy"]###
-                if accuracy >= 0:
+                if accuracy > 0:
                     # set to above 0 to show some accuracy was retruend from model
                     accuracy_achieved = True
                     bci.StopThreads()
