### 2023-10-28 LMBooth/pybci#14 @jsheunis
> Since CLI based examples are available i feel adding argument parsing to the other examples will increase the cognitive workload to deduce what is necessary or not. Would you agree that with the addition of the CLI examples the addition of arg passing to the other examples is no longer needed and makes them clearer to read and understand?

Yes I agree 👍
