### 2023-10-28 94d3f25
Update LSLScanner.py
