### 2023-10-28 743dee9
remove debug prints for ubuntu
waiting till can test ubuntu with own device, keeps printing marker stream cannot be found.
