+++ pybci/Utils/PseudoDevice.py
@@ -30,7 +30,6 @@ class PseudoDeviceController:
             self.command_queue = queue.Queue()
             self.worker = threading.Thread(target=self._run_device)
         else:
-            print("got this mode:"+execution_mode)
             raise ValueError(f"Unsupported execution mode: {execution_mode}")
 
         self.worker.start()
@@ -38,17 +37,17 @@ class PseudoDeviceController:
     def _run_device(self):
         device = PseudoDevice(*self.args, **self.kwargs, stop_signal=self.stop_signal)
         while not self.stop_signal.is_set():
-            #if self.execution_mode == 'process':
-            try:
-                command = self.command_queue.get_nowait()
-                if command == "BeginStreaming":
-                    device.BeginStreaming()
-            except queue.Empty:
-                pass
+            if self.execution_mode == 'process':
+                try:
+                    command = self.command_queue.get_nowait()
+                    if command == "BeginStreaming":
+                        device.BeginStreaming()
+                except queue.Empty:
+                    pass
             #elif self.execution_mode == 'thread':
             #    device.update()
 
-            time.sleep(0.5)
+            time.sleep(0.01)
 
     def BeginStreaming(self):
         #if self.execution_mode == 'process':
@@ -100,7 +99,7 @@ class PseudoDevice:
         self.pseudoMarkerConfig = pseudoMarkerConfig
         self.sampleRate = sampleRate
         self.channelCount = channelCount
-        markerInfo = pylsl.StreamInfo(pseudoMarkerConfig.markerName, pseudoMarkerConfig.markerType, 1, 0, 'string', 'DevMarker')
+        markerInfo = pylsl.StreamInfo(pseudoMarkerConfig.markerName, pseudoMarkerConfig.markerType, 1, 0, 'string', 'Dev')
         self.markerOutlet = pylsl.StreamOutlet(markerInfo)
         info = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
         chns = info.desc().append_child("channels")
