+++ pybci/Utils/LSLScanner.py
@@ -66,8 +66,6 @@ class LSLScanner:
         markerStreams = []
         self.markerStream = None
         for stream in streams:
-            print(stream.name())
-            print(stream.type())
             if stream.type() in self.markerTypes:
                 markerStreams.append(StreamInlet(stream))
         if self.markerStreamPredefined:
