+++ appveyor.yml
@@ -6,9 +6,6 @@ environment:
     PYLSL_LIB: /home/appveyor/.local/lib/python3.10/site-packages/pylsl/lib/liblsl.so
     #PYLSL_LIB: lsl.dll
   matrix:
-    - ID: Ubuntu
-      APPVEYOR_BUILD_WORKER_IMAGE: ubuntu2204
-      PYTHON: 3.10  # Added Python version
     - ID: macOS
       APPVEYOR_BUILD_WORKER_IMAGE: macOS
       PYTHON: 3.9
@@ -21,6 +18,9 @@ environment:
     - ID: VS2019_Py39
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       PYTHON: "C:\\Python39-x64"
+    - ID: Ubuntu
+      APPVEYOR_BUILD_WORKER_IMAGE: ubuntu2204
+      PYTHON: 3.10  # Added Python version
 
 matrix:
   fast_finish: true
