### 2023-10-28 d263565
tweaks and corrections to doc styling and syntax use
