+++ Tests/test_PseudoDevice.py
@@ -7,8 +7,10 @@ def test_run_pseudo():
     pd.BeginStreaming()
     time.sleep(5)
     pd.StopStreaming()
+    time.sleep(5)
 
     pd = PseudoDeviceController(execution_mode="process")
     pd.BeginStreaming()
     time.sleep(5)
     pd.StopStreaming()
+    time.sleep(5)
