### 2023-10-28 9775758
Update test_PseudoDevice.py
