+++ pybci/pybci.py
@@ -109,7 +109,7 @@ class PyBCI:
                 current_script_path = os.path.abspath(__file__)
                 current_script_path = current_script_path.replace("pybci.py", "", 1)  
                 print("Mac current_script_path: "+current_script_path)
-                desiredpath = current_script_path + "/Utils/PseudoSubprocess.py"
+                desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
                 self.process = subprocess.Popen(["python3",desiredpath], stdin=subprocess.PIPE)
                 self.process.stdin.write(b'begin\n')
                 self.process.stdin.flush()
@@ -118,7 +118,7 @@ class PyBCI:
                 current_script_path = os.path.abspath(__file__)
                 current_script_path = current_script_path.replace("pybci.py", "", 1)  
                 print("Linux current_script_path: "+current_script_path) 
-                desiredpath = current_script_path + "/PseudoSubprocess.py"
+                desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
                 self.process = subprocess.Popen(["python3", desiredpath], stdin=subprocess.PIPE)
                 self.process.stdin.write(b'begin\n')
                 self.process.stdin.flush()
