+++ pybci/Utils/PseudoSubprocess.py
@@ -1,4 +1,5 @@
-from PyBCI.Utils.PseudoDevice import PseudoDeviceController
+
+from pybci import PyBCIfrom PyBCI.Utils.PseudoDevice import PseudoDeviceController
 import sys
 
 class PseudoSubprocess:
