+++ Tests/pytest.ini
@@ -1,3 +1,1 @@
 [pytest]
-filterwarnings =
-    ignore::DeprecationWarning:sre_constants:19
