+++ pybci/pybci.py
@@ -106,28 +106,30 @@ class PyBCI:
                     pseudoDevice.BeginStreaming()
                 self.pseudoDevice = pseudoDevice
             elif current_os == "macOS":
-                current_script_path = os.path.abspath(__file__)
-                current_script_path = current_script_path.replace("pybci.py", "", 1)  
-                print("Mac current_script_path: "+current_script_path)
-                desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
-                self.process = subprocess.Popen([sys.executable,"-u", desiredpath], stdin=subprocess.PIPE)
-                print(f"Return code: {self.process.returncode}")
-                self.process.stdin.write(b'begin\n')
-                print(f"Return code: {self.process.returncode}")
-                self.process.stdin.flush()
-                print(f"Return code: {self.process.returncode}")
+                self.pseudoDevice = pseudoDevice
+                #current_script_path = os.path.abspath(__file__)
+                #current_script_path = current_script_path.replace("pybci.py", "", 1)  
+                #print("Mac current_script_path: "+current_script_path)
+                #desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
+                #self.process = subprocess.Popen([sys.executable,"-u", desiredpath], stdin=subprocess.PIPE)
+                #print(f"Return code: {self.process.returncode}")
+                #self.process.stdin.write(b'begin\n')
+                #print(f"Return code: {self.process.returncode}")
+                #self.process.stdin.flush()
+                #print(f"Return code: {self.process.returncode}")
 
             elif current_os == "Linux":
-                current_script_path = os.path.abspath(__file__)
-                current_script_path = current_script_path.replace("pybci.py", "", 1)  
-                print("Linux current_script_path: "+current_script_path) 
-                desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
-                self.process = subprocess.Popen([sys.executable,"-u",  desiredpath], stdin=subprocess.PIPE)
-                print(f"Return code: {self.process.returncode}")
-                self.process.stdin.write(b'begin\n')
-                print(f"Return code: {self.process.returncode}")
-                self.process.stdin.flush()
-                print(f"Return code: {self.process.returncode}")
+                self.pseudoDevice = pseudoDevice
+                #current_script_path = os.path.abspath(__file__)
+                #current_script_path = current_script_path.replace("pybci.py", "", 1)  
+                #print("Linux current_script_path: "+current_script_path) 
+                #desiredpath = current_script_path + "Utils/PseudoSubprocess.py"
+                #self.process = subprocess.Popen([sys.executable,"-u",  desiredpath], stdin=subprocess.PIPE)
+                #print(f"Return code: {self.process.returncode}")
+                #self.process.stdin.write(b'begin\n')
+                #print(f"Return code: {self.process.returncode}")
+                #self.process.stdin.flush()
+                #print(f"Return code: {self.process.returncode}")
         self.lslScanner = LSLScanner(self, dataStreams, markerStream,streamTypes, markerTypes, logger =self.logger)
         self.ConfigureMachineLearning(minimumEpochsRequired,  clf, model, torchModel) # configure first, connect second   
         #self.Connect()
