+++ Tests/test_Simple.py
@@ -1,9 +1,17 @@
-from pybci import PyBCI
+from pybci import PyBCI, get_os
 import time
+from pybci.Utils.PseudoDevice import PseudoDeviceController
 # Test case using the fixture
 #@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_bci():
-    bci = PyBCI(minimumEpochsRequired=4, createPseudoDevice=True, loggingLevel="TIMING")
+    current_os = get_os()
+    if current_os == "Windows":
+        bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True,)
+    else:
+        pdc = PseudoDeviceController(execution_mode="process")
+        pdc.BeginStreaming()
+        time.sleep(10)
+        bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True, pseudoDeviceController=pdc)
     while not bci.connected:
         bci.Connect()
         time.sleep(1)
