+++ Tests/test_Pytorch.py
@@ -1,9 +1,11 @@
-from pybci import PyBCI
+from pybci import PyBCI, get_os
+
 import time
 import torch
 from torch.utils.data import DataLoader, TensorDataset
 from torch import nn
 from pybci.Utils.PseudoDevice import PseudoDeviceController
+
 
 class SimpleNN(nn.Module):
     def __init__(self, input_size, hidden_size, num_classes):
@@ -55,10 +57,14 @@ def PyTorchModel(x_train, x_test, y_train, y_test):
 
 #@pytest.mark.timeout(300)  # Extended timeout to 5 minutes
 def test_run_bci():
-    pdc = PseudoDeviceController(execution_mode="process")
-    pdc.BeginStreaming()
-    time.sleep(10)
-    bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True,torchModel = PyTorchModel, pseudoDeviceController=pdc)
+    current_os = get_os()
+    if current_os == "Windows":
+        bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True,torchModel = PyTorchModel)
+    else:
+        pdc = PseudoDeviceController(execution_mode="process")
+        pdc.BeginStreaming()
+        time.sleep(10)
+        bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True,torchModel = PyTorchModel, pseudoDeviceController=pdc)
     while not bci.connected:
         bci.Connect()
         time.sleep(1)
