### 2023-10-28 1089c5a
test_cli fix
