+++ Tests/test_cli.py
@@ -12,49 +12,49 @@ def test_cli_simple_timeout():
         timeout = 30  # timeout in seconds
         my_bci_wrapper = None
 
-        def run_main():
-            nonlocal my_bci_wrapper
-            my_bci_wrapper = mainSimple(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
+        #def run_main():
+        #    nonlocal my_bci_wrapper
+        my_bci_wrapper = mainSimple(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
 
-        main_thread = threading.Thread(target=run_main)
-        main_thread.start()
-        main_thread.join()
+        #main_thread = threading.Thread(target=run_main)
+        #main_thread.start()
+        #main_thread.join()
 
 def test_cli_sklearn_timeout():
     with patch('builtins.input', return_value='stop'):
         timeout = 30  # timeout in seconds
         my_bci_wrapper = None
 
-        def run_main():
-            nonlocal my_bci_wrapper
-            my_bci_wrapper = mainSklearn(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
+        #def run_main():
+        #    nonlocal my_bci_wrapper
+        my_bci_wrapper = mainSklearn(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
 
-        main_thread = threading.Thread(target=run_main)
-        main_thread.start()
-        main_thread.join()
+        #main_thread = threading.Thread(target=run_main)
+        #main_thread.start()
+        #main_thread.join()
 
 def test_cli_pytorch_timeout():
     with patch('builtins.input', return_value='stop'):
         timeout = 30  # timeout in seconds
         my_bci_wrapper = None
 
-        def run_main():
-            nonlocal my_bci_wrapper
-            my_bci_wrapper = mainPyTorch(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
+        #def run_main():
+        #    nonlocal my_bci_wrapper
+        my_bci_wrapper = mainPyTorch(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
 
-        main_thread = threading.Thread(target=run_main)
-        main_thread.start()
-        main_thread.join()
+        #main_thread = threading.Thread(target=run_main)
+        #main_thread.start()
+        #main_thread.join()
 
 def test_cli_tensorflow_timeout():
     with patch('builtins.input', return_value='stop'):
         timeout = 30  # timeout in seconds
         my_bci_wrapper = None
 
-        def run_main():
-            nonlocal my_bci_wrapper
-            my_bci_wrapper = mainTensorflow(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
+        #def run_main():
+            #nonlocal my_bci_wrapper
+        my_bci_wrapper = mainTensorflow(createPseudoDevice=True, min_epochs_train=1, min_epochs_test=2, timeout=timeout)
 
-        main_thread = threading.Thread(target=run_main)
-        main_thread.start()
-        main_thread.join()
+        #main_thread = threading.Thread(target=run_main)
+        #main_thread.start()
+        #main_thread.join()
