+++ pybci/CliTests/testSklearn.py
@@ -1,4 +1,5 @@
-from ..pybci import PyBCI
+from ..pybci import PyBCI, get_os
+from pybci.Utils.PseudoDevice import PseudoDeviceController
 import time
 import argparse
 from sklearn.neural_network import MLPClassifier
@@ -24,7 +25,16 @@ class CLI_testSklearnWrapper:
         if self.min_epochs_test <= self.min_epochs_train:
             self.min_epochs_test = self.min_epochs_train+1
         clf = MLPClassifier(max_iter = 1000, solver ="lbfgs")#solver=clf, alpha=alpha,hidden_layer_sizes=hid)
-        self.bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=createPseudoDevice, clf = clf)
+        current_os = get_os()
+        if current_os == "Windows":
+            self.bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True, clf = clf)
+        else:
+            pdc = PseudoDeviceController(execution_mode="process")
+            pdc.BeginStreaming()
+            time.sleep(10)
+            self.bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True, pseudoDeviceController=pdc, clf = clf)
+
+        #self.bci = PyBCI(minimumEpochsRequired = min_epochs_train, createPseudoDevice=createPseudoDevice, clf = clf)
         main_thread = threading.Thread(target=self.loop)
         main_thread.start()
         if self.timeout:
