### 2023-10-29 LMBooth/pybci#25 @anilbey
Constant warning messages while running the exercises
Hi, I am having difficulty running the exercises.

I installed `pybci` and tried the following tests under the Examples directory

- testSimple.py
- testSklearn.py

Both of them keep giving the following warnings in a loop.

```
PyBCI: [WARNING] - No Marker streams available, make sure your accepted marker data Type have been set in bci.lslScanner.markerTypes correctly.
PyBCI: [WARNING] - No data streams available, make sure your streamTypes have been set in bci.lslScanner.dataStream correctly.

```

In `testSimple.py` only the first 2 print messages below are displayed. The rest of them do not get displayed.

> Starting bci
> attemting connect to bci

Am I missing some dependency/configuration? Are the examples running fine on the CI?

The full log is below. The system I am testing is an Ubuntu 22.04 LTS

```
❯ python testSimple.py
Starting bci
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'lo' (status: 1, multicast: 0, broadcast: 0)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'enp0s31f6' (status: 1, multicast: 4096, broadcast: 2)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'lo' (status: 1, multicast: 0, broadcast: 0)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'enp0s31f6' (status: 1, multicast: 4096, broadcast: 2)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:102   INFO|       IPv4 addr: c0a80108
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'lo' (status: 1, multicast: 0, broadcast: 0)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'enp0s31f6' (status: 1, multicast: 4096, broadcast: 2)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:105   INFO|       IPv6 addr: 2a04:ee41:82:7519:76d7:8d91:667c:f269
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'enp0s31f6' (status: 1, multicast: 4096, broadcast: 2)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:105   INFO|       IPv6 addr: 2a04:ee41:82:7519:55b3:6db:9a75:4d6b
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:89    INFO| netif 'enp0s31f6' (status: 1, multicast: 4096, broadcast: 2)
2023-10-29 20:13:50.492 (   1.384s) [python          ]      netinterfaces.cpp:105   INFO|       IPv6 addr: fe80::f07c:2a0a:1f5d:1f06%enp0s31f6
2023-10-29 20:13:50.492 (   1.384s) [python          ]         api_config.cpp:270   INFO| Loaded default config
PyBCI: [INFO] - Invalid or no sklearn classifier passed to clf. Checking tensorflow model...
2023-10-29 20:13:52.689121: E tensorflow/compiler/xla/stream_executor/cuda/cuda_dnn.cc:9342] Unable to register cuDNN factory: Attempting to register factory for plugin cuDNN when one has already been registered
2023-10-29 20:13:52.689167: E tensorflow/compiler/xla/stream_executor/cuda/cuda_fft.cc:609] Unable to register cuFFT factory: Attempting to register factory for plugin cuFFT when one has already been registered
2023-10-29 20:13:52.689206: E tensorflow/compiler/xla/stream_executor/cuda/cuda_blas.cc:1518] Unable to register cuBLAS factory: Attempting to register factory for plugin cuBLAS when one has already been registered
2023-10-29 20:13:52.695314: I tensorflow/core/platform/cpu_feature_guard.cc:182] This TensorFlow binary is optimized to use available CPU instructions in performance-critical operations.
To enable the following instructions: AVX2 FMA, in other operations, rebuild TensorFlow with the appropriate compiler flags.
2023-10-29 20:13:53.352814: W tensorflow/compiler/tf2tensorrt/utils/py_utils.cc:38] TF-TRT Warning: Could not find TensorRT
PyBCI: [INFO] - Invalid or no tensorflow model passed to model.  Checking pytorch torchModel...
PyBCI: [INFO] - Invalid or no PyTorch model passed to model. Defaulting to SVM by SkLearn
attemting connect to bci
PyBCI: [WARNING] - No Marker streams available, make sure your accepted marker data Type have been set in bci.lslScanner.markerTypes correctly.
PyBCI: [WARNING] - No data streams available, make sure your streamTypes have been set in bci.lslScanner.dataStream correctly.
PyBCI: [WARNING] - No Marker streams available, make sure your accepted marker data Type have been set in bci.lslScanner.markerTypes correctly.
PyBCI: [WARNING] - No data streams available, make sure your streamTypes have been set in bci.lslScanner.dataStream correctly.
PyBCI: [WARNING] - No Marker streams available, make sure your accepted marker data Type have been set in bci.lslScanner.markerTypes correctly.
...
```

This is part of the https://github.com/openjournals/joss-reviews/issues/5706 review.
