### 2023-10-29 #5706 @anilbey
Hi, @emdupre I am still having a problem running the exercises. I created an issue with the details. I will update the checklist once I can run the exercises.
