### 2023-10-29 LMBooth/pybci#25 @LMBooth
Hi @anilbey thank you for trying, this also relates to issue #24. For unix based operating systems there is an issue with starting lsl threads outside of the main thread, I'd only tested the pseudo device on my windows machine but since integrating in to appveyor I've been attempting a fix and plan to update the documentation accordingly where the user will be required to initialise the pseudo device and pass as an argument to the pybci class.

I'll be home later in the week and will hopefully have a fix when appveyor returns that all the tests complete successfully.
