### 2023-10-30 1679def
added stopstreaming in pseudodevicecontroller of ps
