+++ appveyor.yml
@@ -30,6 +30,7 @@ install:
   - ps: |
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -eq "Visual Studio 2019") {
         # Windows-specific setup here
+        pip install --upgrade setuptools wheel
         pip install --upgrade urllib3>=2.0.5
         pip install .
         pip install pytest pytest-timeout ruff
