+++ appveyor.yml
@@ -30,6 +30,8 @@ install:
   - ps: |
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -eq "Visual Studio 2019") {
         # Windows-specific setup here
+        python -m pip install --upgrade pip
+        pip uninstall -y urllib3
         pip install --upgrade setuptools wheel
         pip install --upgrade urllib3>=2.0.5
         pip install .
