+++ appveyor.yml
@@ -8,13 +8,13 @@ environment:
   matrix:
     - ID: VS2019_Py311
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python311-x64"
+      PYTHON: "C:\Python311"
     - ID: VS2019_Py310
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python310-x64"
+      PYTHON: "C:\Python310"
     - ID: VS2019_Py39
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python39-x64"
+      PYTHON: "C:\Python39"
     - ID: macOS
       APPVEYOR_BUILD_WORKER_IMAGE: macos-monterey
       PYTHON: 3.9
