+++ appveyor.yml
@@ -8,13 +8,13 @@ environment:
   matrix:
     - ID: VS2019_Py311
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\Python311"
+      PY: 311-x64
     - ID: VS2019_Py310
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\Python310"
+      PY: 310-x64
     - ID: VS2019_Py39
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\Python39"
+      PY: 39-x64
     - ID: macOS
       APPVEYOR_BUILD_WORKER_IMAGE: macos-monterey
       PYTHON: 3.9
