+++ appveyor.yml
@@ -34,7 +34,7 @@ install:
         #SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%
         python -m pip install --upgrade pip
         python --version
-        dir
+        dir C:\
         pip --version
         #pip uninstall -y urllib3
         #pip install --upgrade setuptools wheel
