+++ appveyor.yml
@@ -35,7 +35,6 @@ install:
         where python
         python -m pip install --upgrade pip
         python --version
-        dir C:\
         pip --version
         #pip uninstall -y urllib3
         #pip install --upgrade setuptools wheel
