+++ appveyor.yml
@@ -30,7 +30,6 @@ install:
   - ps: |
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -eq "Visual Studio 2019") {
         # Windows-specific setup here
-        #set PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%
         set PATH="%PYTHON%;%PYTHON%\Scripts;%PATH%"
         where python
         python -m pip install --upgrade pip
