+++ appveyor.yml
@@ -35,8 +35,6 @@ install:
         python -m pip install --upgrade pip
         python --version
         pip --version
-        #pip uninstall -y urllib3
-        #pip install --upgrade setuptools wheel
         pip install --upgrade urllib3>=2.0.5
         pip install .
         pip install pytest pytest-timeout ruff
