+++ appveyor.yml
@@ -45,8 +45,8 @@ install:
           $env:PYTHONPATH = "%PYTHON_PATH%\Lib;%PYTHON_PATH%\Lib\site-packages"
           setx PYTHONPATH "$env:PYTHONPATH" /M
         }
-        echo PYTHONPATH
-        echo python
+        echo %PYTHON_PATH%
+        echo %PYTHON%
         where python
         python -m pip install --upgrade pip
         python --version
