+++ appveyor.yml
@@ -41,10 +41,10 @@ install:
         # Other commands
         python -m pip install --upgrade pip
         python --version
-        pip --version
-        pip install --upgrade urllib3>=2.0.5
-        pip install .
-        pip install pytest pytest-timeout ruff
+        python -m pip --version
+        python -m pip install --upgrade urllib3>=2.0.5
+        python -m pip install .
+        python -m pip install pytest pytest-timeout ruff
         $env:PYLSL_LIB = "lsl.dll"
       }
   - sh: |
