### 2023-11-01 1aee61b
pybci pseudo device stop update
