+++ Tests/pytest.ini
@@ -1,1 +1,3 @@
 [pytest]
+filterwarnings =
+    ignore::DeprecationWarning:tensorflow\.python\.framework\.dtypes:35
