+++ appveyor.yml
@@ -129,7 +129,7 @@ test_script:
         echo 'About to run tests on non-Ubuntu...'
         $env:PYTEST_ERRORLEVEL=0
         #try {
-        python -m pytest -vv -s Tests/ || PYTEST_ERRORLEVEL=$?
+        python -m pytest -vv -s Tests/ #|| PYTEST_ERRORLEVEL=$?
         #} catch {
         #  $env:PYTEST_ERRORLEVEL=$_.Exception.HResult
         #}
