+++ Tests/pytest.ini
@@ -1,3 +1,5 @@
 [pytest]
 filterwarnings =
     ignore::DeprecationWarning:tensorflow\.python\.framework\.dtypes:35
+    ignore::RuntimeWarning:numpy\.core\.fromnumeric:3504
+    ignore::RuntimeWarning:numpy\.core\._methods:129
