+++ appveyor.yml
@@ -127,7 +127,7 @@ test_script:
   - ps: |
       if ($env:APPVEYOR_BUILD_WORKER_IMAGE -ne "ubuntu2204") {
         $PYTEST_ERRORLEVEL = 0
-        pytest Tests/ -W "ignore::DeprecationWarning" -W "ignore::RuntimeWarning"
+        pytest -vv -s Tests/ 
         $PYTEST_ERRORLEVEL = $LASTEXITCODE
         if ($PYTEST_ERRORLEVEL -ne 0) {
           Write-Host 'Pytest failed. Terminating python process.'
