+++ appveyor.yml
@@ -5,6 +5,10 @@ environment:
   global:
     PYLSL_LIB: /home/appveyor/.local/lib/python3.10/site-packages/pylsl/lib/liblsl.so
   matrix:
+    - ID: Ubuntu
+      APPVEYOR_BUILD_WORKER_IMAGE: ubuntu2204
+      PYTHON: "3.10"
+      PYTHON_VERSION: "3.10"
     - ID: VS2019_Py311
       APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       PYTHON: "C:\\Python311-x64"
@@ -21,10 +25,7 @@ environment:
       APPVEYOR_BUILD_WORKER_IMAGE: macos-monterey
       PYTHON: "3.9"
       PYTHON_VERSION: "3.9"
-    - ID: Ubuntu
-      APPVEYOR_BUILD_WORKER_IMAGE: ubuntu2204
-      PYTHON: "3.10"
-      PYTHON_VERSION: "3.10"
+
 
 matrix:
   fast_finish: true
