+++ Tests/test_Simple.py
@@ -6,6 +6,11 @@ def test_run_bci():
     current_os = get_os()
     if current_os == "Windows":
         bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True)
+    elif current_os == "Linux":
+        pdc = PseudoDeviceController(execution_mode="thread")
+        pdc.BeginStreaming()
+        time.sleep(10)
+        bci = PyBCI(minimumEpochsRequired = 3, createPseudoDevice=True, pseudoDeviceController=pdc)
     else:
         pdc = PseudoDeviceController(execution_mode="process")
         pdc.BeginStreaming()
