+++ pybci/Utils/PseudoDevice.py
@@ -38,13 +38,13 @@ class PseudoDeviceController:
     def _run_device(self):
         device = PseudoDevice(*self.args, **self.kwargs, stop_signal=self.stop_signal)
         while not self.stop_signal.is_set():
-            if self.execution_mode == 'process':
-                try:
-                    command = self.command_queue.get_nowait()
-                    if command == "BeginStreaming":
-                        device.BeginStreaming()
-                except queue.Empty:
-                    pass
+            #if self.execution_mode == 'process':
+            try:
+                command = self.command_queue.get_nowait()
+                if command == "BeginStreaming":
+                    device.BeginStreaming()
+            except queue.Empty:
+                pass
             #elif self.execution_mode == 'thread':
             #    device.update()
 
