### 2023-11-02 2bf19e0
Update LSLScanner.py
