### 2023-11-02 6826d8a
Update LSLScanner.py
