### 2023-11-02 cc7c16c
Update LSLScanner.py
