### 2023-11-02 0c5ba25
Update LSLScanner.py
