+++ pybci/Utils/Logger.py
@@ -1,5 +1,5 @@
 import multiprocessing as mp
-mp.set_start_method('spawn')
+#mp.set_start_method('spawn')
 
 class Logger:
     INFO = "INFO"
