+++ pybci/Utils/PseudoDevice.py
@@ -104,11 +104,7 @@ class PseudoDevice:
         self.sampleRate = sampleRate
         self.channelCount = channelCount
         markerInfo = pylsl.StreamInfo(pseudoMarkerConfig.markerName, pseudoMarkerConfig.markerType, 1, 0, 'string', 'Dev')
-        self.markerOutlet = pylsl.StreamOutlet(markerInfo)
-        streams = pylsl.resolve_stream()
-        for stream in streams:
-            self.logger.log(Logger.INFO," Found stream name: " + stream.name())
-            self.logger.log(Logger.INFO," Found stream type: " + stream.type())
+        
         info = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
         chns = info.desc().append_child("channels")
         for label in range(self.channelCount):
@@ -116,11 +112,17 @@ class PseudoDevice:
             ch.append_child_value("label", str(label+1))
             ch.append_child_value("type", dataStreamType)
         self.outlet = pylsl.StreamOutlet(info)
-
+        
         streams = pylsl.resolve_stream()
         for stream in streams:
             self.logger.log(Logger.INFO," Found stream name: " + stream.name())
             self.logger.log(Logger.INFO," Found stream type: " + stream.type())
+        self.markerOutlet = pylsl.StreamOutlet(markerInfo)
+        streams = pylsl.resolve_stream()
+        for stream in streams:
+            
+            self.logger.log(Logger.INFO," Marker found stream name: " + stream.name())
+            self.logger.log(Logger.INFO," Marker found stream type: " + stream.type())
         self.last_update_time = time.time()
         self.phase_offset = 0.0
 
