+++ pybci/Utils/PseudoDevice.py
@@ -14,6 +14,7 @@ from multiprocessing import Process, Queue, Event
 import multiprocessing as mp
 #mp.set_start_method('spawn')
 import numpy as np
+import asyncio
 
 class PseudoDeviceController:
     log_queue = None
@@ -104,6 +105,7 @@ class PseudoDevice:
         self.sampleRate = sampleRate
         self.channelCount = channelCount
         self.markerInfo = pylsl.StreamInfo(self.pseudoMarkerConfig.markerName, self.pseudoMarkerConfig.markerType, 1, pylsl.IRREGULAR_RATE, 'string', 'Dev')
+        self.markerOutlet = pylsl.StreamOutlet(self.markerInfo)
         
         self.dataInfo = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
         chns = self.dataInfo.desc().append_child("channels")
@@ -112,16 +114,6 @@ class PseudoDevice:
             ch.append_child_value("label", str(label+1))
             ch.append_child_value("type", dataStreamType)
         self.outlet = pylsl.StreamOutlet(self.dataInfo)
-        
-        #streams = pylsl.resolve_stream()
-        #for stream in streams:
-        #    self.logger.log(Logger.INFO," Found stream name: " + stream.name())
-        #    self.logger.log(Logger.INFO," Found stream type: " + stream.type())
-        #self.markerOutlet = pylsl.StreamOutlet(self.markerInfo)
-        #streams = pylsl.resolve_stream()
-        #for stream in streams:   
-        #    self.logger.log(Logger.INFO," Marker found stream name: " + stream.name())
-        #    self.logger.log(Logger.INFO," Marker found stream type: " + stream.type())
         self.last_update_time = time.time()
         self.phase_offset = 0.0
 
@@ -145,7 +137,25 @@ class PseudoDevice:
         else:  # boolean flag for threads
             return self.stop_signal
 
-    def update(self):
+    async def _generate_signal(self):
+        #self.outlet = pylsl.StreamOutlet(self.dataInfo)
+        self.marker_task = asyncio.create_task(self._maker_timing())  # Create marker task after starting event loop
+        while not self._should_stop():
+            start_time = time.time()
+            await self.update()
+            end_time = time.time()
+            elapsed_time = end_time - start_time
+            expected_samples = int(self.sampleRate * elapsed_time)
+            if self.samples_generated >= expected_samples:
+                sleep_duration = max(0, (1.0 / 10) - elapsed_time)
+                await asyncio.sleep(sleep_duration)
+        #while not self._should_stop():
+        #    start_time = time.time()
+        #    await self.update()
+        #    sleep_duration = max(0, (1.0/10) - (start_time - time.time()))
+        #    await asyncio.sleep(sleep_duration)
+
+    async def update(self):
         with self.lock:
             if not self._should_stop():
                 current_time = time.time()
@@ -197,51 +207,33 @@ class PseudoDevice:
             self.stop_signal.clear()
         else:
             self.stop_signal = False
-        #else:  # For threading
-        self.thread = threading.Thread(target=self._generate_signal)
-        self.thread.start()
-        if self.pseudoMarkerConfig.autoplay:
-            self.StartMarkers()
-        self.log_message(Logger.INFO, " PseudoDevice - Begin streaming.")
-
-    def _generate_signal(self):
-        self.outlet = pylsl.StreamOutlet(self.dataInfo)
-        
-        while not self._should_stop():
-            start_time = time.time()
-            self.update()
-            sleep_duration = max(0, (1.0 / 10) - (start_time - time.time()))
-            time.sleep(sleep_duration)
-#            precise_sleep(sleep_duration)
[…]
