+++ pybci/Utils/PseudoDevice.py
@@ -11,10 +11,8 @@ import threading
 import pylsl
 import queue
 from multiprocessing import Process, Queue, Event
-import multiprocessing as mp
-#mp.set_start_method('spawn')
+import multiprocessing
 import numpy as np
-import asyncio
 
 class PseudoDeviceController:
     log_queue = None
@@ -40,13 +38,13 @@ class PseudoDeviceController:
     def _run_device(self):
         device = PseudoDevice(*self.args, **self.kwargs, stop_signal=self.stop_signal)
         while not self.stop_signal.is_set():
-            #if self.execution_mode == 'process':
-            try:
-                command = self.command_queue.get_nowait()
-                if command == "BeginStreaming":
-                    device.BeginStreaming()
-            except queue.Empty:
-                pass
+            if self.execution_mode == 'process':
+                try:
+                    command = self.command_queue.get_nowait()
+                    if command == "BeginStreaming":
+                        device.BeginStreaming()
+                except queue.Empty:
+                    pass
             #elif self.execution_mode == 'thread':
             #    device.update()
 
@@ -104,16 +102,15 @@ class PseudoDevice:
         self.pseudoMarkerConfig = pseudoMarkerConfig
         self.sampleRate = sampleRate
         self.channelCount = channelCount
-        self.markerInfo = pylsl.StreamInfo(self.pseudoMarkerConfig.markerName, self.pseudoMarkerConfig.markerType, 1, pylsl.IRREGULAR_RATE, 'string', 'Dev')
-        self.markerOutlet = pylsl.StreamOutlet(self.markerInfo)
-        
-        self.dataInfo = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
-        chns = self.dataInfo.desc().append_child("channels")
+        markerInfo = pylsl.StreamInfo(pseudoMarkerConfig.markerName, pseudoMarkerConfig.markerType, 1, 0, 'string', 'Dev')
+        self.markerOutlet = pylsl.StreamOutlet(markerInfo)
+        info = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
+        chns = info.desc().append_child("channels")
         for label in range(self.channelCount):
             ch = chns.append_child("channel")
             ch.append_child_value("label", str(label+1))
             ch.append_child_value("type", dataStreamType)
-        self.outlet = pylsl.StreamOutlet(self.dataInfo)
+        self.outlet = pylsl.StreamOutlet(info)
         self.last_update_time = time.time()
         self.phase_offset = 0.0
 
@@ -126,36 +123,18 @@ class PseudoDevice:
         return np.tile(full_signal, (self.channelCount, 1)).T
 
     def log_message(self, level='INFO', message = ""):
-        if self.log_queue is not None and isinstance(self.log_queue, type(mp.Queue)):
+        if self.log_queue is not None and isinstance(self.log_queue, type(multiprocessing.Queue)):
             self.log_queue.put(f'PyBCI: [{level}] - {message}')
         else:
             self.logger.log(level, message)
     
     def _should_stop(self):
-        if isinstance(self.stop_signal, mp.synchronize.Event):
+        if isinstance(self.stop_signal, multiprocessing.synchronize.Event):
             return self.stop_signal.is_set()
         else:  # boolean flag for threads
             return self.stop_signal
 
-    async def _generate_signal(self):
-        #self.outlet = pylsl.StreamOutlet(self.dataInfo)
-        self.marker_task = asyncio.create_task(self._maker_timing())  # Create marker task after starting event loop
-        while not self._should_stop():
-            start_time = time.time()
-            await self.update()
-            end_time = time.time()
[…]
