+++ pybci/Utils/PseudoDevice.py
@@ -104,6 +104,7 @@ class PseudoDevice:
         self.channelCount = channelCount
         markerInfo = pylsl.StreamInfo(pseudoMarkerConfig.markerName, pseudoMarkerConfig.markerType, 1, 0, 'string', 'Dev')
         self.markerOutlet = pylsl.StreamOutlet(markerInfo)
+        self.markerOutlet.push_sample([self.pseudoMarkerConfig.baselineMarkerString])
         info = pylsl.StreamInfo(dataStreamName, dataStreamType, self.channelCount, self.sampleRate, 'float32', 'Dev')
         chns = info.desc().append_child("channels")
         for label in range(self.channelCount):
