+++ appveyor.yml
@@ -50,9 +50,17 @@ install:
       }
   - sh: |
       case $APPVEYOR_BUILD_WORKER_IMAGE in
-        "ubuntu2204")
+        "ubuntu2004")
           echo "Updating ubuntu system..."
           sudo apt-get update
+          sudo apt-get install python3.10
+          # Install pip for Python 3.10
+          sudo apt-get install python3.10-distutils
+          wget https://bootstrap.pypa.io/get-pip.py
+          python3.10 get-pip.py
+
+          # Set Python 3.10 as the default Python version
+          sudo update-alternatives --install /usr/bin/python python /usr/bin/python3.10 1
           sudo apt-get install -y python3-pip cmake git
           python3 -m pip install --upgrade urllib3>=2.0.5
           python3 -m pip install .
