+++ appveyor.yml
@@ -55,16 +55,16 @@ install:
           sudo apt-get update
           sudo apt-get install python3.10
           # Install pip for Python 3.10
-          sudo apt-get install python3.10-distutils
+          sudo apt-get install python3-distutils
           wget https://bootstrap.pypa.io/get-pip.py
           python3.10 get-pip.py
 
           # Set Python 3.10 as the default Python version
           sudo update-alternatives --install /usr/bin/python python /usr/bin/python3.10 1
           sudo apt-get install -y python3-pip cmake git
-          python3 -m pip install --upgrade urllib3>=2.0.5
-          python3 -m pip install .
-          python3 -m pip install pytest pytest-timeout ruff
+          python3.10 -m pip install --upgrade urllib3>=2.0.5
+          python3.10 -m pip install .
+          python3.10 -m pip install pytest pytest-timeout ruff
           ;;
         "macos-monterey")
           echo "Updating Homebrew and installing packages..."
