+++ appveyor.yml
@@ -56,6 +56,7 @@ install:
           sudo apt-get install python3.10
           # Install pip for Python 3.10
           sudo apt-get install python3-distutils
+          ls /usr/bin/python*
           wget https://bootstrap.pypa.io/get-pip.py
           python310 get-pip.py
 
