+++ appveyor.yml
@@ -68,6 +68,7 @@ install:
           /opt/az/bin/python3.10 -m pip install --upgrade urllib3>=2.0.5
           /opt/az/bin/python3.10 -m pip install .
           /opt/az/bin/python3.10 -m pip install pytest pytest-timeout ruff
+          /opt/az/bin/python3.10 -m site
           ;;
         "macos-monterey")
           echo "Updating Homebrew and installing packages..."
