### 2023-11-03 LMBooth/pybci#25 @LMBooth
Hi again @anilbey, for an update of where i've progressed to - i have the mac distribution working with the pytests on appveyor, however Linux is proving far more problematic. When creating two lsl outlets the second creation seems to overwrite the first meaning  only one stream (either data or marker depending on order of creation) will be seen on the network. I've opened an [issue on the pylsl git here](https://github.com/labstreaminglayer/pylsl/issues/74) to see if they have any suggestions of how to overcome this issue.

If it's not possible then i may have to omit the PseudoDevice class from Linux meaning no automated tests can be run and only streaming hardware would be compatible. I'm curious to know your thoughts on this (should i omit the PseudoDevice or just say pybci does not support Linux) and if you have a windows/mac OS available to verify functionality on?
