### 2023-11-03 LMBooth/pybci#25 @anilbey
Hello @LMBooth thanks for the information.

Given that the underlying technologies (Python, LSL and pylsl) support cross-platform use, it would be unusual for software built upon them to not offer the same level of compatibility.

The bcipy package, which was mentioned in the 'state of the field' section of your paper,  operates on the three major operating systems as well: [bcipy GitHub Actions](https://github.com/CAMBI-tech/BciPy/actions/runs/6724778648).

I can run the tests on Mac and Windows, but is the package not going to work on Linux?

Also I'm curious, does the PseudoDevice (used in tests and examples) have to depend on the pylsl implementation? Can it not generate and stream mock data without having a dependency on pylsl?
