### 2023-11-03 LMBooth/pybci#25 @LMBooth
Hi @anilbey, thankyou for the quick response.

It is an odd error and i'm wondering if it may be due to recent updates in the LSL packages since my previous example worked a few months ago which upon review follows the same methodology of lsl creation. I'm going to set the download for lsl to be an earlier version on linux to see if the tests run successfully, if so i can notify the maintainers of pyLSL that there is an issue.

PyLSL would ideally be the way to create pseudo streams for data and markers since this is the method data would be input from [many off the shelf devices ](https://labstreaminglayer.readthedocs.io/info/supported_devices.html)  or markers from software like Unity,  PsychoPy, or custom languages interfaces be it C, C++, python, Java, MATLAB.

I'll post an update if any previous versions of the liblsl.so has any luck or not.
