### 2023-11-03 LMBooth/pybci#25 @LMBooth
Okayyyy, reverting to earlier liblsl.so versions or pylsl versions didn't help, though if i run my old separate script in one terminal and set pseudodevice to false in another the tests work on ubuntu.

To speed things up I've partitioned my home machine and installed Ubuntu on it too so i don't need to go in the office to check or do it via appveyor anymore so that will hopefully speed up finding a fix, i'm thinking if i can make the original pseudodevice script run fine then i could take a part of that and subprocess to it so i create another instance of python, *fingers crossed* that'll do it?

I'm out tomorrow day and evening with friends but will hopefully be able to get something working come Sunday night. So sorry about this @anilbey
