### 2023-11-06 LMBooth/pybci#25 @LMBooth
Hi @anilbey , so i have changed the pseudo device class to separate out the threaded and multi-processed functions as i had the multi-processed functions called in the same class inheriting self. Multiprocessing in Windows and Mac is spawned where in Linux it's forked, i believe this is why there was an issue. Now i have the program working correctly on my own Linux build (22.04) and Python version (3.10.12) however it still can only find one stream in the appveyor process.

Do you have any success if you try to install and run an example of pybci 1.4.3?
