### 2023-11-11 LMBooth/pybci#18 @LMBooth
Update: target coverage >70%
