### 2023-11-14 LMBooth/pybci#25 @emdupre
Thank you for identifying and working on this, @anilbey and @LMBooth ! @anilbey, please let us know when you've had an opportunity to assess the proposed fix in v1.4.3.
