### 2023-11-14 LMBooth/pybci#25 @anilbey
So on v1.4.3
```
❯ pip list | grep pybci
pybci-package                1.4.3
```

Errors occur while running the tests.
```
2023-11-14 17:11:52.914 ( 111.886s) [R_PyBCIPseudoM  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
2023-11-14 17:11:52.918 ( 111.889s) [R_PyBCIPseudoD  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
PyBCI: [INFO] - Threads stopped.
.
Tests/test_Simple.py PyBCI: [INFO] -Initialised PseudoDevice...
PyBCI: [INFO] - PseudoDevice - Begin streaming.
PyBCI: [INFO] - PseudoDevice - sending marker Marker1
PyBCI: [INFO] - Found stream name: PyBCIPseudoDataStream
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [INFO] - Found stream name: LSL_Tags
PyBCI: [INFO] - Found stream type: Tags
PyBCI: [INFO] - Found stream name: PyBCIPseudoMarkers
PyBCI: [INFO] - Found stream type: Markers
PyBCI: [INFO] - Invalid or no sklearn classifier passed to clf. Checking tensorflow model...
PyBCI: [INFO] - Invalid or no tensorflow model passed to model.  Checking pytorch torchModel...
PyBCI: [INFO] - Invalid or no PyTorch model passed to model. Defaulting to SVM by SkLearn
PyBCI: [INFO] - Found stream name: PyBCIPseudoDataStream
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [INFO] - Found stream name: LSL_Tags
PyBCI: [INFO] - Found stream type: Tags
PyBCI: [INFO] - Found stream name: PyBCIPseudoMarkers
PyBCI: [INFO] - Found stream type: Markers
PyBCI: [INFO] - Success - 1 data stream(s) found, 1 marker stream found
PyBCI: [INFO] - Starting threads initialisation...
PyBCI: [INFO] - Threads initialised.
PyBCI: [INFO] - Started training...
PyBCI: [INFO] - PseudoDevice - sending marker Marker2
PyBCI: [INFO] - PseudoDevice - sending marker Marker3
PyBCI: [INFO] - PseudoDevice - sending baseline
PyBCI: [INFO] - PseudoDevice - sending marker Marker1
PyBCI: [INFO] - PseudoDevice - sending marker Marker2
PyBCI: [INFO] - PseudoDevice - sending marker Marker3
PyBCI: [INFO] - PseudoDevice - sending baseline
PyBCI: [INFO] - PseudoDevice - sending marker Marker1
2023-11-14 17:13:24.117 ( 203.088s) [R_PyBCIPseudoM  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
2023-11-14 17:13:24.123 ( 203.094s) [R_PyBCIPseudoD  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
PyBCI: [INFO] - Threads stopped.
```
