### 2023-11-14 LMBooth/pybci#25 @LMBooth
> 2023-11-14 17:13:24.117 ( 203.088s) [R_PyBCIPseudoM  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
> 2023-11-14 17:13:24.123 ( 203.094s) [R_PyBCIPseudoD  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...

Hi  @anilbey, these are generated in the LSL libraries instead of pybci when the receiver inlets are disconnected, which occurs when the Pseudo Device outlets are closed at the end of the test, it looks like the test was successful.
