### 2023-11-14 LMBooth/pybci#25 @anilbey
But it is later followed by a segmentation fault.

```
PyBCI: [INFO] - Starting threads initialisation...
PyBCI: [INFO] - Threads initialised.
PyBCI: [INFO] - Started training...
2023-11-14 17:18:40.831 ( 519.803s) [R_PyBCIPseudoM  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
2023-11-14 17:18:40.833 ( 519.804s) [R_PyBCIPseudoD  ]      data_receiver.cpp:344    ERR| Stream transmission broke off (Input stream error.); re-connecting...
PyBCI: [INFO] - Threads stopped.

Timeout reached. Stopping threads.
.
Tests/test_zDualStream.py Fatal Python error: Segmentation fault

```
