### 2023-11-14 LMBooth/pybci#25 @anilbey
Here also, another one

```PyBCI: [INFO] - Invalid or no sklearn classifier passed to clf. Checking tensorflow model...
PyBCI: [INFO] - Invalid or no tensorflow model passed to model.  Checking pytorch torchModel...
PyBCI: [INFO] - Invalid or no PyTorch model passed to model. Defaulting to SVM by SkLearn
PyBCI: [INFO] - Found stream name: dev2
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [INFO] - Found stream name: LSL_Tags
PyBCI: [INFO] - Found stream type: Tags
PyBCI: [INFO] - Found stream name: dev1
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [WARNING] - No Marker streams available, make sure your accepted marker data Type have been set in bci.lslScanner.markerTypes correctly.
Fatal Python error: Segmentation fault

Current thread 0x00007fd084aa5000 (most recent call first):
  Garbage-collecting

```
