### 2023-11-14 LMBooth/pybci#25 @LMBooth
This is odd, when i call `pytest Tests/test_zDualStream.py` and `pytest Tests/test_zSimpleOtherFeatures.py` they both pass fine on Ubuntu 22.04 and Python 3.10.12. Just ran the pytest for all 11 items  to see if it's an ordered effect, but passed with the following output:
```
/Documents/GitHub/pybci$ pytest Tests
================================================================================= test session starts =================================================================================
rootdir: /home/liam/Documents/GitHub/pybci/Tests
configfile: pytest.ini
collected 11 items

Tests/test_PseudoDevice.py .                                                                                                                                                    [  9%]
Tests/test_Pytorch.py .                                                                                                                                                         [ 18%]
Tests/test_Simple.py .                                                                                                                                                          [ 27%]
Tests/test_Sklearn.py .                                                                                                                                                         [ 36%]
Tests/test_Tensorflow.py .                                                                                                                                                      [ 45%]
Tests/test_cli.py ....                                                                                                                                                          [ 81%]
Tests/test_zDualStream.py .                                                                                                                                                     [ 90%]
Tests/test_zSimpleOtherFeatures.py .                                                                                                                                            [100%]

================================================================================== warnings summary ===================================================================================
../../../.local/lib/python3.10/site-packages/tensorflow/__init__.py:29
  /home/liam/.local/lib/python3.10/site-packages/tensorflow/__init__.py:29: DeprecationWarning: The distutils package is deprecated and slated for removal in Python 3.12. Use setuptools or check PEP 632 for potential alternatives
    import distutils as _distutils

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
====================================================================== 11 passed, 1 warning in 656.05s (0:10:56) ======================================================================
```

Does turning off the LSL_Tags stream make a difference?
