### 2023-11-14 LMBooth/pybci#25 @LMBooth
or printing the  output from `pytest -s Tests/test_zDualStream.py`:

```
PyBCI: [INFO] - Invalid or no tensorflow model passed to model.  Checking pytorch torchModel...
PyBCI: [INFO] - Invalid or no PyTorch model passed to model. Defaulting to SVM by SkLearn
PyBCI: [INFO] - Found stream name: dev2
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [INFO] - Found stream name: dev1
PyBCI: [INFO] - Found stream type: EMG
PyBCI: [INFO] - Found stream name: PyBCIPseudoMarkers
PyBCI: [INFO] - Found stream type: Markers
PyBCI: [INFO] - Success - 2 data stream(s) found, 1 marker stream found
PyBCI: [INFO] - Starting threads initialisation...
PyBCI: [INFO] - PseudoDevice - sending marker Marker2
PyBCI: [INFO] - Threads initialised.
PyBCI: [INFO] - Started training...
PyBCI: [INFO] - PseudoDevice - sending marker Marker3
PyBCI: [INFO] - PseudoDevice - sending baseline
PyBCI: [INFO] - PseudoDevice - sending marker Marker1
PyBCI: [INFO] - PseudoDevice - sending marker Marker2
PyBCI: [INFO] - PseudoDevice - sending marker Marker3
PyBCI: [INFO] - PseudoDevice - sending baseline
PyBCI: [INFO] - PseudoDevice - sending marker Marker1
PyBCI: [INFO] - PseudoDevice - sending marker Marker2
2023-11-14 16:51:37.578 (  92.421s) [R_dev1          ]      data_receiver.cpp:342    ERR| Stream transmission broke off (Input stream error.); re-connecting...
2023-11-14 16:51:37.584 (  92.427s) [R_PyBCIPseudoM  ]      data_receiver.cpp:342    ERR| Stream transmission broke off (Input stream error.); re-connecting...
2023-11-14 16:51:39.580 (  94.422s) [R_dev2          ]      data_receiver.cpp:342    ERR| Stream transmission broke off (Input stream error.); re-connecting...
PyBCI: [INFO] - Threads stopped.
.

================================================================================== warnings summary ===================================================================================
test_zDualStream.py::test_run_dual
  /home/liam/.local/lib/python3.10/site-packages/tensorflow/__init__.py:29: DeprecationWarning: The distutils package is deprecated and slated for removal in Python 3.12. Use setuptools or check PEP 632 for potential alternatives
    import distutils as _distutils

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
====================================================================== 1 passed, 1 warning in 103.48s (0:01:43) =======================================================================
```
