### 2023-11-14 LMBooth/pybci#25 @anilbey
> Does turning off the LSL_Tags stream make a difference?

I don't know.

 A docker image may help in this case with reproducibility and it will also simplify the installation.
