### 2023-11-14 LMBooth/pybci#25 @LMBooth
Ahh, i ask about the LSL_Tags stream as that isn't set by pybci so was unsure if you had other LSL streams running from another source.

I've started putting together a dockerfile in the root of the directory, hopefully i'll have time tomorrow evening to verify and that will give a more clear - and fully-functional! - method of installation for Linux.
