### 2023-11-16 LMBooth/pybci#25 @LMBooth
Hi @anilbey , can you run with the [Dockerfile](https://github.com/LMBooth/pybci/blob/main/Dockerfile)?
