### 2023-11-16 LMBooth/pybci#25 @anilbey
Thanks a lot @LMBooth for the quick implementation. I am on a leave now and don't have access to a linux machine however I will check it in detail after I return (on the 26th).
