### 2023-11-29 LMBooth/pybci#25 @anilbey
Brilliant, I can run both the tests and the examples with the docker image and the instructions provided on the README.
Tests:
```
11 passed, 1 warning in 640.70s (0:10:40)
```
Examples:
```
PyBCI: [INFO] - PseudoDevice - sending marker Marker36], 'baseline': [2, 6], 'Marker1': [3, 6]} Accuracy: 1.0
Markers received: {'Marker2': [0, 6], 'Marker3': [1, 9], 'baseline': [2, 6], 'Marker1': [3, 6]} Accuracy: 1.0
```

Thanks @LMBooth. I think now this issue can be marked as resolved.
