### 2023-11-29 LMBooth/pybci#26 @anilbey
small issue: logo on readme becomes invisible on github's dark theme
Not a serious one but may be worth addressing.

Part of https://github.com/openjournals/joss-reviews/issues/5706.
