### 2023-11-29 LMBooth/pybci#25 @LMBooth
Excellent, thankyou for providing your input to help resolve this @anilbey, a meme comes to mind:
![image](https://github.com/LMBooth/pybci/assets/31129153/089402a9-28fe-42f9-8470-f71be40de11d)

I've just today extended some of the pytest coverage too to help ensure some of the additional functionality functions correctly.
