### 2023-11-29 LMBooth/pybci#25 @anilbey
> I've just today extended some of the pytest coverage too to help ensure some of the additional functionality functions correctly.

Great news!
