### 2023-11-29 LMBooth/pybci#26 @LMBooth
I have pondered on this, whether to block white the background or not, i suppose it would cover more basis. I'll give it a go.
