+++ docs/source/conf.py
@@ -5,7 +5,7 @@ copyright = '2023, Liam Booth'
 author = 'Liam Booth'
 
 # The full version, including alpha/beta/rc tags
-release = '1.4.2'
+release = '1.4.4'
 
 # -- General configuration ---------------------------------------------------
 extensions = [
