### 2023-11-29 #5706 @emdupre
Thank you, @jsheunis and @anilbey ! It looks like both of your reviewer checklists are complete (and all associated issues have been closed), so I am ready to take this submission forward. Of course, if you have any remaining concerns in publishing _PyBCI_, please let me know.

I'll go ahead and perform a few editorial checks now.
