### 2023-11-30 #5706 @emdupre
I've reviewed the software and am also very happy with this submission 🎉

The software paper itself is also very well-written. I did, though, have a few small editorial requests :

- Please update your author affiliations to include the country; e.g., `University of Hull, United Kingdom`
- L31-32 : This sentence appears to be missing a word ; I would suggest inserting "that" before "offer ease of use"
- L31-33 : Please add spaces before each in-text citation
- L43 : Are these two sentences intended to be joined ? That is, "...stored in pre-allocated NumPy arrays when in training mode, based on a configurable time window before and after each marker type."
- L44 : Please add a comma after "When in test mode"
- L46 : Please add a comma after the Antropy citation
- References : In the [`vallat_antropy_2023`](https://github.com/LMBooth/pybci/blob/1158144822e67066eb4b8490ad4c97809c8914cf/JOSS/paper.bib#L77C7-L77C26) and  [`BciPy`](https://github.com/LMBooth/pybci/blob/1158144822e67066eb4b8490ad4c97809c8914cf/JOSS/paper.bib#L154) references, please ensure that "Python" is capitalized in the formatted reference by enclosing it in brackets; e.g.,
  > title = {BciPy: brain–computer interface software in {P}ython},
- References : I see that PyLSL does [specifically ask](https://labstreaminglayer.readthedocs.io/info/faqs.html?highlight=cite#referencing-lsl) to cite the GitHub repository, but I'm a little hesitant for that to be the only citation as it is a non-persistent link. Would you also be willing to [directly cite `liblsl`](https://zenodo.org/records/7978343), assuming I understand correctly that this is indeed what you've used ? Alternatively, if a pylsl citation would be best, would you be able to cite the pylsl version directly, using something like:
  > @software{pylsl,
  author = {Kothe, Christian},
  title = {PyLSL},
  url = {https://github.com/labstreaminglayer/pylsl},
  version = {v1.16.2},
  date = {2023-11-29},
}

  I recognize that this is also a non-persistent link, but it at least provides more insight into the state of the repository when it was cited.
