### 2023-11-30 #5706 @LMBooth
HI @emdupre,

Thankyou for finding these changes in the paper. I've made my way through most the edits.

For this point:
> * L43 : Are these two sentences intended to be joined ? That is, "...stored in pre-allocated NumPy arrays when in training mode, based on a configurable time window before and after each marker type."

They are intended to be two separate sentences, " Samples are collected in chunks from the LSL data streams and stored in pre-allocated NumPy arrays." Applies to both test and train mode, so we'll keep them separate.

As for changing the LSL link, i've decided to go with the [Zenodo repository](https://zenodo.org/records/7978343).

I'll post back here when i've had chance to archive the pybci the repository.
