### 2023-11-30 #5706 @LMBooth
> After making these changes, could you then please:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in [Zenodo](https://zenodo.org/) or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
> * [ ]  List the DOI of the archived version here.

Hi @emdupre, the DOI archived version can be found here for PyBCI v1.5.0: https://doi.org/10.5281/zenodo.10224399, If there are any issue or you need anything else from me let me know!
