### 2023-11-30 #5706 @emdupre
Thank you, @LMBooth !

A few small requests on the archive :

- [ ] The license on the archive is currently CC-BY 4.0. Could you please update this to match the software license of MIT ?
- [ ] It looks like the author listing is slightly different on the archive than the software paper (i.e., second and third authors have exchanged places). Could you please harmonize the ordering or confirm if these differences are intentional ?
- [ ] The second and third authors appear to have their first and last names flipped in the archive. For example, in the paper one author is listed as "Anthony Bateson" but in the archive they are listed as "Anthony, Bateson" implying their full name is Bateson Anthony. This is also the case for "Aziz, Asghar." Could you please confirm these names are entered correctly?

You should be able to make all of these changes by editing the archive metadata ; if it requires minting a new DOI, please let me know.
