### 2023-12-01 #5706 @LMBooth
Hi @emdupre, the license and authors have been updated on Zenodo and should be in line with the JOSS submission, please don't hesitate to get in touch if anything else is needed.
