### 2023-12-01 #5706 @emdupre
Thank you for your engagement throughout the review process, @LMBooth ! I'm now happy to recommend _PyBCI_ to the EiC team for publication in JOSS 🎉

Thank you, too, to @anilbey and @jsheunis for your reviews and input throughout this process ! JOSS runs thanks to volunteers like you, so thank you 💐
