### 2023-12-01 #5706 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
